Cardano.Ledger.Conway

Documentation

type Conway = ConwayEra

Deprecated: In favor of ConwayEra

data ConwayEra

hardforkConwayBootstrapPhase ∷ ProtVer → Bool

Bootstrap phase

hardforkConwayDisallowUnelectedCommitteeFromVoting ∷ ProtVer → Bool

Starting with protocol version 11, we do not allow unelected committee members to submit votes.

hardforkConwayDELEGIncorrectDepositsAndRefunds ∷ ProtVer → Bool

Starting with protocol version 11, we report incorrect deposit and refunds better

hardforkConwayMoveWithdrawalsAndDRepChecksToLedgerRule ∷ ProtVer → Bool

Starting with protocol version 11, we move a predicate check and updates related to DRep dormancy and expiry from CERTS to LEDGER since that is a more suitable place for them.

updates to drep expiry for all voting dreps

drep dormancy tracking updates

withdrawals draining - (split into two)

NOTE: In addition, we split the predicate check for withdrawals into two to make it better: both invalid withdrawals (submitted in the wrong network or with missing reward accounts) and incomplete withdrawals were being reported with WithdrawalsNotInRewardsCERTS but now ConwayWithdrawalsMissingAccounts and ConwayIncompleteWithdrawals are the new predicate failures we use to report the two separate cases in LEDGER

data family Tx era

Orphan instances

RunConwayRatify ConwayEra

Instance details

Methods

runConwayRatify ∷ Globals → RatifyEnv ConwayEra → RatifyState ConwayEra → RatifySignal ConwayEra → RatifyState ConwayEra

ApplyTx ConwayEra

Instance details

Methods

applyTxValidation ∷ ValidationPolicy → Globals → MempoolEnv ConwayEra → MempoolState ConwayEra → Tx ConwayEra → Either (ApplyTxError ConwayEra) (MempoolState ConwayEra, Validated (Tx ConwayEra))

ApplyBlock ConwayEra

Instance details

Methods

applyBlock ∷ ∀ (ep ∷ EventPolicy). SingEP ep → ValidationPolicy → Globals → NewEpochState ConwayEra → Block BHeaderView ConwayEra → (NewEpochState ConwayEra, [PredicateFailure (EraRule "BBODY" ConwayEra)], [Event (EraRule "BBODY" ConwayEra)])

applyTick ∷ ∀ (ep ∷ EventPolicy). SingEP ep → Globals → NewEpochState ConwayEra → SlotNo → (NewEpochState ConwayEra, [Event (EraRule "TICK" ConwayEra)])