Cardano.Ledger.Conway.Core

Documentation

class (BabbageEraTxBody era, ConwayEraTxCert era, ConwayEraPParams era, ConwayEraScript era) ⇒ ConwayEraTxBody era where

Methods

currentTreasuryValueTxBodyL ∷ Lens' (TxBody era) (StrictMaybe Coin)

Lens for getting and setting number of Coin that is expected to be in the Treasury at the current Epoch

votingProceduresTxBodyL ∷ Lens' (TxBody era) (VotingProcedures era)

Lens for getting and setting VotingProcedures.

proposalProceduresTxBodyL ∷ Lens' (TxBody era) (OSet (ProposalProcedure era))

Lens for getting and setting ProposalProcedures.

treasuryDonationTxBodyL ∷ Lens' (TxBody era) Coin

class BabbageEraPParams era ⇒ ConwayEraPParams era

Minimal complete definition

modifiedPPGroups, ppuWellFormed, hkdPoolVotingThresholdsL, hkdDRepVotingThresholdsL, hkdCommitteeMinSizeL, hkdCommitteeMaxTermLengthL, hkdGovActionLifetimeL, hkdGovActionDepositL, hkdDRepDepositL, hkdDRepActivityL, hkdMinFeeRefScriptCostPerByteL

ppPoolVotingThresholdsL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) PoolVotingThresholds

ppDRepVotingThresholdsL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) DRepVotingThresholds

ppCommitteeMinSizeL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) Natural

ppCommitteeMaxTermLengthL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) EpochInterval

ppGovActionLifetimeL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) EpochInterval

ppGovActionDepositL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) Coin

ppDRepDepositL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) Coin

ppDRepActivityL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) EpochInterval

ppuPoolVotingThresholdsL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe PoolVotingThresholds)

ppuDRepVotingThresholdsL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe DRepVotingThresholds)

ppuCommitteeMinSizeL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

ppuCommitteeMaxTermLengthL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

ppuGovActionLifetimeL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

ppuGovActionDepositL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

ppuDRepDepositL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

ppuDRepActivityL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

data PoolVotingThresholds

Constructors

PoolVotingThresholds

Fields

pvtMotionNoConfidence ∷ !UnitInterval

pvtCommitteeNormal ∷ !UnitInterval

pvtCommitteeNoConfidence ∷ !UnitInterval

pvtHardForkInitiation ∷ !UnitInterval

pvtPPSecurityGroup ∷ !UnitInterval

data DRepVotingThresholds

Constructors

DRepVotingThresholds

Fields

dvtMotionNoConfidence ∷ !UnitInterval

dvtCommitteeNormal ∷ !UnitInterval

dvtCommitteeNoConfidence ∷ !UnitInterval

dvtUpdateToConstitution ∷ !UnitInterval

dvtHardForkInitiation ∷ !UnitInterval

dvtPPNetworkGroup ∷ !UnitInterval

dvtPPEconomicGroup ∷ !UnitInterval

dvtPPTechnicalGroup ∷ !UnitInterval

dvtPPGovGroup ∷ !UnitInterval

dvtTreasuryWithdrawal ∷ !UnitInterval

dvtPPNetworkGroupL ∷ Lens' DRepVotingThresholds UnitInterval

dvtPPGovGroupL ∷ Lens' DRepVotingThresholds UnitInterval

dvtPPTechnicalGroupL ∷ Lens' DRepVotingThresholds UnitInterval

dvtPPEconomicGroupL ∷ Lens' DRepVotingThresholds UnitInterval

dvtUpdateToConstitutionL ∷ Lens' DRepVotingThresholds UnitInterval

class AlonzoEraScript era ⇒ ConwayEraScript era where

Methods

mkVotingPurpose ∷ f Word32 (Voter (EraCrypto era)) → PlutusPurpose f era

toVotingPurpose ∷ PlutusPurpose f era → Maybe (f Word32 (Voter (EraCrypto era)))

mkProposingPurpose ∷ f Word32 (ProposalProcedure era) → PlutusPurpose f era

toProposingPurpose ∷ PlutusPurpose f era → Maybe (f Word32 (ProposalProcedure era))

pattern VotingPurpose ∷ ConwayEraScript era ⇒ f Word32 (Voter (EraCrypto era)) → PlutusPurpose f era

pattern ProposingPurpose ∷ ConwayEraScript era ⇒ f Word32 (ProposalProcedure era) → PlutusPurpose f era

class ShelleyEraTxCert era ⇒ ConwayEraTxCert era

Minimal complete definition

mkRegDepositTxCert, getRegDepositTxCert, mkUnRegDepositTxCert, getUnRegDepositTxCert, mkDelegTxCert, getDelegTxCert, mkRegDepositDelegTxCert, getRegDepositDelegTxCert, mkAuthCommitteeHotKeyTxCert, getAuthCommitteeHotKeyTxCert, mkResignCommitteeColdTxCert, getResignCommitteeColdTxCert, mkRegDRepTxCert, getRegDRepTxCert, mkUnRegDRepTxCert, getUnRegDRepTxCert, mkUpdateDRepTxCert, getUpdateDRepTxCert

pattern RegDepositTxCert ∷ ConwayEraTxCert era ⇒ StakeCredential (EraCrypto era) → Coin → TxCert era

pattern UnRegDepositTxCert ∷ ConwayEraTxCert era ⇒ StakeCredential (EraCrypto era) → Coin → TxCert era

pattern DelegTxCert ∷ ConwayEraTxCert era ⇒ StakeCredential (EraCrypto era) → Delegatee (EraCrypto era) → TxCert era

pattern RegDepositDelegTxCert ∷ ConwayEraTxCert era ⇒ StakeCredential (EraCrypto era) → Delegatee (EraCrypto era) → Coin → TxCert era

pattern AuthCommitteeHotKeyTxCert ∷ ConwayEraTxCert era ⇒ Credential 'ColdCommitteeRole (EraCrypto era) → Credential 'HotCommitteeRole (EraCrypto era) → TxCert era

pattern ResignCommitteeColdTxCert ∷ ConwayEraTxCert era ⇒ Credential 'ColdCommitteeRole (EraCrypto era) → StrictMaybe (Anchor (EraCrypto era)) → TxCert era

pattern RegDRepTxCert ∷ ConwayEraTxCert era ⇒ Credential 'DRepRole (EraCrypto era) → Coin → StrictMaybe (Anchor (EraCrypto era)) → TxCert era

pattern UnRegDRepTxCert ∷ ConwayEraTxCert era ⇒ Credential 'DRepRole (EraCrypto era) → Coin → TxCert era

pattern UpdateDRepTxCert ∷ ConwayEraTxCert era ⇒ Credential 'DRepRole (EraCrypto era) → StrictMaybe (Anchor (EraCrypto era)) → TxCert era

module Cardano.Ledger.Babbage.Core