Cardano.Ledger.Conway.Governance

Documentation

class (EraPParams era, Eq (GovState era), Show (GovState era), NoThunks (GovState era), NFData (GovState era), EncCBOR (GovState era), DecCBOR (GovState era), DecShareCBOR (GovState era), Share (GovState era) ~ (Interns (Credential 'Staking), Interns (KeyHash 'StakePool), Interns (Credential 'DRepRole), Interns (Credential 'HotCommitteeRole)), ToCBOR (GovState era), FromCBOR (GovState era), Default (GovState era), ToJSON (GovState era)) ⇒ EraGov era where

Minimal complete definition

curPParamsGovStateL, prevPParamsGovStateL, futurePParamsGovStateL, obligationGovState

Associated Types

type GovState era = (r ∷ Type) | r → era

Methods

emptyGovState ∷ GovState era

Construct empty governance state

curPParamsGovStateL ∷ Lens' (GovState era) (PParams era)

Lens for accessing current protocol parameters

prevPParamsGovStateL ∷ Lens' (GovState era) (PParams era)

Lens for accessing the previous protocol parameters

futurePParamsGovStateL ∷ Lens' (GovState era) (FuturePParams era)

Lens for accessing the future protocol parameters.

This lens will produce DefinitePParamsUpdate whenever we are absolutely sure that the new PParams will be updated. Which means there will be no chance of a DefinitePParamsUpdate value until we are past the point of no return, which is 2 stability windows before the end of the epoch. This lens is mostly intended for ledger usage and nextEpochUpdatedPParams should be used instead whenever definite results are desired.

obligationGovState ∷ GovState era → Obligations

data EnactState era

Constructors

EnactState

Fields

ensCommittee ∷ !(StrictMaybe (Committee era))

Constitutional Committee

ensConstitution ∷ !(Constitution era)

Constitution

ensCurPParams ∷ !(PParams era)

ensPrevPParams ∷ !(PParams era)

ensTreasury ∷ !Coin

ensWithdrawals ∷ !(Map (Credential 'Staking) Coin)

ensPrevGovActionIds ∷ !(GovRelation StrictMaybe)

Last enacted GovAction Ids

data RatifyState era

RatifyState stores information about what will happen to the active governance actions at the next epoch boundary.

Constructors

RatifyState

Fields

rsEnactState ∷ !(EnactState era)

This is the currently active EnactState. It contains all the changes that were applied to it at the last epoch boundary by all the proposals that were enacted.

rsEnacted ∷ !(Seq (GovActionState era))

Governance actions that are going to be enacted at the next epoch boundary.

rsExpired ∷ !(Set GovActionId)

Governance actions that are going to be removed at the next epoch boundary, either due to expiring or because they would become invalid after another governance action gets enacted or expired before it

rsDelayed ∷ !Bool

This flag is set to true if one of the proposals that was ratified at the last epoch boundary was a delaying action. This means that no other proposals will be ratified this epoch and each active proposal that has not become invalid will have its expiry date extended by one epoch.

This flag is reset at each epoch boundary before the RATIFY rule gets called, but it might immediately be set to True again after the RATIFY rule has finished execution.

data RatifyEnv era

Constructors

RatifyEnv

Fields

reInstantStake ∷ InstantStake era

reStakePoolDistr ∷ PoolDistr

reDRepDistr ∷ Map DRep (CompactForm Coin)

reDRepState ∷ Map (Credential 'DRepRole) DRepState

reCurrentEpoch ∷ EpochNo

reCommitteeState ∷ CommitteeState era

reAccounts ∷ Accounts era

reStakePools ∷ Map (KeyHash 'StakePool) StakePoolState

newtype RatifySignal era

Constructors

RatifySignal

Fields

unRatifySignal ∷ StrictSeq (GovActionState era)

data ConwayGovState era

Conway governance state

Constructors

ConwayGovState

Fields

cgsProposals ∷ !(Proposals era)

cgsCommittee ∷ !(StrictMaybe (Committee era))

cgsConstitution ∷ !(Constitution era)

cgsCurPParams ∷ !(PParams era)

cgsPrevPParams ∷ !(PParams era)

cgsFuturePParams ∷ !(FuturePParams era)

cgsDRepPulsingState ∷ !(DRepPulsingState era)

The cgsDRepPulsingState field is a pulser that incrementally computes the stake distribution of the DReps over the Epoch following the close of voting at end of the previous Epoch. It assembles this with some of its other internal components into a (RatifyEnv era) when it completes, and then calls the RATIFY rule and eventually returns the updated RatifyState. The pulser is created at the Epoch boundary, but does no work until it is pulsed in the NEWEPOCH rule, whenever the system is NOT at the epoch boundary.

predictFuturePParams ∷ (ConwayEraAccounts era, EraStake era) ⇒ ConwayGovState era → ConwayGovState era

This function updates the thunk, which will contain new PParams once evaluated or Nothing when there was no update. At the same time if we already know the future of PParams, then it will act as an identity function.

data Committee era

Constructors

Committee

Fields

committeeMembers ∷ !(Map (Credential 'ColdCommitteeRole) EpochNo)

Committee members with epoch number when each of them expires

committeeThreshold ∷ !UnitInterval

Threshold of the committee that is necessary for a successful vote

committeeMembersL ∷ Lens' (Committee era) (Map (Credential 'ColdCommitteeRole) EpochNo)

committeeThresholdL ∷ Lens' (Committee era) UnitInterval

authorizedElectedHotCommitteeCredentials ∷ StrictMaybe (Committee era) → CommitteeState era → Set (Credential 'HotCommitteeRole)

Extract all unique hot credential authorizations for the current committee that is elected.

data GovAction era

Note that the previous governance action id is only optional for the very first governance action of the same purpose.

Constructors

ParameterChange

Fields

!(StrictMaybe (GovPurposeId 'PParamUpdatePurpose))

Previous governance action id of ParameterChange type, which corresponds to PParamUpdatePurpose.

!(PParamsUpdate era)

Proposed changes to PParams

!(StrictMaybe ScriptHash)

Policy hash protection

HardForkInitiation

Fields

!(StrictMaybe (GovPurposeId 'HardForkPurpose))

Previous governance action id of HardForkInitiation type, which corresponds to HardForkPurpose

!ProtVer

Proposed new protocol version

TreasuryWithdrawals

Fields

!(Map RewardAccount Coin)

Proposed treasury withdrawals

!(StrictMaybe ScriptHash)

Policy hash protection

NoConfidence

Fields

!(StrictMaybe (GovPurposeId 'CommitteePurpose))

Previous governance action id of NoConfidence or UpdateCommittee type, which corresponds to CommitteePurpose

UpdateCommittee

Fields

!(StrictMaybe (GovPurposeId 'CommitteePurpose))

Previous governance action id of UpdateCommittee or NoConfidence type, which corresponds to CommitteePurpose

!(Set (Credential 'ColdCommitteeRole))

Constitutional Committe members to be removed

!(Map (Credential 'ColdCommitteeRole) EpochNo)

Constitutional committee members to be added

!UnitInterval

New Threshold

NewConstitution

Fields

!(StrictMaybe (GovPurposeId 'ConstitutionPurpose))

Previous governance action id of NewConstitution type, which corresponds to ConstitutionPurpose

!(Constitution era)

InfoAction

data GovActionState era

Constructors

GovActionState

Fields

gasId ∷ !GovActionId

gasCommitteeVotes ∷ !(Map (Credential 'HotCommitteeRole) Vote)

gasDRepVotes ∷ !(Map (Credential 'DRepRole) Vote)

gasStakePoolVotes ∷ !(Map (KeyHash 'StakePool) Vote)

gasProposalProcedure ∷ !(ProposalProcedure era)

gasProposedIn ∷ !EpochNo

gasExpiresAfter ∷ !EpochNo

newtype GovActionIx

Constructors

GovActionIx

Fields

unGovActionIx ∷ Word16

data GovActionId

Constructors

GovActionId

Fields

gaidTxId ∷ !TxId

gaidGovActionIx ∷ !GovActionIx

data GovActionPurpose

Constructors

PParamUpdatePurpose

HardForkPurpose

CommitteePurpose

ConstitutionPurpose

class ToGovActionPurpose (p ∷ GovActionPurpose)

Minimal complete definition

toGovActionPurpose

isGovActionWithPurpose ∷ ∀ p era. ToGovActionPurpose p ⇒ GovAction era → Bool

data DRepPulsingState era

Constructors

DRPulsing !(DRepPulser era Identity (RatifyState era))

DRComplete !(PulsingSnapshot era) !(RatifyState era)

data DRepPulser era (m ∷ Type → Type) ans where

The type of a Pulser which uses computeDRepDistr as its underlying function. Note that we use two type equality (~) constraints to fix both the monad m and the ans type, to the context where we will use the type as a Pulser. The type DRepPulser must have m and ans as its last two parameters so we can make a Pulsable instance. We will always use this instantiation (DRepPulser era Identity (RatifyState era))

Constructors

DRepPulser

Fields

∷ ∀ era ans m. (ans ~ RatifyState era, m ~ Identity, RunConwayRatify era)

⇒ { dpPulseSize ∷ !Int

How many elements of dpAccounts to consume each pulse.

, dpAccounts ∷ !(Accounts era)

Snapshot containing the mapping of stake credentials to DReps, Pools and Rewards.

, dpIndex ∷ !Int

The index of the iterator over dpAccounts. Grows with each pulse.

, dpInstantStake ∷ !(InstantStake era)

Snapshot of the stake distr (comes from the IncrementalStake)

, dpStakePoolDistr ∷ PoolDistr

Snapshot of the pool distr. Lazy on purpose: See $sel:ssStakeMarkPoolDistr:SnapShots and ADR-7 for explanation.

, dpDRepDistr ∷ !(Map DRep (CompactForm Coin))

The partial result that grows with each pulse. The purpose of the pulsing.

, dpDRepState ∷ !(Map (Credential 'DRepRole) DRepState)

Snapshot of registered DRep credentials

, dpCurrentEpoch ∷ !EpochNo

Snapshot of the EpochNo this pulser will complete in.

, dpCommitteeState ∷ !(CommitteeState era)

Snapshot of the CommitteeState

, dpEnactState ∷ !(EnactState era)

Snapshot of the EnactState, Used to build the Env of the RATIFY rule

, dpProposals ∷ !(StrictSeq (GovActionState era))

Snapshot of the proposals. This is the Signal for the RATIFY rule

, dpProposalDeposits ∷ !(Map (Credential 'Staking) (CompactForm Coin))

Snapshot of the proposal-deposits per reward-account-staking-credential

, dpGlobals ∷ !Globals

, dpStakePools ∷ !(Map (KeyHash 'StakePool) StakePoolState)

Snapshot of the parameters of stake pools - this is needed to get the reward account for SPO vote calculation

} → DRepPulser era m ans

govActionIdToText ∷ GovActionId → Text

data Voter

Constructors

CommitteeVoter !(Credential 'HotCommitteeRole)

DRepVoter !(Credential 'DRepRole)

StakePoolVoter !(KeyHash 'StakePool)

data Vote

Constructors

VoteNo

VoteYes

Abstain

data VotingProcedure era

Constructors

VotingProcedure

Fields

vProcVote ∷ !Vote

vProcAnchor ∷ !(StrictMaybe Anchor)

newtype VotingProcedures era

Constructors

VotingProcedures

Fields

unVotingProcedures ∷ Map Voter (Map GovActionId (VotingProcedure era))

foldlVotingProcedures

Arguments

∷ (c → Voter → GovActionId → VotingProcedure era → c)
Accumulating function

→ c
Initial accumulator

→ VotingProcedures era
Procedures to fold over

→ c

foldrVotingProcedures

Arguments

∷ (Voter → GovActionId → VotingProcedure era → c → c)
Accumulating function

→ c
Initial accumulator

→ VotingProcedures era
Procedures to fold over

→ c

data ProposalProcedure era

Constructors

ProposalProcedure

Fields

pProcDeposit ∷ !Coin

pProcReturnAddr ∷ !RewardAccount

pProcGovAction ∷ !(GovAction era)

pProcAnchor ∷ !Anchor

data Anchor

Constructors

Anchor

Fields

anchorUrl ∷ !Url

anchorDataHash ∷ !(SafeHash AnchorData)

newtype AnchorData

Constructors

AnchorData ByteString

indexedGovProps ∷ Seq (ProposalProcedure era) → Seq (GovActionIx, ProposalProcedure era)

Attaches indices to a sequence of proposal procedures. The indices grow from left to right.

data Constitution era

Constructors

Constitution

Fields

constitutionAnchor ∷ !Anchor

constitutionScript ∷ !(StrictMaybe ScriptHash)

class (EraGov era, EraStake era, ConwayEraAccounts era) ⇒ ConwayEraGov era where

Methods

constitutionGovStateL ∷ Lens' (GovState era) (Constitution era)

proposalsGovStateL ∷ Lens' (GovState era) (Proposals era)

drepPulsingStateGovStateL ∷ Lens' (GovState era) (DRepPulsingState era)

committeeGovStateL ∷ Lens' (GovState era) (StrictMaybe (Committee era))

votingStakePoolThreshold ∷ ConwayEraPParams era ⇒ RatifyState era → GovAction era → StrictMaybe UnitInterval

votingDRepThreshold ∷ ConwayEraPParams era ⇒ RatifyState era → GovAction era → StrictMaybe UnitInterval

votingCommitteeThreshold ∷ ConwayEraPParams era ⇒ EpochNo → RatifyState era → CommitteeState era → GovAction era → StrictMaybe UnitInterval

isStakePoolVotingAllowed ∷ ConwayEraPParams era ⇒ GovAction era → Bool

isDRepVotingAllowed ∷ ConwayEraPParams era ⇒ GovAction era → Bool

isCommitteeVotingAllowed ∷ ConwayEraPParams era ⇒ EpochNo → CommitteeState era → GovAction era → Bool

reorderActions ∷ StrictSeq (GovActionState era) → StrictSeq (GovActionState era)

actionPriority ∷ GovAction era → Int

data Proposals era

Self-contained representation of all 4 proposals trees. This forest is made up of only action-ids for nodes - full GovActionStates are stored only once in the OMap. All functions in this module prefixed with the string proposals- operate on this data-type keeping it consistent.

NOTE: The correct way to think about this data-structure is similar to 4 of the following, one for each GovActionPurpose

  data Tree a = Node (StrictMaybe a) [Tree a]

but because this does not allow us to look-up a node's edges in predictable time, we use a map from nodes to their edges (parent and children) to capture the graph (PGraph). We also need to always know the roots of the 4 trees, and those we store in the PRoot

NOTE: At the end of an epoch boundary, we expect pRoots to be the same as the PrevGovActionIds from the EnactState

mkProposals ∷ (EraPParams era, MonadFail m) ⇒ GovRelation StrictMaybe → OMap GovActionId (GovActionState era) → m (Proposals era)

Reconstruct the Proposals forest from an OMap of GovActionStates and the 4 roots (PrevGovActionIds)

unsafeMkProposals ∷ (HasCallStack, EraPParams era) ⇒ GovRelation StrictMaybe → OMap GovActionId (GovActionState era) → Proposals era

Reconstruct the Proposals forest from an OMap of GovActionStates and the 4 roots (PrevGovActionIds). This function can fail and may return a malformed Proposals if not given correct inputs.

WARNING: Should only be used for testing!

newtype GovPurposeId (p ∷ GovActionPurpose)

Constructors

GovPurposeId

Fields

unGovPurposeId ∷ GovActionId

data PRoot a

The root of a single Proposals tree. prRoot is always expected to be equal to the respective PrevGovActionId at the end of every epoch boundary

Constructors

PRoot

Fields

prRoot ∷ !(StrictMaybe a)

prChildren ∷ !(Set a)

data PEdges a

A non-root edges in a Proposals tree. peParent is expected to be a SNothing only at the begining when no governance actions has been enacted yet.

Constructors

PEdges

Fields

peParent ∷ !(StrictMaybe a)

peChildren ∷ !(Set a)

newtype PGraph a

A single proposal-tree. This map represents all the action-ids that form a tree.

Constructors

PGraph

Fields

unPGraph ∷ Map a (PEdges a)

pRootsL ∷ Lens' (Proposals era) (GovRelation PRoot)

pPropsL ∷ Lens' (Proposals era) (OMap GovActionId (GovActionState era))

prRootL ∷ Lens' (PRoot a) (StrictMaybe a)

prChildrenL ∷ Lens' (PRoot a) (Set a)

peChildrenL ∷ Lens' (PEdges a) (Set a)

pGraphL ∷ Lens' (Proposals era) (GovRelation PGraph)

pGraphNodesL ∷ Lens' (PGraph a) (Map a (PEdges a))

data GovRelation (f ∷ Type → Type)

Abstract data type for representing relationship of governance action with the same purpose

Constructors

GovRelation

Fields

grPParamUpdate ∷ !(f (GovPurposeId 'PParamUpdatePurpose))

grHardFork ∷ !(f (GovPurposeId 'HardForkPurpose))

grCommittee ∷ !(f (GovPurposeId 'CommitteePurpose))

grConstitution ∷ !(f (GovPurposeId 'ConstitutionPurpose))

hoistGovRelation ∷ (∀ a. f a → g a) → GovRelation f → GovRelation g

withGovActionParent

Arguments

∷ GovActionState era

→ a
The result to be used for governance actions that can't have a parent

→ (∀ p. (∀ f. Lens' (GovRelation f) (f (GovPurposeId p))) → StrictMaybe (GovPurposeId p) → GovPurposeId p → a)
Function that will be applied to a lens and a parent

→ a

Apply a function to a GovAction that can have a parent.

toPrevGovActionIds ∷ GovRelation PRoot → GovRelation StrictMaybe

fromPrevGovActionIds ∷ GovRelation StrictMaybe → GovRelation PRoot

grPParamUpdateL ∷ Lens' (GovRelation f) (f (GovPurposeId 'PParamUpdatePurpose))

grHardForkL ∷ Lens' (GovRelation f) (f (GovPurposeId 'HardForkPurpose))

grCommitteeL ∷ Lens' (GovRelation f) (f (GovPurposeId 'CommitteePurpose))

grConstitutionL ∷ Lens' (GovRelation f) (f (GovPurposeId 'ConstitutionPurpose))

proposalsActions ∷ Proposals era → StrictSeq (GovActionState era)

Get the sequence of GovActionStates

proposalsDeposits ∷ Proposals era → Map (Credential 'Staking) (CompactForm Coin)

Get a mapping from the reward-account staking credentials to deposits of all proposals.

proposalsAddAction ∷ ∀ era. (EraPParams era, HasCallStack) ⇒ GovActionState era → Proposals era → Maybe (Proposals era)

Add a single GovActionState to the Proposals forest. The tree to which it is added is picked according to its GovActionPurpose. Returns Nothing when the operation cannot succeed.

proposalsRemoveWithDescendants ∷ EraPParams era ⇒ Set GovActionId → Proposals era → (Proposals era, Map GovActionId (GovActionState era))

Remove the set of given action-ids with their descendants from the Proposals forest. Cannot be used for removing enacted GovActionIds (i.e. roots)

proposalsAddVote ∷ Voter → Vote → GovActionId → Proposals era → Proposals era

Add a vote to an existing GovActionState. This is a no-op if the provided GovActionId does not already exist

proposalsIds ∷ Proposals era → StrictSeq GovActionId

Get the sequence of GovActionIds

proposalsApplyEnactment ∷ ∀ era. EraPParams era ⇒ Seq (GovActionState era) → Set GovActionId → Proposals era → (Proposals era, Map GovActionId (GovActionState era), Map GovActionId (GovActionState era), Map GovActionId (GovActionState era))

For use in the EPOCH rule. Apply the result of extractDRepPulsingState to the Proposals forest, so that: i. all the expired action-ids and their descendants are removed, and ii. the sequence of enacted action-ids is promoted to the root, removing competing/sibling action-ids and their descendants at each step

proposalsSize ∷ Proposals era → Int

proposalsLookupId ∷ GovActionId → Proposals era → Maybe (GovActionState era)

proposalsActionsMap ∷ Proposals era → Map GovActionId (GovActionState era)

Get the unordered map of GovActionIds and GovActionStates

proposalsWithPurpose ∷ ∀ p era. ToGovActionPurpose p ⇒ (∀ f. Lens' (GovRelation f) (f (GovPurposeId p))) → Proposals era → Map (GovPurposeId p) (GovActionState era)

cgsProposalsL ∷ Lens' (ConwayGovState era) (Proposals era)

cgsDRepPulsingStateL ∷ Lens' (ConwayGovState era) (DRepPulsingState era)

cgsCurPParamsL ∷ Lens' (ConwayGovState era) (PParams era)

cgsPrevPParamsL ∷ Lens' (ConwayGovState era) (PParams era)

cgsFuturePParamsL ∷ Lens' (ConwayGovState era) (FuturePParams era)

cgsCommitteeL ∷ Lens' (ConwayGovState era) (StrictMaybe (Committee era))

cgsConstitutionL ∷ Lens' (ConwayGovState era) (Constitution era)

ensCommitteeL ∷ Lens' (EnactState era) (StrictMaybe (Committee era))

ensConstitutionL ∷ Lens' (EnactState era) (Constitution era)

ensCurPParamsL ∷ Lens' (EnactState era) (PParams era)

ensPrevPParamsL ∷ Lens' (EnactState era) (PParams era)

ensWithdrawalsL ∷ Lens' (EnactState era) (Map (Credential 'Staking) Coin)

ensTreasuryL ∷ Lens' (EnactState era) Coin

ensPrevGovActionIdsL ∷ Lens' (EnactState era) (GovRelation StrictMaybe)

ensPrevPParamUpdateL ∷ Lens' (EnactState era) (StrictMaybe (GovPurposeId 'PParamUpdatePurpose))

ensPrevHardForkL ∷ Lens' (EnactState era) (StrictMaybe (GovPurposeId 'HardForkPurpose))

ensPrevCommitteeL ∷ Lens' (EnactState era) (StrictMaybe (GovPurposeId 'CommitteePurpose))

ensPrevConstitutionL ∷ Lens' (EnactState era) (StrictMaybe (GovPurposeId 'ConstitutionPurpose))

ensProtVerL ∷ EraPParams era ⇒ Lens' (EnactState era) ProtVer

rsEnactStateL ∷ Lens' (RatifyState era) (EnactState era)

rsExpiredL ∷ Lens' (RatifyState era) (Set GovActionId)

rsEnactedL ∷ Lens' (RatifyState era) (Seq (GovActionState era))

rsDelayedL ∷ Lens' (RatifyState era) Bool

constitutionScriptL ∷ Lens' (Constitution era) (StrictMaybe ScriptHash)

constitutionAnchorL ∷ Lens' (Constitution era) Anchor

gasIdL ∷ Lens' (GovActionState era) GovActionId

gasDepositL ∷ Lens' (GovActionState era) Coin

gasCommitteeVotesL ∷ Lens' (GovActionState era) (Map (Credential 'HotCommitteeRole) Vote)

gasDRepVotesL ∷ Lens' (GovActionState era) (Map (Credential 'DRepRole) Vote)

gasStakePoolVotesL ∷ Lens' (GovActionState era) (Map (KeyHash 'StakePool) Vote)

gasExpiresAfterL ∷ Lens' (GovActionState era) EpochNo

gasActionL ∷ Lens' (GovActionState era) (GovAction era)

gasReturnAddrL ∷ Lens' (GovActionState era) RewardAccount

gasProposedInL ∷ Lens' (GovActionState era) EpochNo

gasProposalProcedureL ∷ Lens' (GovActionState era) (ProposalProcedure era)

gasDeposit ∷ GovActionState era → Coin

gasAction ∷ GovActionState era → GovAction era

gasReturnAddr ∷ GovActionState era → RewardAccount

pProcDepositL ∷ Lens' (ProposalProcedure era) Coin

pProcGovActionL ∷ Lens' (ProposalProcedure era) (GovAction era)

pProcReturnAddrL ∷ Lens' (ProposalProcedure era) RewardAccount

pProcAnchorL ∷ Lens' (ProposalProcedure era) Anchor

newEpochStateDRepPulsingStateL ∷ ConwayEraGov era ⇒ Lens' (NewEpochState era) (DRepPulsingState era)

epochStateDRepPulsingStateL ∷ ConwayEraGov era ⇒ Lens' (EpochState era) (DRepPulsingState era)

epochStateStakeDistrL ∷ ∀ era f. Functor f ⇒ (VMap VB VP (Credential 'Staking) (CompactForm Coin) → f (VMap VB VP (Credential 'Staking) (CompactForm Coin))) → EpochState era → f (EpochState era)

epochStateRegDrepL ∷ ConwayEraCertState era ⇒ Lens' (EpochState era) (Map (Credential 'DRepRole) DRepState)

pulseDRepPulsingState ∷ (EraStake era, ConwayEraAccounts era) ⇒ DRepPulsingState era → DRepPulsingState era

completeDRepPulsingState ∷ (EraStake era, ConwayEraAccounts era) ⇒ DRepPulsingState era → DRepPulsingState era

extractDRepPulsingState ∷ (EraStake era, ConwayEraAccounts era) ⇒ DRepPulsingState era → RatifyState era

forceDRepPulsingState ∷ ConwayEraGov era ⇒ NewEpochState era → NewEpochState era

Force computation of DRep stake distribution and figure out the next enact state. This operation is useful in cases when access to new EnactState or DRep stake distribution is needed more than once. It is safe to call this function at any point. Whenever pulser is already in computed state this will be a noop.

finishDRepPulser ∷ (EraStake era, ConwayEraAccounts era) ⇒ DRepPulsingState era → (PulsingSnapshot era, RatifyState era)

computeDRepDistr ∷ (EraStake era, ConwayEraAccounts era) ⇒ InstantStake era → Map (Credential 'DRepRole) DRepState → Map (Credential 'Staking) (CompactForm Coin) → PoolDistr → Map DRep (CompactForm Coin) → Map (Credential 'Staking) (AccountState era) → (Map DRep (CompactForm Coin), PoolDistr)

We iterate over a pulse-sized chunk of the Accounts.

For each staking credential in the chunk that has delegated to a DRep, add the stake distribution, rewards, and proposal deposits for that credential to the DRep distribution, if the DRep is a DRepCredential (also, AlwaysAbstain or AlwaysNoConfidence) and a member of the registered DReps. If the DRepCredential is not a member of the registered DReps, ignore and skip that DRep.

For each staking credential in the chunk that has delegated to an SPO, add only the proposal deposits for that credential to the stake pool distribution, since the rewards and stake are already added to it by the SNAP rule.

Give or take, this operation has roughly O (a * (log(b) + log(c) + log(d) + log(e) + log(f))) complexity, where, (a) is the size of the chunk of the Accounts, which is the pulse-size, iterate over (b) is the size of the StakeDistr, lookup (c) is the size of the DRepDistr, insertWith (d) is the size of the dpProposalDeposits, lookup (e) is the size of the registered DReps, lookup (f) is the size of the PoolDistr, insert

getRatifyState ∷ (ConwayEraAccounts era, EraStake era) ⇒ ConwayGovState era → RatifyState era

conwayGovStateDRepDistrG ∷ (ConwayEraAccounts era, EraStake era) ⇒ SimpleGetter (ConwayGovState era) (Map DRep (CompactForm Coin))

psDRepDistrG ∷ (EraStake era, ConwayEraAccounts era) ⇒ SimpleGetter (DRepPulsingState era) (Map DRep (CompactForm Coin))

This is potentially an expensive getter. Make sure not to use it in the first 80% of the epoch.

data PulsingSnapshot era

A snapshot of information from the previous epoch stored inside the Pulser. After the pulser completes, but before the epoch turns, this information is store in the DRComplete constructor of the DRepPulsingState These are the values at the start of the current epoch. This allows the API To access these "previous" values, both during and after pulsing.

Constructors

PulsingSnapshot

Fields

psProposals ∷ !(StrictSeq (GovActionState era))

psDRepDistr ∷ !(Map DRep (CompactForm Coin))

psDRepState ∷ !(Map (Credential 'DRepRole) DRepState)

psPoolDistr ∷ Map (KeyHash 'StakePool) (CompactForm Coin)

setCompleteDRepPulsingState ∷ ConwayEraGov era ⇒ PulsingSnapshot era → RatifyState era → EpochState era → EpochState era

setFreshDRepPulsingState ∷ (GovState era ~ ConwayGovState era, Monad m, RunConwayRatify era, ConwayEraGov era, ConwayEraCertState era) ⇒ EpochNo → PoolDistr → EpochState era → ReaderT Globals m (EpochState era)

Refresh the pulser in the EpochState using all the new data that is needed to compute the RatifyState when pulsing completes.

psProposalsL ∷ Lens' (PulsingSnapshot era) (StrictSeq (GovActionState era))

psDRepDistrL ∷ Lens' (PulsingSnapshot era) (Map DRep (CompactForm Coin))

psDRepStateL ∷ Lens' (PulsingSnapshot era) (Map (Credential 'DRepRole) DRepState)

psPoolDistrL ∷ Lens' (PulsingSnapshot era) (Map (KeyHash 'StakePool) (CompactForm Coin))

class (STS (ConwayRATIFY era), Signal (ConwayRATIFY era) ~ RatifySignal era, BaseM (ConwayRATIFY era) ~ Reader Globals, Environment (ConwayRATIFY era) ~ RatifyEnv era, State (ConwayRATIFY era) ~ RatifyState era, PredicateFailure (ConwayRATIFY era) ~ Void) ⇒ RunConwayRatify era where

Minimal complete definition

Nothing

Methods

runConwayRatify ∷ Globals → RatifyEnv era → RatifyState era → RatifySignal era → RatifyState era

govStatePrevGovActionIds ∷ ConwayEraGov era ⇒ GovState era → GovRelation StrictMaybe

mkEnactState ∷ ConwayEraGov era ⇒ GovState era → EnactState era

ratifySignalL ∷ Lens' (RatifySignal era) (StrictSeq (GovActionState era))

reDRepDistrL ∷ Lens' (RatifyEnv era) (Map DRep (CompactForm Coin))

reDRepStateL ∷ Lens' (RatifyEnv era) (Map (Credential 'DRepRole) DRepState)

reCurrentEpochL ∷ Lens' (RatifyEnv era) EpochNo

reCommitteeStateL ∷ Lens' (RatifyEnv era) (CommitteeState era)

data DefaultVote

Default vote that will be used for Stake Pool.

Constructors

DefaultNo
Reward account is delegated to a DRepKeyHash, DRepScriptHash or undelegated: default vote is No.

DefaultAbstain
Reward account is delegated to DRepAlwaysAbstain: default vote is Abstain, except for HardForkInitiation actions.

DefaultNoConfidence
Reward account is delegated to DRepAlwaysNoConfidence: default vote is Yes in case of a NoConfidence action, otherwise No.

defaultStakePoolVote

Arguments

∷ ConwayEraAccounts era

⇒ KeyHash 'StakePool
Specify the key hash of the pool whose default vote should be returned.

→ Map (KeyHash 'StakePool) StakePoolState
Registered Stake Pools

→ Accounts era
Delegations of staking credneitals to a DRep

→ DefaultVote

translateProposals ∷ (TranslateEra era GovActionState, TranslationError era GovActionState ~ Void) ⇒ TranslationContext era → Proposals (PreviousEra era) → Proposals era

Exported for testing

pparamsUpdateThreshold ∷ ∀ era. ConwayEraPParams era ⇒ DRepVotingThresholds → PParamsUpdate era → UnitInterval

data TreeMaybe a

Wraper type, which serves as a composition of Tree . StrictMaybe

Also its Show instance will print a nice tree structure.

Constructors

TreeMaybe

Fields

unTreeMaybe ∷ Tree (StrictMaybe a)

toGovRelationTree ∷ HasCallStack ⇒ Proposals era → GovRelation TreeMaybe

Partial version of toGovRelationTreeEither

toGovRelationTreeEither ∷ Proposals era → Either String (GovRelation TreeMaybe)

Convert Proposals into a valid Tree

showGovActionType ∷ GovAction era → String

Orphan instances

EraGov ConwayEra

Instance details

Associated Types

type GovState ConwayEra = (r ∷ Type)

Methods

emptyGovState ∷ GovState ConwayEra

curPParamsGovStateL ∷ Lens' (GovState ConwayEra) (PParams ConwayEra)

prevPParamsGovStateL ∷ Lens' (GovState ConwayEra) (PParams ConwayEra)

futurePParamsGovStateL ∷ Lens' (GovState ConwayEra) (FuturePParams ConwayEra)

obligationGovState ∷ GovState ConwayEra → Obligations