Cardano.Ledger.Conway.Governance.DRepPulser

Documentation

data DRepPulsingState era

Constructors

DRPulsing !(DRepPulser era Identity (RatifyState era))

DRComplete !(PulsingSnapshot era) !(RatifyState era)

data DRepPulser era (m ∷ Type → Type) ans where

The type of a Pulser which uses accumDRepDistr as its underlying function. accumDRepDistr will be partially applied to the components of type (UMap c) and (Map (Credential 'DRepRole c) (DRepState c)) when pulsing. Note that we use two type equality (~) constraints to fix both the monad m and the ans type, to the context where we will use the type as a Pulser. The type DRepPulser must have m and ans as its last two parameters so we can make a Pulsable instance. We will always use this instantiation (DRepPulser era Identity (RatifyState era))

Constructors

DRepPulser

Fields

∷ ∀ era ans m. (ans ~ RatifyState era, m ~ Identity, RunConwayRatify era)

⇒ { dpPulseSize ∷ !Int

How many elements of dpBalance to consume each pulse.

, dpUMap ∷ !(UMap (EraCrypto era))

Snapshot containing the mapping of stake credentials to Pools.

, dpBalance ∷ !(Map (Credential 'Staking (EraCrypto era)) (CompactForm Coin))

The object we are iterating over. Shrinks with each pulse

, dpStakeDistr ∷ !(Map (Credential 'Staking (EraCrypto era)) (CompactForm Coin))

Snapshot of the stake distr (comes from the IncrementalStake)

, dpStakePoolDistr ∷ PoolDistr (EraCrypto era)

Snapshot of the pool distr. Lazy on purpose: See ssStakeMarkPoolDistr and ADR-7 for explanation.

, dpDRepDistr ∷ !(Map (DRep (EraCrypto era)) (CompactForm Coin))

The partial result that grows with each pulse. The purpose of the pulsing.

, dpDRepState ∷ !(Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

Snapshot of registered DRep credentials

, dpCurrentEpoch ∷ !EpochNo

Snapshot of the Epoch this pulser will complete in.

, dpCommitteeState ∷ !(CommitteeState era)

Snapshot of the CommitteeState

, dpEnactState ∷ !(EnactState era)

Snapshot of the EnactState, Used to build the Env of the RATIFY rule

, dpProposals ∷ !(StrictSeq (GovActionState era))

Snapshot of the proposals. This is the Signal for the RATIFY rule

, dpGlobals ∷ !Globals

} → DRepPulser era m ans

pulseDRepPulsingState ∷ DRepPulsingState era → DRepPulsingState era

completeDRepPulsingState ∷ DRepPulsingState era → DRepPulsingState era

extractDRepPulsingState ∷ DRepPulsingState era → RatifyState era

finishDRepPulser ∷ DRepPulsingState era → (PulsingSnapshot era, RatifyState era)

computeDrepDistr ∷ UMap c → Map (Credential 'DRepRole c) (DRepState c) → Map (Credential 'Staking c) (CompactForm Coin) → Map (DRep c) (CompactForm Coin)

Given three inputs

1) Map (Credential 'Staking c) (DRep c). The delegation map. Inside the DRepUView of the UMap um from the DState.

2) regDreps :: Map (Credential 'DRepRole c) (DRepState c). The map of registered DReps to their state. The first part of the VState.

3) stakeDistr :: VMap VB VP (Credential 'Staking c) (CompactForm Coin). The aggregated stake distr extracted from the first component of the IncrementalStake i.e. (IStake credmap _) where credmap is converted to a VMap

Compute the Drep distribution of stake(Coin) cost is expected to be O(size of stakeDistr * log (size of um) * log (size of regDreps)) This is going to be expensive, so we will want to pulse it. Without pulsing, we estimate 3-5 seconds

psDRepDistrG ∷ SimpleGetter (DRepPulsingState era) (Map (DRep (EraCrypto era)) (CompactForm Coin))

This is potentially an expensive getter. Make sure not to use it in the first 80% of the epoch.

dormantEpoch ∷ DRepPulsingState era → Bool

data PulsingSnapshot era

A snapshot of information from the previous epoch stored inside the Pulser. After the pulser completes, but before the epoch turns, this information is store in the DRComplete constructor of the DRepPulsingState These are the values at the start of the current epoch. This allows the API To access these "previous" values, both during and after pulsing.

Constructors

PulsingSnapshot

Fields

psProposals ∷ !(StrictSeq (GovActionState era))

psDRepDistr ∷ !(Map (DRep (EraCrypto era)) (CompactForm Coin))

psDRepState ∷ !(Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

psProposalsL ∷ Lens' (PulsingSnapshot era) (StrictSeq (GovActionState era))

psDRepDistrL ∷ Lens' (PulsingSnapshot era) (Map (DRep (EraCrypto era)) (CompactForm Coin))

psDRepStateL ∷ Lens' (PulsingSnapshot era) (Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

class (STS (ConwayRATIFY era), Signal (ConwayRATIFY era) ~ RatifySignal era, BaseM (ConwayRATIFY era) ~ Reader Globals, Environment (ConwayRATIFY era) ~ RatifyEnv era, State (ConwayRATIFY era) ~ RatifyState era) ⇒ RunConwayRatify era where

Minimal complete definition

Nothing

Methods

runConwayRatify ∷ Globals → RatifyEnv era → RatifyState era → RatifySignal era → RatifyState era