Cardano.Ledger.Conway.Governance.Internal

Documentation

data EnactState era

Constructors

EnactState

Fields

ensCommittee ∷ !(StrictMaybe (Committee era))

Constitutional Committee

ensConstitution ∷ !(Constitution era)

Constitution

ensCurPParams ∷ !(PParams era)

ensPrevPParams ∷ !(PParams era)

ensTreasury ∷ !Coin

ensWithdrawals ∷ !(Map (Credential 'Staking (EraCrypto era)) Coin)

ensPrevGovActionIds ∷ !(GovRelation StrictMaybe era)

Last enacted GovAction Ids

data RatifyState era

Constructors

RatifyState

Fields

rsEnactState ∷ !(EnactState era)

rsEnacted ∷ !(Seq (GovActionState era))

rsExpired ∷ !(Set (GovActionId (EraCrypto era)))

rsDelayed ∷ !Bool

data RatifyEnv era

Constructors

RatifyEnv

Fields

reStakeDistr ∷ !(Map (Credential 'Staking (EraCrypto era)) (CompactForm Coin))

reStakePoolDistr ∷ !(PoolDistr (EraCrypto era))

reDRepDistr ∷ !(Map (DRep (EraCrypto era)) (CompactForm Coin))

reDRepState ∷ !(Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

reCurrentEpoch ∷ !EpochNo

reCommitteeState ∷ !(CommitteeState era)

newtype RatifySignal era

Constructors

RatifySignal (StrictSeq (GovActionState era))

votingStakePoolThreshold ∷ ConwayEraPParams era ⇒ RatifyState era → GovAction era → StrictMaybe UnitInterval

votingDRepThreshold ∷ ConwayEraPParams era ⇒ RatifyState era → GovAction era → StrictMaybe UnitInterval

votingCommitteeThreshold ∷ ConwayEraPParams era ⇒ EpochNo → RatifyState era → CommitteeState era → GovAction era → StrictMaybe UnitInterval

isStakePoolVotingAllowed ∷ ConwayEraPParams era ⇒ GovAction era → Bool

isDRepVotingAllowed ∷ ConwayEraPParams era ⇒ GovAction era → Bool

isCommitteeVotingAllowed ∷ ConwayEraPParams era ⇒ EpochNo → CommitteeState era → GovAction era → Bool

reorderActions ∷ StrictSeq (GovActionState era) → StrictSeq (GovActionState era)

actionPriority ∷ GovAction era → Int

hoistGovRelation ∷ (∀ a. f a → g a) → GovRelation f era → GovRelation g era

withGovActionParent

Arguments

∷ GovActionState era

→ a
The result to be used for governance actions that can't have a parent

→ (∀ p. (∀ f. Lens' (GovRelation f era) (f (GovPurposeId p era))) → StrictMaybe (GovPurposeId p era) → GovPurposeId p era → a)
Function that will be applied to a lens and a parent

→ a

Apply a function to a GovAction that can have a parent.

ensCommitteeL ∷ Lens' (EnactState era) (StrictMaybe (Committee era))

ensConstitutionL ∷ Lens' (EnactState era) (Constitution era)

ensCurPParamsL ∷ Lens' (EnactState era) (PParams era)

ensPrevPParamsL ∷ Lens' (EnactState era) (PParams era)

ensWithdrawalsL ∷ Lens' (EnactState era) (Map (Credential 'Staking (EraCrypto era)) Coin)

ensTreasuryL ∷ Lens' (EnactState era) Coin

ensPrevGovActionIdsL ∷ Lens' (EnactState era) (GovRelation StrictMaybe era)

ensPrevPParamUpdateL ∷ Lens' (EnactState era) (StrictMaybe (GovPurposeId 'PParamUpdatePurpose era))

ensPrevHardForkL ∷ Lens' (EnactState era) (StrictMaybe (GovPurposeId 'HardForkPurpose era))

ensPrevCommitteeL ∷ Lens' (EnactState era) (StrictMaybe (GovPurposeId 'CommitteePurpose era))

ensPrevConstitutionL ∷ Lens' (EnactState era) (StrictMaybe (GovPurposeId 'ConstitutionPurpose era))

ensProtVerL ∷ EraPParams era ⇒ Lens' (EnactState era) ProtVer

rsEnactStateL ∷ Lens' (RatifyState era) (EnactState era)

rsExpiredL ∷ Lens' (RatifyState era) (Set (GovActionId (EraCrypto era)))

rsEnactedL ∷ Lens' (RatifyState era) (Seq (GovActionState era))

rsDelayedL ∷ Lens' (RatifyState era) Bool

epochStateStakeDistrL ∷ Lens' (EpochState era) (VMap VB VP (Credential 'Staking (EraCrypto era)) (CompactForm Coin))

epochStateIncrStakeDistrL ∷ Lens' (EpochState era) (Map (Credential 'Staking (EraCrypto era)) (CompactForm Coin))

epochStateRegDrepL ∷ Lens' (EpochState era) (Map (Credential 'DRepRole (EraCrypto era)) (DRepState (EraCrypto era)))

epochStateUMapL ∷ Lens' (EpochState era) (UMap (EraCrypto era))

reDRepDistrL ∷ Lens' (RatifyEnv era) (Map (DRep (EraCrypto era)) (CompactForm Coin))

Exported for testing

pparamsUpdateThreshold ∷ ∀ era. ConwayEraPParams era ⇒ DRepVotingThresholds → PParamsUpdate era → UnitInterval