Cardano.Ledger.Conway.Governance.Procedures

Documentation

newtype VotingProcedures era

Constructors

VotingProcedures

Fields

unVotingProcedures ∷ Map (Voter (EraCrypto era)) (Map (GovActionId (EraCrypto era)) (VotingProcedure era))

data VotingProcedure era

Constructors

VotingProcedure

Fields

vProcVote ∷ !Vote

vProcAnchor ∷ !(StrictMaybe (Anchor (EraCrypto era)))

foldlVotingProcedures

Arguments

∷ (c → Voter (EraCrypto era) → GovActionId (EraCrypto era) → VotingProcedure era → c)
Accumulating function

→ c
Initial accumulator

→ VotingProcedures era
Procedures to fold over

→ c

foldrVotingProcedures

Arguments

∷ (Voter (EraCrypto era) → GovActionId (EraCrypto era) → VotingProcedure era → c → c)
Accumulating function

→ c
Initial accumulator

→ VotingProcedures era
Procedures to fold over

→ c

data ProposalProcedure era

Constructors

ProposalProcedure

Fields

pProcDeposit ∷ !Coin

pProcReturnAddr ∷ !(RewardAccount (EraCrypto era))

pProcGovAction ∷ !(GovAction era)

pProcAnchor ∷ !(Anchor (EraCrypto era))

data Anchor c

Constructors

Anchor

Fields

anchorUrl ∷ !Url

anchorDataHash ∷ !(SafeHash c AnchorData)

newtype AnchorData

Constructors

AnchorData ByteString

data Vote

Constructors

VoteNo

VoteYes

Abstain

data Voter c

Constructors

CommitteeVoter !(Credential 'HotCommitteeRole c)

DRepVoter !(Credential 'DRepRole c)

StakePoolVoter !(KeyHash 'StakePool c)

data Committee era

Constructors

Committee

Fields

committeeMembers ∷ !(Map (Credential 'ColdCommitteeRole (EraCrypto era)) EpochNo)

Committee members with epoch number when each of them expires

committeeThreshold ∷ !UnitInterval

Threshold of the committee that is necessary for a successful vote

data GovAction era

Note that the previous governance action id is only optional for the very first governance action of the same purpose.

Constructors

ParameterChange

Fields

!(StrictMaybe (GovPurposeId 'PParamUpdatePurpose era))

Previous governance action id of ParameterChange type, which corresponds to PParamUpdatePurpose.

!(PParamsUpdate era)

Proposed changes to PParams

!(StrictMaybe (ScriptHash (EraCrypto era)))

Policy hash protection

HardForkInitiation

Fields

!(StrictMaybe (GovPurposeId 'HardForkPurpose era))

Previous governance action id of HardForkInitiation type, which corresponds to HardForkPurpose

!ProtVer

Proposed new protocol version

TreasuryWithdrawals

Fields

!(Map (RewardAccount (EraCrypto era)) Coin)

Proposed treasury withdrawals

!(StrictMaybe (ScriptHash (EraCrypto era)))

Policy hash protection

NoConfidence

Fields

!(StrictMaybe (GovPurposeId 'CommitteePurpose era))

Previous governance action id of NoConfidence or UpdateCommittee type, which corresponds to CommitteePurpose

UpdateCommittee

Fields

!(StrictMaybe (GovPurposeId 'CommitteePurpose era))

Previous governance action id of UpdateCommittee or NoConfidence type, which corresponds to CommitteePurpose

!(Set (Credential 'ColdCommitteeRole (EraCrypto era)))

Constitutional Committe members to be removed

!(Map (Credential 'ColdCommitteeRole (EraCrypto era)) EpochNo)

Constitutional committee members to be added

!UnitInterval

New Threshold

NewConstitution

Fields

!(StrictMaybe (GovPurposeId 'ConstitutionPurpose era))

Previous governance action id of NewConstitution type, which corresponds to ConstitutionPurpose

!(Constitution era)

InfoAction

data GovActionId c

Constructors

GovActionId

Fields

gaidTxId ∷ !(TxId c)

gaidGovActionIx ∷ !GovActionIx

newtype GovActionIx

Constructors

GovActionIx

Fields

unGovActionIx ∷ Word16

newtype GovPurposeId (p ∷ GovActionPurpose) era

Constructors

GovPurposeId

Fields

unGovPurposeId ∷ GovActionId (EraCrypto era)

data GovActionPurpose

Constructors

PParamUpdatePurpose

HardForkPurpose

CommitteePurpose

ConstitutionPurpose

class ToGovActionPurpose (p ∷ GovActionPurpose)

Minimal complete definition

toGovActionPurpose

isGovActionWithPurpose ∷ ∀ p era. ToGovActionPurpose p ⇒ GovAction era → Bool

data GovRelation (f ∷ Type → Type) era

Abstract data type for representing relationship of governance action with the same purpose

Constructors

GovRelation

Fields

grPParamUpdate ∷ !(f (GovPurposeId 'PParamUpdatePurpose era))

grHardFork ∷ !(f (GovPurposeId 'HardForkPurpose era))

grCommittee ∷ !(f (GovPurposeId 'CommitteePurpose era))

grConstitution ∷ !(f (GovPurposeId 'ConstitutionPurpose era))

grPParamUpdateL ∷ Lens' (GovRelation f era) (f (GovPurposeId 'PParamUpdatePurpose era))

grHardForkL ∷ Lens' (GovRelation f era) (f (GovPurposeId 'HardForkPurpose era))

grCommitteeL ∷ Lens' (GovRelation f era) (f (GovPurposeId 'CommitteePurpose era))

grConstitutionL ∷ Lens' (GovRelation f era) (f (GovPurposeId 'ConstitutionPurpose era))

hoistGovRelation ∷ (∀ a. f a → g a) → GovRelation f era → GovRelation g era

withGovActionParent

Arguments

∷ GovActionState era

→ a
The result to be used for governance actions that can't have a parent

→ (∀ p. (∀ f. Lens' (GovRelation f era) (f (GovPurposeId p era))) → StrictMaybe (GovPurposeId p era) → GovPurposeId p era → a)
Function that will be applied to a lens and a parent

→ a

Apply a function to a GovAction that can have a parent.

data GovActionState era

Constructors

GovActionState

Fields

gasId ∷ !(GovActionId (EraCrypto era))

gasCommitteeVotes ∷ !(Map (Credential 'HotCommitteeRole (EraCrypto era)) Vote)

gasDRepVotes ∷ !(Map (Credential 'DRepRole (EraCrypto era)) Vote)

gasStakePoolVotes ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) Vote)

gasProposalProcedure ∷ !(ProposalProcedure era)

gasProposedIn ∷ !EpochNo

gasExpiresAfter ∷ !EpochNo

govActionIdToText ∷ GovActionId c → Text

indexedGovProps ∷ Seq (ProposalProcedure era) → Seq (GovActionIx, ProposalProcedure era)

Attaches indices to a sequence of proposal procedures. The indices grow from left to right.

data Constitution era

Constructors

Constitution

Fields

constitutionAnchor ∷ !(Anchor (EraCrypto era))

constitutionScript ∷ !(StrictMaybe (ScriptHash (EraCrypto era)))

constitutionAnchorL ∷ Lens' (Constitution era) (Anchor (EraCrypto era))

constitutionScriptL ∷ Lens' (Constitution era) (StrictMaybe (ScriptHash (EraCrypto era)))

pProcDepositL ∷ Lens' (ProposalProcedure era) Coin

pProcGovActionL ∷ Lens' (ProposalProcedure era) (GovAction era)

pProcReturnAddrL ∷ Lens' (ProposalProcedure era) (RewardAccount (EraCrypto era))

pProcAnchorL ∷ Lens' (ProposalProcedure era) (Anchor (EraCrypto era))

committeeMembersL ∷ Lens' (Committee era) (Map (Credential 'ColdCommitteeRole (EraCrypto era)) EpochNo)

committeeThresholdL ∷ Lens' (Committee era) UnitInterval

gasDRepVotesL ∷ Lens' (GovActionState era) (Map (Credential 'DRepRole (EraCrypto era)) Vote)

gasStakePoolVotesL ∷ Lens' (GovActionState era) (Map (KeyHash 'StakePool (EraCrypto era)) Vote)

gasCommitteeVotesL ∷ Lens' (GovActionState era) (Map (Credential 'HotCommitteeRole (EraCrypto era)) Vote)

gasExpiresAfterL ∷ Lens' (GovActionState era) EpochNo

gasProposalProcedureL ∷ Lens' (GovActionState era) (ProposalProcedure era)

gasActionL ∷ Lens' (GovActionState era) (GovAction era)

gasReturnAddrL ∷ Lens' (GovActionState era) (RewardAccount (EraCrypto era))

gasProposedInL ∷ Lens' (GovActionState era) EpochNo

gasIdL ∷ Lens' (GovActionState era) (GovActionId (EraCrypto era))

gasDepositL ∷ Lens' (GovActionState era) Coin

gasDeposit ∷ GovActionState era → Coin

gasAction ∷ GovActionState era → GovAction era

gasReturnAddr ∷ GovActionState era → RewardAccount (EraCrypto era)