Cardano.Ledger.Conway.Governance.Proposals

Description

This module isolates all the types and functionality around Governance Proposals.

It is important to note that there are two sets of state that we maintain around proposals and their enactment. One is processed with transactions (Proposals) and another at the epoch boundary (PrevGovActionIds, DRepPulser). These two sets work together: the incoming proposals and votes are collected continuously in the state that lives with transaction processing, and at each epoch boundary a snapshot of this state is taken to perform pulsing computations on - after it has been adjusted (the proposals are enacted or expired) based on the result of the previous pulsing computaiion.

Below is a typical timeline of the processing of these states.

Epoch n: Proposals and votes are continuously collected from incoming transactions into Proposals

Epoch n boundary: The DRepPulser contains all proposals and votes from epoch (n - 1). Its calculation is completed, ratified and enacted or expired. Ratification and enactment do not affect Proposals directly. They only update the PrevGovActionIds directly and return the sequence of enacted action-ids and the set of expired action-ids that inform us of the changes pending on Proposals.

1. We take this sequence of enacted action-ids and set of expired action-ids and apply them to the Proposals in the ledger state that now includes all the newly collected proposals and votes from epoch n and epoch (n - 1), as this is a superset of the pulsed set of proposals and votes. We do not expect this operation to fail, since all invariants are expected to hold and only an implementation bug could cause this operation to fail. After applying this operation we expect the Proposals to be in a state, where (i) all expired actions and their descendants have been pruned, and (ii) the sequence of enacted action-ids have been promoted to be the root of the respective tree and their competing or sibling action-ids and their descendants have been pruned from the Proposals tree.

2. The resultant Proposals forest has all the latest proposals and votes collected and with enactments and expirations applied to existing ones, so we take a new snapshot to perform pulsing computations on and start the new pulser (DRepPulser), before entering the new epoch to collect more proposals and votes in Proposals. Here we trust that the pulser accounts correctly for newly collected votes on proposals from previous epochs that haven't been ratified yet.

Epoch (n + 1): New proposals and votes are collected from incoming transactions into Proposals.

Epoch (n + 1) boundary: The DRepPulser now contains all unratified proposals and votes from epoch n. Its calculation is completed, ratified and enacted or expired. This updates PrevGovActionIds and gives us a new sequence of enacted action-ids and set of expired actions-ids to apply to the Proposals, which have been collecting even newer proposals and votes to be a superset of our set of pulsed proposals and votes. And so on...

Intended interface to be used for all implementation

data Proposals era

Self-contained representation of all 4 proposals trees. This forest is made up of only action-ids for nodes - full GovActionStates are stored only once in the OMap. All functions in this module prefixed with the string proposals- operate on this data-type keeping it consistent.

NOTE: The correct way to think about this data-structure is similar to 4 of the following, one for each GovActionPurpose

  data Tree a = Node (StrictMaybe a) [Tree a]

but because this does not allow us to look-up a node's edges in predictable time, we use a map from nodes to their edges (parent and children) to capture the graph (PGraph). We also need to always know the roots of the 4 trees, and those we store in the PRoot

NOTE: At the end of an epoch boundary, we expect pRoots to be the same as the PrevGovActionIds from the EnactState

mapProposals ∷ (GovActionState era → GovActionState era) → Proposals era → Proposals era

Make sure not to change the gasId, otherwise all hell will break loose.

proposalsIds ∷ Proposals era → StrictSeq (GovActionId (EraCrypto era))

Get the sequence of GovActionIds

proposalsActions ∷ Proposals era → StrictSeq (GovActionState era)

Get the sequence of GovActionStates

proposalsSize ∷ Proposals era → Int

proposalsAddAction ∷ ∀ era. (EraPParams era, HasCallStack) ⇒ GovActionState era → Proposals era → Maybe (Proposals era)

Add a single GovActionState to the Proposals forest. The tree to which it is added is picked according to its GovActionPurpose. Returns Nothing when the operation cannot succeed.

proposalsApplyEnactment ∷ ∀ era. EraPParams era ⇒ Seq (GovActionState era) → Set (GovActionId (EraCrypto era)) → Proposals era → (Proposals era, Map (GovActionId (EraCrypto era)) (GovActionState era), Map (GovActionId (EraCrypto era)) (GovActionState era))

For use in the EPOCH rule. Apply the result of extractDRepPulsingState to the Proposals forest, so that: i. all the expired action-ids and their descendants are removed, and ii. the sequence of enacted action-ids is promoted to the root, removing competing/sibling action-ids and their descendants at each step

proposalsAddVote ∷ Voter (EraCrypto era) → Vote → GovActionId (EraCrypto era) → Proposals era → Proposals era

Add a vote to an existing GovActionState. This is a no-op if the provided GovActionId does not already exist

proposalsLookupId ∷ GovActionId (EraCrypto era) → Proposals era → Maybe (GovActionState era)

proposalsActionsMap ∷ Proposals era → Map (GovActionId (EraCrypto era)) (GovActionState era)

Get the unordered map of GovActionIds and GovActionStates

proposalsWithPurpose ∷ ∀ p era. ToGovActionPurpose p ⇒ (∀ f. Lens' (GovRelation f era) (f (GovPurposeId p era))) → Proposals era → Map (GovPurposeId p era) (GovActionState era)

toPrevGovActionIds ∷ GovRelation PRoot era → GovRelation StrictMaybe era

fromPrevGovActionIds ∷ GovRelation StrictMaybe era → GovRelation PRoot era

To be used only for testing

proposalsRemoveWithDescendants ∷ EraPParams era ⇒ Set (GovActionId (EraCrypto era)) → Proposals era → (Proposals era, Map (GovActionId (EraCrypto era)) (GovActionState era))

Remove the set of given action-ids with their descendants from the Proposals forest. Cannot be used for removing enacted GovActionIds (i.e. roots)

data TreeMaybe a

Wraper type, which serves as a composition of Tree . StrictMaybe

Also its Show instance will print a nice tree structure.

Constructors

TreeMaybe

Fields

unTreeMaybe ∷ Tree (StrictMaybe a)

toGovRelationTree ∷ (Era era, HasCallStack) ⇒ Proposals era → GovRelation TreeMaybe era

Partial version of toGovRelationTreeEither

toGovRelationTreeEither ∷ Era era ⇒ Proposals era → Either String (GovRelation TreeMaybe era)

Convert Proposals into a valid Tree

pPropsL ∷ Lens' (Proposals era) (OMap (GovActionId (EraCrypto era)) (GovActionState era))

pRootsL ∷ Lens' (Proposals era) (GovRelation PRoot era)

pGraphL ∷ Lens' (Proposals era) (GovRelation PGraph era)

mkProposals ∷ (EraPParams era, MonadFail m) ⇒ GovRelation StrictMaybe era → OMap (GovActionId (EraCrypto era)) (GovActionState era) → m (Proposals era)

Reconstruct the Proposals forest from an OMap of GovActionStates and the 4 roots (PrevGovActionIds)

unsafeMkProposals ∷ HasCallStack ⇒ GovRelation StrictMaybe era → OMap (GovActionId (EraCrypto era)) (GovActionState era) → Proposals era

Reconstruct the Proposals forest from an OMap of GovActionStates and the 4 roots (PrevGovActionIds). This function can fail and may return a malformed Proposals if not given correct inputs.

WARNING: Should only be used for testing!

data PRoot a

The root of a single Proposals tree. prRoot is always expected to be equal to the respective PrevGovActionId at the end of every epoch boundary

Constructors

PRoot

Fields

prRoot ∷ !(StrictMaybe a)

prChildren ∷ !(Set a)

prRootL ∷ Lens' (PRoot a) (StrictMaybe a)

prChildrenL ∷ Lens' (PRoot a) (Set a)

data PEdges a

A non-root edges in a Proposals tree. peParent is expected to be a SNothing only at the begining when no governance actions has been enacted yet.

Constructors

PEdges

Fields

peParent ∷ !(StrictMaybe a)

peChildren ∷ !(Set a)

peChildrenL ∷ Lens' (PEdges a) (Set a)

newtype PGraph a

A single proposal-tree. This map represents all the action-ids that form a tree.

Constructors

PGraph

Fields

unPGraph ∷ Map a (PEdges a)

pGraphNodesL ∷ Lens' (PGraph a) (Map a (PEdges a))

proposalsDeposits ∷ Proposals era → Map (Credential 'Staking (EraCrypto era)) (CompactForm Coin)

Get a mapping from the reward-account staking credentials to deposits of all proposals.