Cardano.Ledger.Conway.PParams

Description

This module contains the type of protocol parameters and EraPParams instance

Documentation

data BabbagePParams (f ∷ Type → Type) era

Babbage Protocol parameters. Ways in which parameters have changed from Alonzo: lack of d, extraEntropy and replacement of coinsPerUTxOWord with coinsPerUTxOByte

Constructors

BabbagePParams

Fields

bppMinFeeA ∷ !(HKD f Coin)

The linear factor for the minimum fee calculation

bppMinFeeB ∷ !(HKD f Coin)

The constant factor for the minimum fee calculation

bppMaxBBSize ∷ !(HKD f Natural)

Maximal block body size

bppMaxTxSize ∷ !(HKD f Natural)

Maximal transaction size

bppMaxBHSize ∷ !(HKD f Natural)

Maximal block header size

bppKeyDeposit ∷ !(HKD f Coin)

The amount of a key registration deposit

bppPoolDeposit ∷ !(HKD f Coin)

The amount of a pool registration deposit

bppEMax ∷ !(HKD f EpochNo)

Maximum number of epochs in the future a pool retirement is allowed to be scheduled for.

bppNOpt ∷ !(HKD f Natural)

Desired number of pools

bppA0 ∷ !(HKD f NonNegativeInterval)

Pool influence

bppRho ∷ !(HKD f UnitInterval)

Monetary expansion

bppTau ∷ !(HKD f UnitInterval)

Treasury expansion

bppProtocolVersion ∷ !(HKD f ProtVer)

Protocol version

bppMinPoolCost ∷ !(HKD f Coin)

Minimum Stake Pool Cost

bppCoinsPerUTxOByte ∷ !(HKD f CoinPerByte)

Cost in lovelace per byte of UTxO storage (instead of bppCoinsPerUTxOByte)

bppCostModels ∷ !(HKD f CostModels)

Cost models for non-native script languages

bppPrices ∷ !(HKD f Prices)

Prices of execution units (for non-native script languages)

bppMaxTxExUnits ∷ !(HKD f OrdExUnits)

Max total script execution resources units allowed per tx

bppMaxBlockExUnits ∷ !(HKD f OrdExUnits)

Max total script execution resources units allowed per block

bppMaxValSize ∷ !(HKD f Natural)

Max size of a Value in an output

bppCollateralPercentage ∷ !(HKD f Natural)

Percentage of the txfee which must be provided as collateral when including non-native scripts.

bppMaxCollateralInputs ∷ !(HKD f Natural)

Maximum number of collateral inputs allowed in a transaction

data ConwayPParams f era

Conway Protocol parameters. The following parameters have been added since Babbage: * poolVotingThresholds * dRepVotingThresholds * minCommitteeSize * committeeTermLimit * govActionExpiration * govActionDeposit * dRepDeposit * dRepActivity

Constructors

ConwayPParams

Fields

cppMinFeeA ∷ !(HKD f Coin)

The linear factor for the minimum fee calculation

cppMinFeeB ∷ !(HKD f Coin)

The constant factor for the minimum fee calculation

cppMaxBBSize ∷ !(HKD f Natural)

Maximal block body size

cppMaxTxSize ∷ !(HKD f Natural)

Maximal transaction size

cppMaxBHSize ∷ !(HKD f Natural)

Maximal block header size

cppKeyDeposit ∷ !(HKD f Coin)

The amount of a key registration deposit

cppPoolDeposit ∷ !(HKD f Coin)

The amount of a pool registration deposit

cppEMax ∷ !(HKD f EpochNo)

Maximum number of epochs in the future a pool retirement is allowed to be scheduled for.

cppNOpt ∷ !(HKD f Natural)

Desired number of pools

cppA0 ∷ !(HKD f NonNegativeInterval)

Pool influence

cppRho ∷ !(HKD f UnitInterval)

Monetary expansion

cppTau ∷ !(HKD f UnitInterval)

Treasury expansion

cppProtocolVersion ∷ !(HKD f ProtVer)

Protocol version

cppMinPoolCost ∷ !(HKD f Coin)

Minimum Stake Pool Cost

cppCoinsPerUTxOByte ∷ !(HKD f CoinPerByte)

Cost in lovelace per byte of UTxO storage

cppCostModels ∷ !(HKD f CostModels)

Cost models for non-native script languages

cppPrices ∷ !(HKD f Prices)

Prices of execution units (for non-native script languages)

cppMaxTxExUnits ∷ !(HKD f OrdExUnits)

Max total script execution resources units allowed per tx

cppMaxBlockExUnits ∷ !(HKD f OrdExUnits)

Max total script execution resources units allowed per block

cppMaxValSize ∷ !(HKD f Natural)

Max size of a Value in an output

cppCollateralPercentage ∷ !(HKD f Natural)

Percentage of the txfee which must be provided as collateral when including non-native scripts.

cppMaxCollateralInputs ∷ !(HKD f Natural)

Maximum number of collateral inputs allowed in a transaction

New ones for Conway

cppPoolVotingThresholds ∷ !(HKD f PoolVotingThresholds)

Thresholds for SPO votes

cppDRepVotingThresholds ∷ !(HKD f DRepVotingThresholds)

Thresholds for DRep votes

cppMinCommitteeSize ∷ !(HKD f Natural)

Minimum size of the Constitutional Committee

cppCommitteeTermLimit ∷ !(HKD f Natural)

The Constitutional Committee Term limit in number of Slots

cppGovActionExpiration ∷ !(HKD f Natural)

Governance action expiration in number of Slots

cppGovActionDeposit ∷ !(HKD f Coin)

The amount of the Governance Action deposit

cppDRepDeposit ∷ !(HKD f Coin)

The amount of a DRep registration deposit

cppDRepActivity ∷ !(HKD f EpochNo)

The number of Epochs that a DRep can perform no activity without losing their Active status.

getLanguageView ∷ AlonzoEraPParams era ⇒ PParams era → Language → LangDepView

data LangDepView

Constructors

LangDepView

Fields

tag ∷ ByteString

params ∷ ByteString

encodeLangViews ∷ Set LangDepView → Encoding

upgradeConwayPParams ∷ ∀ f c. UpgradeConwayPParams f → PParamsHKD f (BabbageEra c) → ConwayPParams f (ConwayEra c)

data UpgradeConwayPParams f

Constructors

UpgradeConwayPParams

Fields

ucppPoolVotingThresholds ∷ !(HKD f PoolVotingThresholds)

ucppDRepVotingThresholds ∷ !(HKD f DRepVotingThresholds)

ucppMinCommitteeSize ∷ !(HKD f Natural)

ucppCommitteeTermLimit ∷ !(HKD f Natural)

ucppGovActionExpiration ∷ !(HKD f Natural)

ucppGovActionDeposit ∷ !(HKD f Coin)

ucppDRepDeposit ∷ !(HKD f Coin)

ucppDRepActivity ∷ !(HKD f EpochNo)

data PoolVotingThresholds

Constructors

PoolVotingThresholds

Fields

pvtMotionNoConfidence ∷ !UnitInterval

pvtCommitteeNormal ∷ !UnitInterval

pvtCommitteeNoConfidence ∷ !UnitInterval

pvtHardForkInitiation ∷ !UnitInterval

data DRepVotingThresholds

Constructors

DRepVotingThresholds

Fields

dvtMotionNoConfidence ∷ !UnitInterval

dvtCommitteeNormal ∷ !UnitInterval

dvtCommitteeNoConfidence ∷ !UnitInterval

dvtUpdateToConstitution ∷ !UnitInterval

dvtHardForkInitiation ∷ !UnitInterval

dvtPPNetworkGroup ∷ !UnitInterval

dvtPPEconomicGroup ∷ !UnitInterval

dvtPPTechnicalGroup ∷ !UnitInterval

dvtPPGovernanceGroup ∷ !UnitInterval

dvtTreasuryWithdrawal ∷ !UnitInterval

Orphan instances

FromJSON DRepVotingThresholds

Instance details

Methods

parseJSON ∷ Value → Parser DRepVotingThresholds

parseJSONList ∷ Value → Parser [DRepVotingThresholds]

FromJSON PoolVotingThresholds

Instance details

Methods

parseJSON ∷ Value → Parser PoolVotingThresholds

parseJSONList ∷ Value → Parser [PoolVotingThresholds]

Crypto c ⇒ AlonzoEraPParams (ConwayEra c)

Instance details

Methods

hkdCoinsPerUTxOWordL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ExactEra AlonzoEra (ConwayEra c)) ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f CoinPerWord)

hkdCostModelsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f CostModels)

hkdPricesL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Prices)

hkdMaxTxExUnitsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f ExUnits)

hkdMaxBlockExUnitsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f ExUnits)

hkdMaxValSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

hkdCollateralPercentageL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

hkdMaxCollateralInputsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

Crypto c ⇒ BabbageEraPParams (ConwayEra c)

Instance details

Methods

hkdCoinsPerUTxOByteL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f CoinPerByte)

Crypto c ⇒ ConwayEraPParams (ConwayEra c)

Instance details

Methods

hkdPoolVotingThresholdsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f PoolVotingThresholds)

hkdDRepVotingThresholdsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f DRepVotingThresholds)

hkdMinCommitteeSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

hkdCommitteeTermLimitL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

hkdGovActionExpirationL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

hkdGovActionDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)

hkdDRepDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)

hkdDRepActivityL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f EpochNo)

Crypto c ⇒ EraPParams (ConwayEra c)

Instance details

Associated Types

type PParamsHKD f (ConwayEra c) = (r ∷ Type)

type UpgradePParams f (ConwayEra c)

type DowngradePParams f (ConwayEra c)

Methods

applyPPUpdates ∷ PParams (ConwayEra c) → PParamsUpdate (ConwayEra c) → PParams (ConwayEra c)

emptyPParamsIdentity ∷ PParamsHKD Identity (ConwayEra c)

emptyPParamsStrictMaybe ∷ PParamsHKD StrictMaybe (ConwayEra c)

upgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, EraPParams (PreviousEra (ConwayEra c))) ⇒ UpgradePParams f (ConwayEra c) → PParamsHKD f (PreviousEra (ConwayEra c)) → PParamsHKD f (ConwayEra c)

downgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, EraPParams (PreviousEra (ConwayEra c))) ⇒ DowngradePParams f (ConwayEra c) → PParamsHKD f (ConwayEra c) → PParamsHKD f (PreviousEra (ConwayEra c))

hkdMinFeeAL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)

hkdMinFeeBL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)

hkdMaxBBSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

hkdMaxTxSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

hkdMaxBHSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

hkdKeyDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)

hkdPoolDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)

hkdEMaxL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f EpochNo)

hkdNOptL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

hkdA0L ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f NonNegativeInterval)

hkdRhoL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f UnitInterval)

hkdTauL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f UnitInterval)

hkdDL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost (ConwayEra c) 6) ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f UnitInterval)

ppDG ∷ SimpleGetter (PParams (ConwayEra c)) UnitInterval

hkdExtraEntropyL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost (ConwayEra c) 6) ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Nonce)

hkdProtocolVersionL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f ProtVer)

hkdMinUTxOValueL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost (ConwayEra c) 4) ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)

hkdMinPoolCostL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)