Cardano.Ledger.Conway.Plutus.Context

Documentation

pparamUpdateToData ∷ Map Word (PParam era) → PParamsUpdate era → Data

Translate a PParamUpdate to Data, using the PParamMap. The PParamUpdate is Basically (Map (tag t) (Some (StrictMaybe t))) The idea is to only store (Map (tag t) (Some t)) for just the components that had (SJust t), and to leave out the other tags. To transform this back to a (Map (tag t) (Some t)), we start with a full map where everything is SNothing, and then override the tags we find in the data with (SJust t) That is what happens in pparamUpdateFromData below

pparamUpdateFromData ∷ EraPParams era ⇒ Map Word (PParam era) → Data → Maybe (PParamsUpdate era)

Translate Data to a PParamUpdate, using the PParamMap.

conwayPParam ∷ ConwayEraPParams era ⇒ [PParam era]

A list for the Conway era, other eras may have different tags and lenses.

conwayPParamMap ∷ ConwayEraPParams era ⇒ Map Word (PParam era)

A Map for the Conway era

class (ToPlutusData (PParamsUpdate era), EraPlutusTxInfo l era) ⇒ ConwayEraPlutusTxInfo (l ∷ Language) era where

Methods

toPlutusChangedParameters ∷ proxy l → PParamsUpdate era → ChangedParameters

Orphan instances

ToPlutusData CoinPerByte

Instance details

Methods

toPlutusData ∷ CoinPerByte → Data

fromPlutusData ∷ Data → Maybe CoinPerByte

ToPlutusData DRepVotingThresholds

Instance details

Methods

toPlutusData ∷ DRepVotingThresholds → Data

fromPlutusData ∷ Data → Maybe DRepVotingThresholds

ToPlutusData PoolVotingThresholds

Instance details

Methods

toPlutusData ∷ PoolVotingThresholds → Data

fromPlutusData ∷ Data → Maybe PoolVotingThresholds