Cardano.Ledger.Conway.Rules

Documentation

data ConwayBBODY era

data ConwayBbodyPredFailure era

Constructors

WrongBlockBodySizeBBODY (Mismatch 'RelEQ Int)

InvalidBodyHashBBODY (Mismatch 'RelEQ (Hash HASH EraIndependentBlockBody))

LedgersFailure (PredicateFailure (EraRule "LEDGERS" era))
LEDGERS rule subtransition Failures

TooManyExUnits (Mismatch 'RelLTEQ ExUnits)

BodyRefScriptsSizeTooBig (Mismatch 'RelLTEQ Int)

alonzoToConwayBbodyPredFailure ∷ ∀ era. AlonzoBbodyPredFailure era → ConwayBbodyPredFailure era

shelleyToConwayBbodyPredFailure ∷ ∀ era. ShelleyBbodyPredFailure era → ConwayBbodyPredFailure era

totalRefScriptSizeInBlock ∷ (AlonzoEraTx era, BabbageEraTxBody era) ⇒ ProtVer → StrictSeq (Tx era) → UTxO era → Int

data ConwayCERT era

data ConwayCertPredFailure era

Constructors

DelegFailure (PredicateFailure (EraRule "DELEG" era))

PoolFailure (PredicateFailure (EraRule "POOL" era))

GovCertFailure (PredicateFailure (EraRule "GOVCERT" era))

data ConwayCertEvent era

Constructors

DelegEvent (Event (EraRule "DELEG" era))

PoolEvent (Event (EraRule "POOL" era))

GovCertEvent (Event (EraRule "GOVCERT" era))

data CertEnv era

Constructors

CertEnv

Fields

cePParams ∷ PParams era

ceCurrentEpoch ∷ EpochNo

Lazy on purpose, because not all certificates need to know the current EpochNo

ceCurrentCommittee ∷ StrictMaybe (Committee era)

ceCommitteeProposals ∷ Map (GovPurposeId 'CommitteePurpose) (GovActionState era)

data ConwayDELEG era

data ConwayDelegPredFailure era

Constructors

IncorrectDepositDELEG Coin

StakeKeyRegisteredDELEG (Credential 'Staking)

StakeKeyNotRegisteredDELEG (Credential 'Staking)

StakeKeyHasNonZeroRewardAccountBalanceDELEG Coin

DelegateeDRepNotRegisteredDELEG (Credential 'DRepRole)

DelegateeStakePoolNotRegisteredDELEG (KeyHash 'StakePool)

DepositIncorrectDELEG (Mismatch 'RelEQ Coin)

RefundIncorrectDELEG (Mismatch 'RelEQ Coin)

data ConwayDelegEnv era

Constructors

ConwayDelegEnv

Fields

cdePParams ∷ PParams era

cdePools ∷ Map (KeyHash 'StakePool) StakePoolState

processDelegation

Arguments

∷ ConwayEraCertState era

⇒ Credential 'Staking
Delegator

→ Delegatee
New delegatee

→ CertState era

→ CertState era

Apply new delegation, while properly cleaning up older delegations. This function does not enforce that delegatee is registered, that has to be handled by the caller.

data ConwayGOVCERT era

data ConwayGovCertPredFailure era

Constructors

ConwayDRepAlreadyRegistered (Credential 'DRepRole)

ConwayDRepNotRegistered (Credential 'DRepRole)

ConwayDRepIncorrectDeposit (Mismatch 'RelEQ Coin)

ConwayCommitteeHasPreviouslyResigned (Credential 'ColdCommitteeRole)

ConwayDRepIncorrectRefund (Mismatch 'RelEQ Coin)

ConwayCommitteeIsUnknown (Credential 'ColdCommitteeRole)
Predicate failure whenever an update to an unknown committee member is attempted. Current Constitutional Committee and all available proposals will be searched before reporting this predicate failure.

data ConwayGovCertEnv era

Constructors

ConwayGovCertEnv

Fields

cgcePParams ∷ PParams era

cgceCurrentEpoch ∷ EpochNo

Lazy on purpose, because not all certificates need to know the current EpochNo

cgceCurrentCommittee ∷ StrictMaybe (Committee era)

cgceCommitteeProposals ∷ Map (GovPurposeId 'CommitteePurpose) (GovActionState era)

All of the UpdateCommittee proposals

computeDRepExpiry

Arguments

∷ EpochInterval
DRepActivity PParam

→ EpochNo
Current epoch

→ EpochNo
The count of the dormant epochs

→ EpochNo
Computed expiry

data ConwayCERTS era

data ConwayCertsPredFailure era

Constructors

WithdrawalsNotInRewardsCERTS Withdrawals
Withdrawals that are missing or do not withdraw the entire amount (pv < 11)

CertFailure (PredicateFailure (EraRule "CERT" era))
CERT rule subtransition Failures

newtype ConwayCertsEvent era

Constructors

CertEvent (Event (EraRule "CERT" era))

data CertsEnv era

Constructors

CertsEnv

Fields

certsTx ∷ Tx era

certsPParams ∷ PParams era

certsCurrentEpoch ∷ EpochNo

Lazy on purpose, because not all certificates need to know the current EpochNo

certsCurrentCommittee ∷ StrictMaybe (Committee era)

certsCommitteeProposals ∷ Map (GovPurposeId 'CommitteePurpose) (GovActionState era)

updateDormantDRepExpiry

Arguments

∷ EpochNo
Current Epoch

→ VState era

→ VState era

Update dormant expiry for all DReps that are active. And also reset the numDormantEpochs counter.

updateDormantDRepExpiries ∷ (EraTx era, ConwayEraTxBody era, ConwayEraCertState era) ⇒ Tx era → EpochNo → CertState era → CertState era

If there is a new governance proposal to vote on in this transaction, AND the number of dormant-epochs recorded is greater than zero, we bump the expiry for all DReps by the number of dormant epochs, and reset the counter to zero.

It does not matter that this is called _before_ the GOV rule in LEDGER, even though we cannot validate any governance proposal here, since the entire transaction will fail if the proposal is not accepted in GOV, and so will this expiry bump done here.

updateVotingDRepExpiries ∷ (EraTx era, ConwayEraTxBody era, ConwayEraCertState era) ⇒ Tx era → EpochNo → EpochInterval → CertState era → CertState era

Update DRep expiry for all DReps that are voting in this transaction. This will execute in mutual-exclusion to the updates to the dormant DRep expiry, because if there are no proposals to vote on, there will be no votes either.

data ConwayENACT era

data EnactSignal era

Constructors

EnactSignal

Fields

esGovActionId ∷ !GovActionId

esGovAction ∷ !(GovAction era)

data EnactState era

Constructors

EnactState

Fields

ensCommittee ∷ !(StrictMaybe (Committee era))

Constitutional Committee

ensConstitution ∷ !(Constitution era)

Constitution

ensCurPParams ∷ !(PParams era)

ensPrevPParams ∷ !(PParams era)

ensTreasury ∷ !Coin

ensWithdrawals ∷ !(Map (Credential 'Staking) Coin)

ensPrevGovActionIds ∷ !(GovRelation StrictMaybe)

Last enacted GovAction Ids

data ConwayEPOCH era

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ConwayEpochEvent era

Constructors

PoolReapEvent (Event (EraRule "POOLREAP" era))

SnapEvent (Event (EraRule "SNAP" era))

EpochBoundaryRatifyState (RatifyState era)

GovInfoEvent

Fields

(Set (GovActionState era))

Enacted actions

(Set (GovActionState era))

Actions that were removed as conflicting due to enactment

(Set (GovActionState era))

Actions that were removed due to expiration together with their dependees

(Map GovActionId Coin)

Map of removed governance action ids that had an unregistered reward account to their unclaimed deposits so they can be transferred to the treasury.

HardForkEvent (Event (EraRule "HARDFORK" era))

data ConwayHARDFORK era

newtype ConwayHardForkEvent era

Constructors

ConwayHardForkEvent ProtVer

data ConwayLEDGER era

data ConwayLedgerPredFailure era

Constructors

ConwayUtxowFailure (PredicateFailure (EraRule "UTXOW" era))

ConwayCertsFailure (PredicateFailure (EraRule "CERTS" era))

ConwayGovFailure (PredicateFailure (EraRule "GOV" era))

ConwayWdrlNotDelegatedToDRep (NonEmpty (KeyHash 'Staking))

ConwayTreasuryValueMismatch (Mismatch 'RelEQ Coin)

ConwayTxRefScriptsSizeTooBig (Mismatch 'RelLTEQ Int)

ConwayMempoolFailure Text

ConwayWithdrawalsMissingAccounts Withdrawals

ConwayIncompleteWithdrawals Withdrawals

data ConwayLedgerEvent era

Constructors

UtxowEvent (Event (EraRule "UTXOW" era))

CertsEvent (Event (EraRule "CERTS" era))

GovEvent (Event (EraRule "GOV" era))

data ConwayMEMPOOL era

data ConwayNEWEPOCH era

data ConwayNewEpochEvent era

Constructors

DeltaRewardEvent !(Event (EraRule "RUPD" era))

RestrainedRewards !EpochNo !(Map (Credential 'Staking) (Set Reward)) !(Set (Credential 'Staking))

TotalRewardEvent !EpochNo !(Map (Credential 'Staking) (Set Reward))

EpochEvent !(Event (EraRule "EPOCH" era))

TotalAdaPotsEvent !AdaPots

data ConwayTICKF era

data ConwayTickfPredFailure era

data ConwayTickfEvent era

data ConwayRATIFY era

data RatifyState era

RatifyState stores information about what will happen to the active governance actions at the next epoch boundary.

Constructors

RatifyState

Fields

rsEnactState ∷ !(EnactState era)

This is the currently active EnactState. It contains all the changes that were applied to it at the last epoch boundary by all the proposals that were enacted.

rsEnacted ∷ !(Seq (GovActionState era))

Governance actions that are going to be enacted at the next epoch boundary.

rsExpired ∷ !(Set GovActionId)

Governance actions that are going to be removed at the next epoch boundary, either due to expiring or because they would become invalid after another governance action gets enacted or expired before it

rsDelayed ∷ !Bool

This flag is set to true if one of the proposals that was ratified at the last epoch boundary was a delaying action. This means that no other proposals will be ratified this epoch and each active proposal that has not become invalid will have its expiry date extended by one epoch.

This flag is reset at each epoch boundary before the RATIFY rule gets called, but it might immediately be set to True again after the RATIFY rule has finished execution.

committeeAccepted ∷ ConwayEraPParams era ⇒ RatifyEnv era → RatifyState era → GovActionState era → Bool

committeeAcceptedRatio ∷ ∀ era. Map (Credential 'ColdCommitteeRole) EpochNo → Map (Credential 'HotCommitteeRole) Vote → CommitteeState era → EpochNo → Rational

spoAccepted ∷ (ConwayEraPParams era, ConwayEraAccounts era) ⇒ RatifyEnv era → RatifyState era → GovActionState era → Bool

spoAcceptedRatio ∷ ConwayEraAccounts era ⇒ RatifyEnv era → GovActionState era → ProtVer → Rational

Final ratio for totalAcceptedStakePoolsRatio we want during the bootstrap period is: t = y / (s - a) Where: * y - total delegated stake that voted Yes * a - total delegated stake that voted Abstain * s - total delegated stake

For HardForkInitiation all SPOs that didn't vote are considered as No votes. Whereas, for all other GovActions, SPOs that didn't vote are considered as Abstain votes.

No votes are not counted. After the bootstrap period if an SPO didn't vote, it will be considered as a No vote by default. The only exceptions are when a pool delegated to an AlwaysNoConfidence or an AlwaysAbstain DRep. In those cases, behaviour is as expected: vote Yes on NoConfidence proposals in case of the former and and vote Abstain by default in case of the latter. For HardForkInitiation, behaviour is the same as during the bootstrap period: if an SPO didn't vote, their vote will always count as No.

dRepAccepted ∷ ConwayEraPParams era ⇒ RatifyEnv era → RatifyState era → GovActionState era → Bool

dRepAcceptedRatio ∷ ∀ era. RatifyEnv era → Map (Credential 'DRepRole) Vote → GovAction era → Rational

acceptedByEveryone ∷ (ConwayEraPParams era, ConwayEraAccounts era) ⇒ RatifyEnv era → RatifyState era → GovActionState era → Bool

prevActionAsExpected ∷ GovActionState era → GovRelation StrictMaybe → Bool

Check that the previous governance action id specified in the proposal does match the last one of the same purpose that was enacted.

validCommitteeTerm ∷ ConwayEraPParams era ⇒ GovAction era → PParams era → EpochNo → Bool

withdrawalCanWithdraw ∷ GovAction era → Coin → Bool

data ConwayGOV era

data GovEnv era

Constructors

GovEnv

Fields

geTxId ∷ TxId

geEpoch ∷ EpochNo

gePParams ∷ PParams era

gePPolicy ∷ StrictMaybe ScriptHash

geCertState ∷ CertState era

geCommittee ∷ StrictMaybe (Committee era)

data GovSignal era

Constructors

GovSignal

Fields

gsVotingProcedures ∷ !(VotingProcedures era)

gsProposalProcedures ∷ !(OSet (ProposalProcedure era))

gsCertificates ∷ !(StrictSeq (TxCert era))

data ConwayGovEvent era

Constructors

GovNewProposals !TxId !(Proposals era)

GovRemovedVotes

Fields

!TxId

!(Set (Voter, GovActionId))

Votes that were replaced in this tx.

!(Set (Credential 'DRepRole))

Any votes from these DReps in this or in previous txs are removed

data ConwayGovPredFailure era

Constructors

GovActionsDoNotExist (NonEmpty GovActionId)

MalformedProposal (GovAction era)

ProposalProcedureNetworkIdMismatch RewardAccount Network

TreasuryWithdrawalsNetworkIdMismatch (Set RewardAccount) Network

ProposalDepositIncorrect (Mismatch 'RelEQ Coin)

DisallowedVoters (NonEmpty (Voter, GovActionId))
Some governance actions are not allowed to be voted on by certain types of Voters. This failure lists all governance action ids with their respective voters that are not allowed to vote on those governance actions.

ConflictingCommitteeUpdate

Fields

(Set (Credential 'ColdCommitteeRole))

Credentials that are mentioned as members to be both removed and added

ExpirationEpochTooSmall

Fields

(Map (Credential 'ColdCommitteeRole) EpochNo)

Members for which the expiration epoch has already been reached

InvalidPrevGovActionId (ProposalProcedure era)

VotingOnExpiredGovAction (NonEmpty (Voter, GovActionId))

ProposalCantFollow

Fields

(StrictMaybe (GovPurposeId 'HardForkPurpose))

The PrevGovActionId of the HardForkInitiation that fails

(Mismatch 'RelGT ProtVer)

Its protocol version and the protocal version of the previous gov-action pointed to by the proposal

InvalidPolicyHash

Fields

(StrictMaybe ScriptHash)

The policy script hash in the proposal

(StrictMaybe ScriptHash)

The policy script hash of the current constitution

DisallowedProposalDuringBootstrap (ProposalProcedure era)

DisallowedVotesDuringBootstrap (NonEmpty (Voter, GovActionId))

VotersDoNotExist (NonEmpty Voter)
Predicate failure for votes by entities that are not present in the ledger state

ZeroTreasuryWithdrawals (GovAction era)
Treasury withdrawals that sum up to zero are not allowed

ProposalReturnAccountDoesNotExist RewardAccount
Proposals that have an invalid reward account for returns of the deposit

TreasuryWithdrawalReturnAccountsDoNotExist (NonEmpty RewardAccount)
Treasury withdrawal proposals to an invalid reward account

UnelectedCommitteeVoters (NonEmpty (Credential 'HotCommitteeRole))
Disallow votes by unelected committee members

unelectedCommitteeVoters ∷ StrictMaybe (Committee era) → CommitteeState era → VotingProcedures era → Set (Credential 'HotCommitteeRole)

data ConwayUTXO era

allegraToConwayUtxoPredFailure ∷ ∀ era. EraRuleFailure "PPUP" era ~ VoidEraRule "PPUP" era ⇒ AllegraUtxoPredFailure era → ConwayUtxoPredFailure era

babbageToConwayUtxoPredFailure ∷ ∀ era. BabbageUtxoPredFailure era → ConwayUtxoPredFailure era

alonzoToConwayUtxoPredFailure ∷ ∀ era. AlonzoUtxoPredFailure era → ConwayUtxoPredFailure era

data ConwayUtxoPredFailure era

Predicate failure for the Conway Era

Constructors

UtxosFailure (PredicateFailure (EraRule "UTXOS" era))
Subtransition Failures

BadInputsUTxO (Set TxIn)
The bad transaction inputs

OutsideValidityIntervalUTxO

Fields

ValidityInterval

transaction's validity interval

SlotNo

current slot

MaxTxSizeUTxO (Mismatch 'RelLTEQ Word32)

InputSetEmptyUTxO

FeeTooSmallUTxO (Mismatch 'RelGTEQ Coin)

ValueNotConservedUTxO (Mismatch 'RelEQ (Value era))

WrongNetwork
the set of addresses with incorrect network IDs

Fields

Network

the expected network id

(Set Addr)

the set of addresses with incorrect network IDs

WrongNetworkWithdrawal

Fields

Network

the expected network id

(Set RewardAccount)

the set of reward addresses with incorrect network IDs

OutputTooSmallUTxO [TxOut era]
list of supplied transaction outputs that are too small

OutputBootAddrAttrsTooBig [TxOut era]
list of supplied bad transaction outputs

OutputTooBigUTxO [(Int, Int, TxOut era)]
list of supplied bad transaction output triples (actualSize,PParameterMaxValue,TxOut)

InsufficientCollateral

Fields

DeltaCoin

balance computed

Coin

the required collateral for the given fee

ScriptsNotPaidUTxO (UTxO era)
The UTxO entries which have the wrong kind of script

ExUnitsTooBigUTxO (Mismatch 'RelLTEQ ExUnits)

CollateralContainsNonADA (Value era)
The inputs marked for use as fees contain non-ADA tokens

WrongNetworkInTxBody (Mismatch 'RelEQ Network)
Wrong Network ID in body

OutsideForecast SlotNo
slot number outside consensus forecast range

TooManyCollateralInputs (Mismatch 'RelLTEQ Natural)
There are too many collateral inputs

NoCollateralInputs

IncorrectTotalCollateralField
The collateral is not equivalent to the total collateral asserted by the transaction

Fields

DeltaCoin

collateral provided

Coin

collateral amount declared in transaction body

BabbageOutputTooSmallUTxO [(TxOut era, Coin)]
list of supplied transaction outputs that are too small, together with the minimum value for the given output.

BabbageNonDisjointRefInputs (NonEmpty TxIn)
TxIns that appear in both inputs and reference inputs

data ConwayUTXOS era

data ConwayUtxosPredFailure era

Constructors

ValidationTagMismatch IsValid TagMismatchDescription
The isValid tag on the transaction is incorrect. The tag given here is that provided on the transaction (whereas evaluation of the scripts gives the opposite.). The Text tries to explain why it failed.

CollectErrors [CollectError era]
We could not find all the necessary inputs for a Plutus Script. Previous PredicateFailure tests should make this impossible, but the consequences of not detecting this means scripts get dropped, so things might validate that shouldn't. So we double check in the function collectTwoPhaseScriptInputs, it should find data for every Script.

data ConwayUtxosEvent era

Constructors

TotalDeposits (SafeHash EraIndependentTxBody) Coin

SuccessfulPlutusScriptsEvent (NonEmpty PlutusWithContext)

FailedPlutusScriptsEvent (NonEmpty PlutusWithContext)

TxUTxODiff
The UTxOs consumed and created by a signal tx

Fields

(UTxO era)

UTxO consumed

(UTxO era)

UTxO created

alonzoToConwayUtxosPredFailure ∷ ∀ era. EraRuleFailure "PPUP" era ~ VoidEraRule "PPUP" era ⇒ AlonzoUtxosPredFailure era → ConwayUtxosPredFailure era

alonzoToConwayUtxosEvent ∷ ∀ era. EraRuleEvent "PPUP" era ~ VoidEraRule "PPUP" era ⇒ AlonzoUtxosEvent era → ConwayUtxosEvent era

alonzoToConwayUtxowPredFailure ∷ ∀ era. AlonzoUtxowPredFailure era → ConwayUtxowPredFailure era

babbageToConwayUtxowPredFailure ∷ ∀ era. BabbageUtxowPredFailure era → ConwayUtxowPredFailure era

data ConwayUTXOW era

data ConwayUtxowPredFailure era

Predicate failure type for the Conway Era

Constructors

UtxoFailure (PredicateFailure (EraRule "UTXO" era))

InvalidWitnessesUTXOW [VKey 'Witness]

MissingVKeyWitnessesUTXOW
witnesses which failed in verifiedWits function

Fields

(Set (KeyHash 'Witness))

witnesses which were needed and not supplied

MissingScriptWitnessesUTXOW (Set ScriptHash)
missing scripts

ScriptWitnessNotValidatingUTXOW (Set ScriptHash)
failed scripts

MissingTxBodyMetadataHash TxAuxDataHash
hash of the full metadata

MissingTxMetadata TxAuxDataHash
hash of the metadata included in the transaction body

ConflictingMetadataHash (Mismatch 'RelEQ TxAuxDataHash)

InvalidMetadata
Contains out of range values (string`s too long)

ExtraneousScriptWitnessesUTXOW (Set ScriptHash)
extraneous scripts

MissingRedeemers [(PlutusPurpose AsItem era, ScriptHash)]

MissingRequiredDatums

Fields

(Set DataHash)

Set of missing data hashes

(Set DataHash)

Set of received data hashes

NotAllowedSupplementalDatums

Fields

(Set DataHash)

Set of unallowed data hashes.

(Set DataHash)

Set of acceptable supplemental data hashes

PPViewHashesDontMatch (Mismatch 'RelEQ (StrictMaybe ScriptIntegrityHash))

UnspendableUTxONoDatumHash (Set TxIn)
Set of transaction inputs that are TwoPhase scripts, and should have a DataHash but don't

ExtraRedeemers [PlutusPurpose AsIx era]
List of redeemers not needed

MalformedScriptWitnesses (Set ScriptHash)
Embed UTXO rule failures

MalformedReferenceScripts (Set ScriptHash)
the set of malformed script witnesses

ScriptIntegrityHashMismatch (Mismatch 'RelEQ (StrictMaybe ScriptIntegrityHash)) (StrictMaybe ByteString)
The computed script integrity hash does not match the provided script integrity hash

shelleyToConwayUtxowPredFailure ∷ ShelleyUtxowPredFailure era → ConwayUtxowPredFailure era

Orphan instances

InjectRuleEvent "TICK" ConwayEpochEvent ConwayEra

Instance details

Methods

injectEvent ∷ ConwayEpochEvent ConwayEra → EraRuleEvent "TICK" ConwayEra

InjectRuleEvent "TICK" ConwayHardForkEvent ConwayEra

Instance details

Methods

injectEvent ∷ ConwayHardForkEvent ConwayEra → EraRuleEvent "TICK" ConwayEra