Cardano.Ledger.Conway.Rules.Bbody

Documentation

data ConwayBBODY era

data ConwayBbodyPredFailure era

Constructors

WrongBlockBodySizeBBODY

Fields

!Int

Actual Body Size

!Int

Claimed Body Size in Header

InvalidBodyHashBBODY

Fields

!(Hash (EraCrypto era) EraIndependentBlockBody)

Actual Hash

!(Hash (EraCrypto era) EraIndependentBlockBody)

Claimed Hash

LedgersFailure !(PredicateFailure (EraRule "LEDGERS" era))
LEDGERS rule subtransition Failures

TooManyExUnits

Fields

!ExUnits

Computed Sum of ExUnits for all plutus scripts

!ExUnits

Maximum allowed by protocal parameters

BodyRefScriptsSizeTooBig

Fields

Int

Computed sum of reference script size

Int

Maximum allowed total reference script size

maxRefScriptSizePerBlock ∷ Int

In the next era this will become a proper protocol parameter. For now this is a hard coded limit on the total number of bytes of all reference scripts combined from all transactions within a block.

Orphan instances

InjectRuleFailure "BBODY" AllegraUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AllegraUtxoPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" AlonzoBbodyPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AlonzoBbodyPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" AlonzoUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxoPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" AlonzoUtxosPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxosPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" AlonzoUtxowPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxowPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" BabbageUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ BabbageUtxoPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" BabbageUtxowPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ BabbageUtxowPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ConwayCertPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayCertPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ConwayCertsPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayCertsPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ConwayDelegPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayDelegPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ConwayGovPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayGovPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ConwayGovCertPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayGovCertPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ConwayLedgerPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayLedgerPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ConwayUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayUtxoPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ConwayUtxosPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayUtxosPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ConwayUtxowPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayUtxowPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ShelleyBbodyPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ShelleyBbodyPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ShelleyLedgersPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ShelleyLedgersPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ShelleyPoolPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ShelleyPoolPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ShelleyUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxoPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

InjectRuleFailure "BBODY" ShelleyUtxowPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxowPredFailure (ConwayEra c) → EraRuleFailure "BBODY" (ConwayEra c)

(Era era, BaseM ledgers ~ ShelleyBase, ledgers ~ EraRule "LEDGERS" era, STS ledgers, DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), Era era) ⇒ Embed ledgers (ConwayBBODY era)

Instance details

Methods

wrapFailed ∷ PredicateFailure ledgers → PredicateFailure (ConwayBBODY era)

wrapEvent ∷ Event ledgers → Event (ConwayBBODY era)

(DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), Embed (EraRule "LEDGERS" era) (EraRule "BBODY" era), Environment (EraRule "LEDGERS" era) ~ ShelleyLedgersEnv era, State (EraRule "LEDGERS" era) ~ LedgerState era, Signal (EraRule "LEDGERS" era) ~ Seq (AlonzoTx era), AlonzoEraTxWits era, Tx era ~ AlonzoTx era, TxSeq era ~ AlonzoTxSeq era, Tx era ~ AlonzoTx era, EraSegWits era, AlonzoEraPParams era, InjectRuleFailure "BBODY" AlonzoBbodyPredFailure era, InjectRuleFailure "BBODY" ConwayBbodyPredFailure era, EraRule "BBODY" era ~ ConwayBBODY era, EraTx era, BabbageEraTxBody era) ⇒ STS (ConwayBBODY era)

Instance details

Associated Types

type State (ConwayBBODY era)

type Signal (ConwayBBODY era)

type Environment (ConwayBBODY era)

type BaseM (ConwayBBODY era) ∷ Type → Type

type Event (ConwayBBODY era)

type PredicateFailure (ConwayBBODY era)

Methods

initialRules ∷ [InitialRule (ConwayBBODY era)]

transitionRules ∷ [TransitionRule (ConwayBBODY era)]

assertions ∷ [Assertion (ConwayBBODY era)]

renderAssertionViolation ∷ AssertionViolation (ConwayBBODY era) → String