Cardano.Ledger.Conway.Rules.Epoch

Documentation

data ConwayEPOCH era

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

data ConwayEpochEvent era

Constructors

PoolReapEvent (Event (EraRule "POOLREAP" era))

SnapEvent (Event (EraRule "SNAP" era))

EpochBoundaryRatifyState (RatifyState era)

GovInfoEvent (Set (GovActionState era)) (Set (GovActionState era)) (Set (GovActionId (EraCrypto era)))

Orphan instances

(EraTxOut era, RunConwayRatify era, ConwayEraGov era, Embed (EraRule "SNAP" era) (ConwayEPOCH era), Environment (EraRule "SNAP" era) ~ SnapEnv era, State (EraRule "SNAP" era) ~ SnapShots (EraCrypto era), Signal (EraRule "SNAP" era) ~ (), Embed (EraRule "POOLREAP" era) (ConwayEPOCH era), Environment (EraRule "POOLREAP" era) ~ ShelleyPoolreapEnv era, State (EraRule "POOLREAP" era) ~ ShelleyPoolreapState era, Signal (EraRule "POOLREAP" era) ~ EpochNo, Eq (UpecPredFailure era), Show (UpecPredFailure era), Embed (EraRule "RATIFY" era) (ConwayEPOCH era), Environment (EraRule "RATIFY" era) ~ RatifyEnv era, GovState era ~ ConwayGovState era, State (EraRule "RATIFY" era) ~ RatifyState era, Signal (EraRule "RATIFY" era) ~ RatifySignal era) ⇒ STS (ConwayEPOCH era)

Instance details

Associated Types

type State (ConwayEPOCH era)

type Signal (ConwayEPOCH era)

type Environment (ConwayEPOCH era)

type BaseM (ConwayEPOCH era) ∷ Type → Type

type Event (ConwayEPOCH era)

type PredicateFailure (ConwayEPOCH era)

Methods

initialRules ∷ [InitialRule (ConwayEPOCH era)]

transitionRules ∷ [TransitionRule (ConwayEPOCH era)]

assertions ∷ [Assertion (ConwayEPOCH era)]

renderAssertionViolation ∷ AssertionViolation (ConwayEPOCH era) → String

(EraGov era, PredicateFailure (ConwayRATIFY era) ~ Void, STS (ConwayRATIFY era), BaseM (ConwayRATIFY era) ~ ShelleyBase, Event (ConwayRATIFY era) ~ Void) ⇒ Embed (ConwayRATIFY era) (ConwayEPOCH era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ConwayRATIFY era) → PredicateFailure (ConwayEPOCH era)

wrapEvent ∷ Event (ConwayRATIFY era) → Event (ConwayEPOCH era)

(Era era, STS (ShelleyPOOLREAP era), PredicateFailure (EraRule "POOLREAP" era) ~ ShelleyPoolreapPredFailure era, Event (EraRule "POOLREAP" era) ~ ShelleyPoolreapEvent era) ⇒ Embed (ShelleyPOOLREAP era) (ConwayEPOCH era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleyPOOLREAP era) → PredicateFailure (ConwayEPOCH era)

wrapEvent ∷ Event (ShelleyPOOLREAP era) → Event (ConwayEPOCH era)

(EraTxOut era, PredicateFailure (EraRule "SNAP" era) ~ ShelleySnapPredFailure era, Event (EraRule "SNAP" era) ~ SnapEvent era) ⇒ Embed (ShelleySNAP era) (ConwayEPOCH era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ShelleySNAP era) → PredicateFailure (ConwayEPOCH era)

wrapEvent ∷ Event (ShelleySNAP era) → Event (ConwayEPOCH era)