Cardano.Ledger.Conway.Rules.Gov

Documentation

data ConwayGOV era

data GovEnv era

Constructors

GovEnv

Fields

geTxId ∷ !(TxId (EraCrypto era))

geEpoch ∷ !EpochNo

gePParams ∷ !(PParams era)

gePPolicy ∷ !(StrictMaybe (ScriptHash (EraCrypto era)))

geCertState ∷ !(CertState era)

data GovSignal era

Constructors

GovSignal

Fields

gsVotingProcedures ∷ !(VotingProcedures era)

gsProposalProcedures ∷ !(OSet (ProposalProcedure era))

gsCertificates ∷ !(StrictSeq (TxCert era))

data ConwayGovEvent era

Constructors

GovNewProposals !(TxId (EraCrypto era)) !(Proposals era)

data ConwayGovPredFailure era

Constructors

GovActionsDoNotExist (NonEmpty (GovActionId (EraCrypto era)))

MalformedProposal (GovAction era)

ProposalProcedureNetworkIdMismatch (RewardAccount (EraCrypto era)) Network

TreasuryWithdrawalsNetworkIdMismatch (Set (RewardAccount (EraCrypto era))) Network

ProposalDepositIncorrect

Fields

Coin

Submitted deposit

Coin

Expected deposit taken from PParams

DisallowedVoters !(NonEmpty (Voter (EraCrypto era), GovActionId (EraCrypto era)))
Some governance actions are not allowed to be voted on by certain types of Voters. This failure lists all governance action ids with their respective voters that are not allowed to vote on those governance actions.

ConflictingCommitteeUpdate

Fields

(Set (Credential 'ColdCommitteeRole (EraCrypto era)))

Credentials that are mentioned as members to be both removed and added

ExpirationEpochTooSmall

Fields

(Map (Credential 'ColdCommitteeRole (EraCrypto era)) EpochNo)

Members for which the expiration epoch has already been reached

InvalidPrevGovActionId (ProposalProcedure era)

VotingOnExpiredGovAction (NonEmpty (Voter (EraCrypto era), GovActionId (EraCrypto era)))

ProposalCantFollow

Fields

(StrictMaybe (GovPurposeId 'HardForkPurpose era))

The PrevGovActionId of the HardForkInitiation that fails

ProtVer

Its protocol version

ProtVer

The ProtVer of the Previous GovAction pointed to by the one proposed

InvalidPolicyHash

Fields

(StrictMaybe (ScriptHash (EraCrypto era)))

The policy script hash in the proposal

(StrictMaybe (ScriptHash (EraCrypto era)))

The policy script hash of the current constitution

DisallowedProposalDuringBootstrap (ProposalProcedure era)

DisallowedVotesDuringBootstrap (NonEmpty (Voter (EraCrypto era), GovActionId (EraCrypto era)))

VotersDoNotExist (NonEmpty (Voter (EraCrypto era)))
Predicate failure for votes by entities that are not present in the ledger state

Orphan instances

(ConwayEraTxCert era, ConwayEraPParams era, EraRule "GOV" era ~ ConwayGOV era, InjectRuleFailure "GOV" ConwayGovPredFailure era) ⇒ STS (ConwayGOV era)

Instance details

Associated Types

type State (ConwayGOV era)

type Signal (ConwayGOV era)

type Environment (ConwayGOV era)

type BaseM (ConwayGOV era) ∷ Type → Type

type Event (ConwayGOV era)

type PredicateFailure (ConwayGOV era)

Methods

initialRules ∷ [InitialRule (ConwayGOV era)]

transitionRules ∷ [TransitionRule (ConwayGOV era)]

assertions ∷ [Assertion (ConwayGOV era)]

renderAssertionViolation ∷ AssertionViolation (ConwayGOV era) → String