Cardano.Ledger.Conway.Rules.GovCert

Documentation

data ConwayGOVCERT era

data ConwayGovCertPredFailure era

Constructors

ConwayDRepAlreadyRegistered !(Credential 'DRepRole (EraCrypto era))

ConwayDRepNotRegistered !(Credential 'DRepRole (EraCrypto era))

ConwayDRepIncorrectDeposit !Coin !Coin

ConwayCommitteeHasPreviouslyResigned !(Credential 'ColdCommitteeRole (EraCrypto era))

ConwayDRepIncorrectRefund !Coin !Coin

ConwayCommitteeIsUnknown !(Credential 'ColdCommitteeRole (EraCrypto era))
Predicate failure whenever an update to an unknown committee member is attempted. Current Constitutional Committee and all available proposals will be searched before reporting this predicate failure.

data ConwayGovCertEnv era

Constructors

ConwayGovCertEnv

Fields

cgcePParams ∷ !(PParams era)

cgceCurrentEpoch ∷ !EpochNo

cgceCurrentCommittee ∷ StrictMaybe (Committee era)

cgceCommitteeProposals ∷ Map (GovPurposeId 'CommitteePurpose era) (GovActionState era)

All of the UpdateCommittee proposals

updateDRepExpiry

Arguments

∷ EpochInterval
DRepActivity PParam

→ EpochNo
Current epoch

→ EpochNo
The count of the dormant epochs

→ DRepState c

→ DRepState c

Orphan instances

(ConwayEraPParams era, State (EraRule "GOVCERT" era) ~ VState era, Signal (EraRule "GOVCERT" era) ~ ConwayGovCert (EraCrypto era), Environment (EraRule "GOVCERT" era) ~ ConwayGovCertEnv era, EraRule "GOVCERT" era ~ ConwayGOVCERT era, Eq (PredicateFailure (EraRule "GOVCERT" era)), Show (PredicateFailure (EraRule "GOVCERT" era))) ⇒ STS (ConwayGOVCERT era)

Instance details

Associated Types

type State (ConwayGOVCERT era)

type Signal (ConwayGOVCERT era)

type Environment (ConwayGOVCERT era)

type BaseM (ConwayGOVCERT era) ∷ Type → Type

type Event (ConwayGOVCERT era)

type PredicateFailure (ConwayGOVCERT era)

Methods

initialRules ∷ [InitialRule (ConwayGOVCERT era)]

transitionRules ∷ [TransitionRule (ConwayGOVCERT era)]

assertions ∷ [Assertion (ConwayGOVCERT era)]

renderAssertionViolation ∷ AssertionViolation (ConwayGOVCERT era) → String