Cardano.Ledger.Conway.Rules.Ledger

Documentation

data ConwayLEDGER era

data ConwayLedgerPredFailure era

Constructors

ConwayUtxowFailure (PredicateFailure (EraRule "UTXOW" era))

ConwayCertsFailure (PredicateFailure (EraRule "CERTS" era))

ConwayGovFailure (PredicateFailure (EraRule "GOV" era))

ConwayWdrlNotDelegatedToDRep (Set (Credential 'Staking (EraCrypto era)))

ConwayTreasuryValueMismatch

Fields

Coin

Actual

Coin

Submitted in transaction

ConwayTxRefScriptsSizeTooBig

Fields

Int

Computed sum of reference script size

Int

Maximum allowed total reference script size

data ConwayLedgerEvent era

Constructors

UtxowEvent (Event (EraRule "UTXOW" era))

CertsEvent (Event (EraRule "CERTS" era))

GovEvent (Event (EraRule "GOV" era))

maxRefScriptSizePerTx ∷ Int

In the next era this will become a proper protocol parameter. For now this is a hard coded limit on the total number of bytes of reference scripts that a transaction can use.

Orphan instances

InjectRuleFailure "LEDGER" AllegraUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AllegraUtxoPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" AlonzoUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxoPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" AlonzoUtxosPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxosPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" AlonzoUtxowPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxowPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" BabbageUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ BabbageUtxoPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" BabbageUtxowPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ BabbageUtxowPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" ConwayCertPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayCertPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" ConwayCertsPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayCertsPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" ConwayDelegPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayDelegPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" ConwayGovPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayGovPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" ConwayGovCertPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayGovCertPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" ConwayUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayUtxoPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" ConwayUtxosPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayUtxosPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" ConwayUtxowPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayUtxowPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" ShelleyPoolPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ShelleyPoolPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" ShelleyUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxoPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

InjectRuleFailure "LEDGER" ShelleyUtxowPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxowPredFailure (ConwayEra c) → EraRuleFailure "LEDGER" (ConwayEra c)

(AlonzoEraTx era, ConwayEraTxBody era, ConwayEraGov era, GovState era ~ ConwayGovState era, Embed (EraRule "UTXOW" era) (ConwayLEDGER era), Embed (EraRule "GOV" era) (ConwayLEDGER era), Embed (EraRule "CERTS" era) (ConwayLEDGER era), State (EraRule "UTXOW" era) ~ UTxOState era, State (EraRule "CERTS" era) ~ CertState era, State (EraRule "GOV" era) ~ Proposals era, Environment (EraRule "UTXOW" era) ~ UtxoEnv era, Environment (EraRule "CERTS" era) ~ CertsEnv era, Environment (EraRule "GOV" era) ~ GovEnv era, Signal (EraRule "UTXOW" era) ~ Tx era, Signal (EraRule "CERTS" era) ~ Seq (TxCert era), Signal (EraRule "GOV" era) ~ GovSignal era) ⇒ STS (ConwayLEDGER era)

Instance details

Associated Types

type State (ConwayLEDGER era)

type Signal (ConwayLEDGER era)

type Environment (ConwayLEDGER era)

type BaseM (ConwayLEDGER era) ∷ Type → Type

type Event (ConwayLEDGER era)

type PredicateFailure (ConwayLEDGER era)

Methods

initialRules ∷ [InitialRule (ConwayLEDGER era)]

transitionRules ∷ [TransitionRule (ConwayLEDGER era)]

assertions ∷ [Assertion (ConwayLEDGER era)]

renderAssertionViolation ∷ AssertionViolation (ConwayLEDGER era) → String

(EraTx era, ConwayEraTxBody era, ConwayEraPParams era, ConwayEraGov era, Embed (EraRule "CERT" era) (ConwayCERTS era), State (EraRule "CERT" era) ~ CertState era, Environment (EraRule "CERT" era) ~ CertEnv era, Signal (EraRule "CERT" era) ~ TxCert era, PredicateFailure (EraRule "CERTS" era) ~ ConwayCertsPredFailure era, Event (EraRule "CERTS" era) ~ ConwayCertsEvent era, EraRule "CERTS" era ~ ConwayCERTS era) ⇒ Embed (ConwayCERTS era) (ConwayLEDGER era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ConwayCERTS era) → PredicateFailure (ConwayLEDGER era)

wrapEvent ∷ Event (ConwayCERTS era) → Event (ConwayLEDGER era)

(EraPParams era, EraRule "DELEG" era ~ ConwayDELEG era, PredicateFailure (EraRule "CERTS" era) ~ ConwayCertsPredFailure era, PredicateFailure (EraRule "CERT" era) ~ ConwayCertPredFailure era, Event (EraRule "CERTS" era) ~ ConwayCertsEvent era, Event (EraRule "CERT" era) ~ ConwayCertEvent era) ⇒ Embed (ConwayDELEG era) (ConwayLEDGER era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ConwayDELEG era) → PredicateFailure (ConwayLEDGER era)

wrapEvent ∷ Event (ConwayDELEG era) → Event (ConwayLEDGER era)

(ConwayEraTxCert era, ConwayEraPParams era, BaseM (ConwayLEDGER era) ~ ShelleyBase, PredicateFailure (EraRule "GOV" era) ~ ConwayGovPredFailure era, Event (EraRule "GOV" era) ~ ConwayGovEvent era, EraRule "GOV" era ~ ConwayGOV era, InjectRuleFailure "GOV" ConwayGovPredFailure era) ⇒ Embed (ConwayGOV era) (ConwayLEDGER era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ConwayGOV era) → PredicateFailure (ConwayLEDGER era)

wrapEvent ∷ Event (ConwayGOV era) → Event (ConwayLEDGER era)

(Embed (EraRule "UTXOW" era) (ConwayLEDGER era), Embed (EraRule "CERTS" era) (ConwayLEDGER era), Embed (EraRule "GOV" era) (ConwayLEDGER era), ConwayEraGov era, AlonzoEraTx era, ConwayEraTxBody era, ConwayEraPParams era, GovState era ~ ConwayGovState era, Environment (EraRule "UTXOW" era) ~ UtxoEnv era, Environment (EraRule "CERTS" era) ~ CertsEnv era, Environment (EraRule "GOV" era) ~ GovEnv era, Signal (EraRule "UTXOW" era) ~ Tx era, Signal (EraRule "CERTS" era) ~ Seq (TxCert era), Signal (EraRule "GOV" era) ~ GovSignal era, State (EraRule "UTXOW" era) ~ UTxOState era, State (EraRule "CERTS" era) ~ CertState era, State (EraRule "GOV" era) ~ Proposals era, EraRule "GOV" era ~ ConwayGOV era, PredicateFailure (EraRule "LEDGER" era) ~ ConwayLedgerPredFailure era, Event (EraRule "LEDGER" era) ~ ConwayLedgerEvent era, EraGov era) ⇒ Embed (ConwayLEDGER era) (ShelleyLEDGERS era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ConwayLEDGER era) → PredicateFailure (ShelleyLEDGERS era)

wrapEvent ∷ Event (ConwayLEDGER era) → Event (ShelleyLEDGERS era)

(Signable (DSIGN (EraCrypto era)) (Hash (HASH (EraCrypto era)) EraIndependentTxBody), BaseM (ConwayUTXOW era) ~ ShelleyBase, AlonzoEraTx era, EraUTxO era, BabbageEraTxBody era, Embed (EraRule "UTXO" era) (ConwayUTXOW era), State (EraRule "UTXO" era) ~ UTxOState era, Environment (EraRule "UTXO" era) ~ UtxoEnv era, Script era ~ AlonzoScript era, TxOut era ~ BabbageTxOut era, ScriptsNeeded era ~ AlonzoScriptsNeeded era, Signal (EraRule "UTXO" era) ~ Tx era, PredicateFailure (EraRule "UTXOW" era) ~ ConwayUtxowPredFailure era, Event (EraRule "UTXOW" era) ~ AlonzoUtxowEvent era, STS (ConwayUTXOW era), PredicateFailure (ConwayUTXOW era) ~ ConwayUtxowPredFailure era, Event (ConwayUTXOW era) ~ AlonzoUtxowEvent era) ⇒ Embed (ConwayUTXOW era) (ConwayLEDGER era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ConwayUTXOW era) → PredicateFailure (ConwayLEDGER era)

wrapEvent ∷ Event (ConwayUTXOW era) → Event (ConwayLEDGER era)