Cardano.Ledger.Conway.Rules.Ratify

Documentation

data RatifyState era

Constructors

RatifyState

Fields

rsEnactState ∷ !(EnactState era)

rsEnacted ∷ !(Seq (GovActionState era))

rsExpired ∷ !(Set (GovActionId (EraCrypto era)))

rsDelayed ∷ !Bool

committeeAccepted ∷ ConwayEraPParams era ⇒ RatifyEnv era → RatifyState era → GovActionState era → Bool

committeeAcceptedRatio ∷ ∀ era. Map (Credential 'ColdCommitteeRole (EraCrypto era)) EpochNo → Map (Credential 'HotCommitteeRole (EraCrypto era)) Vote → CommitteeState era → EpochNo → Rational

spoAccepted ∷ ConwayEraPParams era ⇒ RatifyEnv era → RatifyState era → GovActionState era → Bool

spoAcceptedRatio ∷ ∀ era. RatifyEnv era → GovActionState era → Rational

Final ratio for totalAcceptedStakePoolsRatio we want is: t = y / (s - a) Where: * y - total delegated stake that voted Yes * a - total delegated stake that voted Abstain * s - total delegated stake

For HardForkInitiation all SPOs that didn't vote are considered as No votes. Whereas, for all other GovActions, SPOs that didn't vote are considered as Abstain votes.

No votes are not counted.

dRepAccepted ∷ ConwayEraPParams era ⇒ RatifyEnv era → RatifyState era → GovActionState era → Bool

dRepAcceptedRatio ∷ ∀ era. RatifyEnv era → Map (Credential 'DRepRole (EraCrypto era)) Vote → GovAction era → Rational

prevActionAsExpected ∷ GovActionState era → GovRelation StrictMaybe era → Bool

Check that the previous governance action id specified in the proposal does match the last one of the same purpose that was enacted.

validCommitteeTerm ∷ ConwayEraPParams era ⇒ GovAction era → PParams era → EpochNo → Bool

withdrawalCanWithdraw ∷ GovAction era → Coin → Bool

Orphan instances

(ConwayEraPParams era, Embed (EraRule "ENACT" era) (ConwayRATIFY era), State (EraRule "ENACT" era) ~ EnactState era, Environment (EraRule "ENACT" era) ~ (), Signal (EraRule "ENACT" era) ~ EnactSignal era) ⇒ STS (ConwayRATIFY era)

Instance details

Associated Types

type State (ConwayRATIFY era)

type Signal (ConwayRATIFY era)

type Environment (ConwayRATIFY era)

type BaseM (ConwayRATIFY era) ∷ Type → Type

type Event (ConwayRATIFY era)

type PredicateFailure (ConwayRATIFY era)

Methods

initialRules ∷ [InitialRule (ConwayRATIFY era)]

transitionRules ∷ [TransitionRule (ConwayRATIFY era)]

assertions ∷ [Assertion (ConwayRATIFY era)]

renderAssertionViolation ∷ AssertionViolation (ConwayRATIFY era) → String

EraGov era ⇒ Embed (ConwayENACT era) (ConwayRATIFY era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ConwayENACT era) → PredicateFailure (ConwayRATIFY era)

wrapEvent ∷ Event (ConwayENACT era) → Event (ConwayRATIFY era)