Cardano.Ledger.Conway.Rules.Utxos

Documentation

data ConwayUTXOS era

data ConwayUtxosPredFailure era

Constructors

ValidationTagMismatch IsValid TagMismatchDescription
The isValid tag on the transaction is incorrect. The tag given here is that provided on the transaction (whereas evaluation of the scripts gives the opposite.). The Text tries to explain why it failed.

CollectErrors [CollectError era]
We could not find all the necessary inputs for a Plutus Script. Previous PredicateFailure tests should make this impossible, but the consequences of not detecting this means scripts get dropped, so things might validate that shouldn't. So we double check in the function collectTwoPhaseScriptInputs, it should find data for every Script.

data ConwayUtxosEvent era

Constructors

TotalDeposits (SafeHash (EraCrypto era) EraIndependentTxBody) Coin

SuccessfulPlutusScriptsEvent (NonEmpty (PlutusWithContext (EraCrypto era)))

FailedPlutusScriptsEvent (NonEmpty (PlutusWithContext (EraCrypto era)))

TxUTxODiff
The UTxOs consumed and created by a signal tx

Fields

(UTxO era)

UTxO consumed

(UTxO era)

UTxO created

Orphan instances

InjectRuleEvent "UTXOS" AlonzoUtxosEvent (ConwayEra c)

Instance details

Methods

injectEvent ∷ AlonzoUtxosEvent (ConwayEra c) → EraRuleEvent "UTXOS" (ConwayEra c)

InjectRuleFailure "UTXOS" AlonzoUtxosPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxosPredFailure (ConwayEra c) → EraRuleFailure "UTXOS" (ConwayEra c)

(AlonzoEraTx era, AlonzoEraUTxO era, ConwayEraTxBody era, ConwayEraPParams era, EraGov era, EraPlutusContext era, GovState era ~ ConwayGovState era, ScriptsNeeded era ~ AlonzoScriptsNeeded era, Signal (ConwayUTXOS era) ~ Tx era, EraRule "UTXOS" era ~ ConwayUTXOS era, InjectRuleFailure "UTXOS" AlonzoUtxosPredFailure era, InjectRuleEvent "UTXOS" AlonzoUtxosEvent era, InjectRuleEvent "UTXOS" ConwayUtxosEvent era) ⇒ STS (ConwayUTXOS era)

Instance details

Associated Types

type State (ConwayUTXOS era)

type Signal (ConwayUTXOS era)

type Environment (ConwayUTXOS era)

type BaseM (ConwayUTXOS era) ∷ Type → Type

type Event (ConwayUTXOS era)

type PredicateFailure (ConwayUTXOS era)

Methods

initialRules ∷ [InitialRule (ConwayUTXOS era)]

transitionRules ∷ [TransitionRule (ConwayUTXOS era)]

assertions ∷ [Assertion (ConwayUTXOS era)]

renderAssertionViolation ∷ AssertionViolation (ConwayUTXOS era) → String

(AlonzoEraTx era, AlonzoEraUTxO era, ConwayEraTxBody era, ConwayEraPParams era, EraGov era, EraPlutusContext era, GovState era ~ ConwayGovState era, PredicateFailure (EraRule "UTXOS" era) ~ ConwayUtxosPredFailure era, ScriptsNeeded era ~ AlonzoScriptsNeeded era, Signal (ConwayUTXOS era) ~ Tx era, EraRule "UTXOS" era ~ ConwayUTXOS era, InjectRuleFailure "UTXOS" AlonzoUtxosPredFailure era, InjectRuleEvent "UTXOS" AlonzoUtxosEvent era, InjectRuleEvent "UTXOS" ConwayUtxosEvent era) ⇒ Embed (ConwayUTXOS era) (BabbageUTXO era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ConwayUTXOS era) → PredicateFailure (BabbageUTXO era)

wrapEvent ∷ Event (ConwayUTXOS era) → Event (BabbageUTXO era)