Cardano.Ledger.Conway.Rules.Utxow

Documentation

alonzoToConwayUtxowPredFailure ∷ ∀ era. AlonzoUtxowPredFailure era → ConwayUtxowPredFailure era

babbageToConwayUtxowPredFailure ∷ ∀ era. BabbageUtxowPredFailure era → ConwayUtxowPredFailure era

data ConwayUTXOW era

data ConwayUtxowPredFailure era

Predicate failure type for the Conway Era

Constructors

UtxoFailure (PredicateFailure (EraRule "UTXO" era))

InvalidWitnessesUTXOW ![VKey 'Witness (EraCrypto era)]

MissingVKeyWitnessesUTXOW
witnesses which failed in verifiedWits function

Fields

!(Set (KeyHash 'Witness (EraCrypto era)))

witnesses which were needed and not supplied

MissingScriptWitnessesUTXOW !(Set (ScriptHash (EraCrypto era)))
missing scripts

ScriptWitnessNotValidatingUTXOW !(Set (ScriptHash (EraCrypto era)))
failed scripts

MissingTxBodyMetadataHash !(AuxiliaryDataHash (EraCrypto era))
hash of the full metadata

MissingTxMetadata !(AuxiliaryDataHash (EraCrypto era))
hash of the metadata included in the transaction body

ConflictingMetadataHash

Fields

!(AuxiliaryDataHash (EraCrypto era))

hash of the metadata included in the transaction body

!(AuxiliaryDataHash (EraCrypto era))

expected hash of the full metadata

InvalidMetadata
Contains out of range values (string`s too long)

ExtraneousScriptWitnessesUTXOW !(Set (ScriptHash (EraCrypto era)))
extraneous scripts

MissingRedeemers ![(PlutusPurpose AsItem era, ScriptHash (EraCrypto era))]

MissingRequiredDatums

Fields

!(Set (DataHash (EraCrypto era)))

Set of missing data hashes

!(Set (DataHash (EraCrypto era)))

Set of received data hashes

NotAllowedSupplementalDatums

Fields

!(Set (DataHash (EraCrypto era)))

Set of unallowed data hashes

!(Set (DataHash (EraCrypto era)))

Set of acceptable supplemental data hashes

PPViewHashesDontMatch

Fields

!(StrictMaybe (ScriptIntegrityHash (EraCrypto era)))

The PPHash in the TxBody

!(StrictMaybe (ScriptIntegrityHash (EraCrypto era)))

Computed from the current Protocol Parameters

UnspendableUTxONoDatumHash (Set (TxIn (EraCrypto era)))
Set of transaction inputs that are TwoPhase scripts, and should have a DataHash but don't

ExtraRedeemers ![PlutusPurpose AsIx era]
List of redeemers not needed

MalformedScriptWitnesses !(Set (ScriptHash (EraCrypto era)))
Embed UTXO rule failures

MalformedReferenceScripts !(Set (ScriptHash (EraCrypto era)))
the set of malformed script witnesses

conwayWitsVKeyNeeded ∷ (EraTx era, ConwayEraTxBody era) ⇒ UTxO era → TxBody era → Set (KeyHash 'Witness (EraCrypto era))

Deprecated: In favor of getConwayWitsVKeyNeeded or getWitsVKeyNeeded

shelleyToConwayUtxowPredFailure ∷ ShelleyUtxowPredFailure era → ConwayUtxowPredFailure era

Orphan instances

InjectRuleFailure "UTXOW" AllegraUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AllegraUtxoPredFailure (ConwayEra c) → EraRuleFailure "UTXOW" (ConwayEra c)

InjectRuleFailure "UTXOW" AlonzoUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxoPredFailure (ConwayEra c) → EraRuleFailure "UTXOW" (ConwayEra c)

InjectRuleFailure "UTXOW" AlonzoUtxosPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxosPredFailure (ConwayEra c) → EraRuleFailure "UTXOW" (ConwayEra c)

InjectRuleFailure "UTXOW" AlonzoUtxowPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ AlonzoUtxowPredFailure (ConwayEra c) → EraRuleFailure "UTXOW" (ConwayEra c)

InjectRuleFailure "UTXOW" BabbageUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ BabbageUtxoPredFailure (ConwayEra c) → EraRuleFailure "UTXOW" (ConwayEra c)

InjectRuleFailure "UTXOW" BabbageUtxowPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ BabbageUtxowPredFailure (ConwayEra c) → EraRuleFailure "UTXOW" (ConwayEra c)

InjectRuleFailure "UTXOW" ConwayUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayUtxoPredFailure (ConwayEra c) → EraRuleFailure "UTXOW" (ConwayEra c)

InjectRuleFailure "UTXOW" ConwayUtxosPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ConwayUtxosPredFailure (ConwayEra c) → EraRuleFailure "UTXOW" (ConwayEra c)

InjectRuleFailure "UTXOW" ShelleyUtxoPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxoPredFailure (ConwayEra c) → EraRuleFailure "UTXOW" (ConwayEra c)

InjectRuleFailure "UTXOW" ShelleyUtxowPredFailure (ConwayEra c)

Instance details

Methods

injectFailure ∷ ShelleyUtxowPredFailure (ConwayEra c) → EraRuleFailure "UTXOW" (ConwayEra c)

(AlonzoEraTx era, AlonzoEraUTxO era, ScriptsNeeded era ~ AlonzoScriptsNeeded era, ConwayEraTxBody era, Signable (DSIGN (EraCrypto era)) (Hash (HASH (EraCrypto era)) EraIndependentTxBody), EraRule "UTXOW" era ~ ConwayUTXOW era, InjectRuleFailure "UTXOW" ShelleyUtxowPredFailure era, InjectRuleFailure "UTXOW" AlonzoUtxowPredFailure era, InjectRuleFailure "UTXOW" BabbageUtxowPredFailure era, Embed (EraRule "UTXO" era) (ConwayUTXOW era), Environment (EraRule "UTXO" era) ~ UtxoEnv era, State (EraRule "UTXO" era) ~ UTxOState era, Signal (EraRule "UTXO" era) ~ Tx era, Eq (PredicateFailure (EraRule "UTXOS" era)), Show (PredicateFailure (EraRule "UTXOS" era))) ⇒ STS (ConwayUTXOW era)

Instance details

Associated Types

type State (ConwayUTXOW era)

type Signal (ConwayUTXOW era)

type Environment (ConwayUTXOW era)

type BaseM (ConwayUTXOW era) ∷ Type → Type

type Event (ConwayUTXOW era)

type PredicateFailure (ConwayUTXOW era)

Methods

initialRules ∷ [InitialRule (ConwayUTXOW era)]

transitionRules ∷ [TransitionRule (ConwayUTXOW era)]

assertions ∷ [Assertion (ConwayUTXOW era)]

renderAssertionViolation ∷ AssertionViolation (ConwayUTXOW era) → String

(Era era, STS (ConwayUTXO era), PredicateFailure (EraRule "UTXO" era) ~ ConwayUtxoPredFailure era, Event (EraRule "UTXO" era) ~ AlonzoUtxoEvent era, BaseM (ConwayUTXOW era) ~ ShelleyBase, PredicateFailure (ConwayUTXOW era) ~ ConwayUtxowPredFailure era, Event (ConwayUTXOW era) ~ AlonzoUtxowEvent era) ⇒ Embed (ConwayUTXO era) (ConwayUTXOW era)

Instance details

Methods

wrapFailed ∷ PredicateFailure (ConwayUTXO era) → PredicateFailure (ConwayUTXOW era)

wrapEvent ∷ Event (ConwayUTXO era) → Event (ConwayUTXOW era)