Cardano.Ledger.Conway.Translation

Documentation

newtype Tx era

Constructors

Tx

Fields

unTx ∷ Tx era

addrPtrNormalize ∷ Addr c → Addr c

This function is implemented solely for the purpose of translating garbage pointers into knowingly invalid ones. Any pointer that contains a SlotNo, TxIx or CertIx that is too large to fit into Word32, Word16 and Word16 respectively, will have all of its values set to 0 using normalizePtr.

There are two reasons why we can safely do that at the Babbage/Conway era boundary:

Invalid pointers are no longer allowed in transactions starting with Babbage era

There are only a handful of Ptrs on mainnet that are invalid.

Once the transition is complete and we are officially in Conway era, this translation logic can be removed in favor of a fixed deserializer that does the same thing for all eras prior to Babbage.

translateDatum ∷ EraCrypto era1 ~ EraCrypto era2 ⇒ Datum era1 → Datum era2

translateTxOut ∷ Crypto c ⇒ TxOut (BabbageEra c) → TxOut (ConwayEra c)

Deprecated: In favor of upgradeTxOut

Filter out TxOuts with zero Coins and normalize Pointers, while converting TxOuts to Conway era.

Orphan instances

Crypto c ⇒ TranslateEra (ConwayEra c) CertState

Instance details

Associated Types

type TranslationError (ConwayEra c) CertState

Methods

translateEra ∷ TranslationContext (ConwayEra c) → CertState (PreviousEra (ConwayEra c)) → Except (TranslationError (ConwayEra c) CertState) (CertState (ConwayEra c))

Crypto c ⇒ TranslateEra (ConwayEra c) CommitteeState

Instance details

Associated Types

type TranslationError (ConwayEra c) CommitteeState

Methods

translateEra ∷ TranslationContext (ConwayEra c) → CommitteeState (PreviousEra (ConwayEra c)) → Except (TranslationError (ConwayEra c) CommitteeState) (CommitteeState (ConwayEra c))

Crypto c ⇒ TranslateEra (ConwayEra c) DState

Instance details

Associated Types

type TranslationError (ConwayEra c) DState

Methods

translateEra ∷ TranslationContext (ConwayEra c) → DState (PreviousEra (ConwayEra c)) → Except (TranslationError (ConwayEra c) DState) (DState (ConwayEra c))

Crypto c ⇒ TranslateEra (ConwayEra c) PState

Instance details

Associated Types

type TranslationError (ConwayEra c) PState

Methods

translateEra ∷ TranslationContext (ConwayEra c) → PState (PreviousEra (ConwayEra c)) → Except (TranslationError (ConwayEra c) PState) (PState (ConwayEra c))

Crypto c ⇒ TranslateEra (ConwayEra c) VState

Instance details

Associated Types

type TranslationError (ConwayEra c) VState

Methods

translateEra ∷ TranslationContext (ConwayEra c) → VState (PreviousEra (ConwayEra c)) → Except (TranslationError (ConwayEra c) VState) (VState (ConwayEra c))

Crypto c ⇒ TranslateEra (ConwayEra c) PParams

Instance details

Associated Types

type TranslationError (ConwayEra c) PParams

Methods

translateEra ∷ TranslationContext (ConwayEra c) → PParams (PreviousEra (ConwayEra c)) → Except (TranslationError (ConwayEra c) PParams) (PParams (ConwayEra c))

Crypto c ⇒ TranslateEra (ConwayEra c) UTxO

Instance details

Associated Types

type TranslationError (ConwayEra c) UTxO

Methods

translateEra ∷ TranslationContext (ConwayEra c) → UTxO (PreviousEra (ConwayEra c)) → Except (TranslationError (ConwayEra c) UTxO) (UTxO (ConwayEra c))

Crypto c ⇒ TranslateEra (ConwayEra c) FuturePParams

Instance details

Associated Types

type TranslationError (ConwayEra c) FuturePParams

Methods

translateEra ∷ TranslationContext (ConwayEra c) → FuturePParams (PreviousEra (ConwayEra c)) → Except (TranslationError (ConwayEra c) FuturePParams) (FuturePParams (ConwayEra c))

Crypto c ⇒ TranslateEra (ConwayEra c) EpochState

Instance details

Associated Types

type TranslationError (ConwayEra c) EpochState

Methods

translateEra ∷ TranslationContext (ConwayEra c) → EpochState (PreviousEra (ConwayEra c)) → Except (TranslationError (ConwayEra c) EpochState) (EpochState (ConwayEra c))

Crypto c ⇒ TranslateEra (ConwayEra c) LedgerState

Instance details

Associated Types

type TranslationError (ConwayEra c) LedgerState

Methods

translateEra ∷ TranslationContext (ConwayEra c) → LedgerState (PreviousEra (ConwayEra c)) → Except (TranslationError (ConwayEra c) LedgerState) (LedgerState (ConwayEra c))

Crypto c ⇒ TranslateEra (ConwayEra c) NewEpochState

Instance details

Associated Types

type TranslationError (ConwayEra c) NewEpochState

Methods

translateEra ∷ TranslationContext (ConwayEra c) → NewEpochState (PreviousEra (ConwayEra c)) → Except (TranslationError (ConwayEra c) NewEpochState) (NewEpochState (ConwayEra c))

Crypto c ⇒ TranslateEra (ConwayEra c) UTxOState

Instance details

Associated Types

type TranslationError (ConwayEra c) UTxOState

Methods

translateEra ∷ TranslationContext (ConwayEra c) → UTxOState (PreviousEra (ConwayEra c)) → Except (TranslationError (ConwayEra c) UTxOState) (UTxOState (ConwayEra c))