Cardano.Ledger.Conway.Tx

Documentation

data family Tx era

data family Tx era

class (EraTx era, AlonzoEraTxBody era, AlonzoEraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTx era where

Methods

isValidTxL ∷ Lens' (Tx era) IsValid

data AlonzoTx era

Constructors

AlonzoTx

Fields

atBody ∷ !(TxBody era)

atWits ∷ !(TxWits era)

atIsValid ∷ !IsValid

atAuxData ∷ !(StrictMaybe (TxAuxData era))

tierRefScriptFee

Arguments

∷ HasCallStack

⇒ Rational
Growth factor or step multiplier

→ Int
Increment size in which price grows linearly according to the price

→ Rational
Base fee. Currently this is customizable by ppMinFeeRefScriptCostPerByteL

→ Int
Total RefScript size in bytes

→ Coin

Calculate the fee for reference scripts using an exponential growth of the price per byte with linear increments

getConwayMinFeeTx ∷ (EraTx era, AlonzoEraTxWits era, ConwayEraPParams era) ⇒ PParams era → Tx era → Int → Coin

data family Tx era

Orphan instances

AlonzoEraTx ConwayEra

Instance details

Methods

isValidTxL ∷ Lens' (Tx ConwayEra) IsValid

EraTx ConwayEra

Instance details

Associated Types

data Tx ConwayEra

Methods

mkBasicTx ∷ TxBody ConwayEra → Tx ConwayEra

bodyTxL ∷ Lens' (Tx ConwayEra) (TxBody ConwayEra)

witsTxL ∷ Lens' (Tx ConwayEra) (TxWits ConwayEra)

auxDataTxL ∷ Lens' (Tx ConwayEra) (StrictMaybe (TxAuxData ConwayEra))

sizeTxF ∷ SimpleGetter (Tx ConwayEra) Word32

sizeTxForFeeCalculation ∷ Tx ConwayEra → Word32

validateNativeScript ∷ Tx ConwayEra → NativeScript ConwayEra → Bool

getMinFeeTx ∷ PParams ConwayEra → Tx ConwayEra → Int → Coin

Generic (Tx ConwayEra)

Instance details

Associated Types

type Rep (Tx ConwayEra) ∷ Type → Type

Methods

from ∷ Tx ConwayEra → Rep (Tx ConwayEra) x

to ∷ Rep (Tx ConwayEra) x → Tx ConwayEra

Show (Tx ConwayEra)

Instance details

Methods

showsPrec ∷ Int → Tx ConwayEra → ShowS

show ∷ Tx ConwayEra → String

showList ∷ [Tx ConwayEra] → ShowS

ToCBOR (Tx ConwayEra)

Instance details

Methods

toCBOR ∷ Tx ConwayEra → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy (Tx ConwayEra) → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [Tx ConwayEra] → Size

DecCBOR (Annotator (Tx ConwayEra))

Instance details

Methods

decCBOR ∷ Decoder s (Annotator (Tx ConwayEra))

dropCBOR ∷ Proxy (Annotator (Tx ConwayEra)) → Decoder s ()

label ∷ Proxy (Annotator (Tx ConwayEra)) → Text

EncCBOR (Tx ConwayEra)

Instance details

Methods

encCBOR ∷ Tx ConwayEra → Encoding

encodedSizeExpr ∷ (∀ t. EncCBOR t ⇒ Proxy t → Size) → Proxy (Tx ConwayEra) → Size

encodedListSizeExpr ∷ (∀ t. EncCBOR t ⇒ Proxy t → Size) → Proxy [Tx ConwayEra] → Size

EqRaw (Tx ConwayEra)

Instance details

Methods

eqRaw ∷ Tx ConwayEra → Tx ConwayEra → Bool

NFData (Tx ConwayEra)

Instance details

Methods

rnf ∷ Tx ConwayEra → ()

Eq (Tx ConwayEra)

Instance details

Methods

(==) ∷ Tx ConwayEra → Tx ConwayEra → Bool

(/=) ∷ Tx ConwayEra → Tx ConwayEra → Bool

NoThunks (Tx ConwayEra)

Instance details

Methods

noThunks ∷ Context → Tx ConwayEra → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → Tx ConwayEra → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (Tx ConwayEra) → String