Cardano.Ledger.Conway.TxBody.Internal

Description

Provides Conway TxBody internals

Warning

This module is considered internal.

The contents of this module may change in any way whatsoever and without any warning between minor versions of this package.

Documentation

class (BabbageEraTxBody era, ConwayEraTxCert era, ConwayEraPParams era, ConwayEraScript era) ⇒ ConwayEraTxBody era where

Methods

currentTreasuryValueTxBodyL ∷ Lens' (TxBody era) (StrictMaybe Coin)

Lens for getting and setting number of Coin that is expected to be in the Treasury at the current Epoch

votingProceduresTxBodyL ∷ Lens' (TxBody era) (VotingProcedures era)

Lens for getting and setting VotingProcedures.

proposalProceduresTxBodyL ∷ Lens' (TxBody era) (OSet (ProposalProcedure era))

Lens for getting and setting ProposalProcedures.

treasuryDonationTxBodyL ∷ Lens' (TxBody era) Coin

newtype ConwayTxBody era

Constructors

TxBodyConstr (MemoBytes ConwayTxBodyRaw era)

Bundled Patterns

pattern ConwayTxBody ∷ ConwayEraTxBody era ⇒ Set (TxIn (EraCrypto era)) → Set (TxIn (EraCrypto era)) → Set (TxIn (EraCrypto era)) → StrictSeq (Sized (TxOut era)) → StrictMaybe (Sized (TxOut era)) → StrictMaybe Coin → OSet (ConwayTxCert era) → Withdrawals (EraCrypto era) → Coin → ValidityInterval → Set (KeyHash 'Witness (EraCrypto era)) → MultiAsset (EraCrypto era) → StrictMaybe (ScriptIntegrityHash (EraCrypto era)) → StrictMaybe (AuxiliaryDataHash (EraCrypto era)) → StrictMaybe Network → VotingProcedures era → OSet (ProposalProcedure era) → StrictMaybe Coin → Coin → ConwayTxBody era

data ConwayTxBodyRaw era

Constructors

ConwayTxBodyRaw

Fields

ctbrSpendInputs ∷ !(Set (TxIn (EraCrypto era)))

ctbrCollateralInputs ∷ !(Set (TxIn (EraCrypto era)))

ctbrReferenceInputs ∷ !(Set (TxIn (EraCrypto era)))

ctbrOutputs ∷ !(StrictSeq (Sized (TxOut era)))

ctbrCollateralReturn ∷ !(StrictMaybe (Sized (TxOut era)))

ctbrTotalCollateral ∷ !(StrictMaybe Coin)

ctbrCerts ∷ !(OSet (ConwayTxCert era))

ctbrWithdrawals ∷ !(Withdrawals (EraCrypto era))

ctbrTxfee ∷ !Coin

ctbrVldt ∷ !ValidityInterval

ctbrReqSignerHashes ∷ !(Set (KeyHash 'Witness (EraCrypto era)))

ctbrMint ∷ !(MultiAsset (EraCrypto era))

ctbrScriptIntegrityHash ∷ !(StrictMaybe (ScriptIntegrityHash (EraCrypto era)))

ctbrAuxDataHash ∷ !(StrictMaybe (AuxiliaryDataHash (EraCrypto era)))

ctbrTxNetworkId ∷ !(StrictMaybe Network)

ctbrVotingProcedures ∷ !(VotingProcedures era)

ctbrProposalProcedures ∷ !(OSet (ProposalProcedure era))

ctbrCurrentTreasuryValue ∷ !(StrictMaybe Coin)

ctbrTreasuryDonation ∷ !Coin

conwayTotalDepositsTxBody ∷ ConwayEraTxBody era ⇒ PParams era → (KeyHash 'StakePool (EraCrypto era) → Bool) → TxBody era → Coin

Compute all the deposits in a TxBody. This includes deposits for:

registering Stake

registering a StakePool

registering a DRep

submitting a Proposal

This is the contribution of a TxBody towards the total Obligations

conwayProposalsDeposits ∷ ConwayEraTxBody era ⇒ PParams era → TxBody era → Coin

Total number of deposits in the proposals in TxBody

Orphan instances

Crypto c ⇒ AllegraEraTxBody (ConwayEra c)

Instance details

Methods

vldtTxBodyL ∷ Lens' (TxBody (ConwayEra c)) ValidityInterval

Crypto c ⇒ AlonzoEraTxBody (ConwayEra c)

Instance details

Methods

collateralInputsTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (Set (TxIn (EraCrypto (ConwayEra c))))

reqSignerHashesTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (Set (KeyHash 'Witness (EraCrypto (ConwayEra c))))

scriptIntegrityHashTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (StrictMaybe (ScriptIntegrityHash (EraCrypto (ConwayEra c))))

networkIdTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (StrictMaybe Network)

redeemerPointer ∷ TxBody (ConwayEra c) → PlutusPurpose AsItem (ConwayEra c) → StrictMaybe (PlutusPurpose AsIx (ConwayEra c))

redeemerPointerInverse ∷ TxBody (ConwayEra c) → PlutusPurpose AsIx (ConwayEra c) → StrictMaybe (PlutusPurpose AsIxItem (ConwayEra c))

Crypto c ⇒ BabbageEraTxBody (ConwayEra c)

Instance details

Methods

sizedOutputsTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (StrictSeq (Sized (TxOut (ConwayEra c))))

referenceInputsTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (Set (TxIn (EraCrypto (ConwayEra c))))

totalCollateralTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (StrictMaybe Coin)

collateralReturnTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (StrictMaybe (TxOut (ConwayEra c)))

sizedCollateralReturnTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (StrictMaybe (Sized (TxOut (ConwayEra c))))

allSizedOutputsTxBodyF ∷ SimpleGetter (TxBody (ConwayEra c)) (StrictSeq (Sized (TxOut (ConwayEra c))))

Crypto c ⇒ EraTxBody (ConwayEra c)

Instance details

Associated Types

type TxBody (ConwayEra c) = (r ∷ Type)

type TxBodyUpgradeError (ConwayEra c)

Methods

mkBasicTxBody ∷ TxBody (ConwayEra c)

inputsTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (Set (TxIn (EraCrypto (ConwayEra c))))

outputsTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (StrictSeq (TxOut (ConwayEra c)))

feeTxBodyL ∷ Lens' (TxBody (ConwayEra c)) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (Withdrawals (EraCrypto (ConwayEra c)))

auxDataHashTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (StrictMaybe (AuxiliaryDataHash (EraCrypto (ConwayEra c))))

spendableInputsTxBodyF ∷ SimpleGetter (TxBody (ConwayEra c)) (Set (TxIn (EraCrypto (ConwayEra c))))

allInputsTxBodyF ∷ SimpleGetter (TxBody (ConwayEra c)) (Set (TxIn (EraCrypto (ConwayEra c))))

certsTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (StrictSeq (TxCert (ConwayEra c)))

getTotalDepositsTxBody ∷ PParams (ConwayEra c) → (KeyHash 'StakePool (EraCrypto (ConwayEra c)) → Bool) → TxBody (ConwayEra c) → Coin

getTotalRefundsTxBody ∷ PParams (ConwayEra c) → (Credential 'Staking (EraCrypto (ConwayEra c)) → Maybe Coin) → (Credential 'DRepRole (EraCrypto (ConwayEra c)) → Maybe Coin) → TxBody (ConwayEra c) → Coin

getGenesisKeyHashCountTxBody ∷ TxBody (ConwayEra c) → Int

upgradeTxBody ∷ TxBody (PreviousEra (ConwayEra c)) → Either (TxBodyUpgradeError (ConwayEra c)) (TxBody (ConwayEra c))

Crypto c ⇒ MaryEraTxBody (ConwayEra c)

Instance details

Methods

mintTxBodyL ∷ Lens' (TxBody (ConwayEra c)) (MultiAsset (EraCrypto (ConwayEra c)))

mintValueTxBodyF ∷ SimpleGetter (TxBody (ConwayEra c)) (Value (ConwayEra c))

mintedTxBodyF ∷ SimpleGetter (TxBody (ConwayEra c)) (Set (PolicyID (EraCrypto (ConwayEra c))))