Cardano.Ledger.Conway.TxInfo

Documentation

data ConwayContextError era

Constructors

BabbageContextError !(BabbageContextError era)

CertificateNotSupported !(TxCert era)

PlutusPurposeNotSupported !(PlutusPurpose AsItem era)

CurrentTreasuryFieldNotSupported !Coin

VotingProceduresFieldNotSupported !(VotingProcedures era)

ProposalProceduresFieldNotSupported !(OSet (ProposalProcedure era))

TreasuryDonationFieldNotSupported !Coin

transTxBodyWithdrawals ∷ EraTxBody era ⇒ TxBody era → Map Credential Lovelace

Translate all Withdrawals from within a TxBody

transTxCert ∷ ConwayEraTxCert era ⇒ TxCert era → TxCert

transDRepCred ∷ Credential 'DRepRole c → DRepCredential

transColdCommitteeCred ∷ Credential 'ColdCommitteeRole c → ColdCommitteeCredential

transHotCommitteeCred ∷ Credential 'HotCommitteeRole c → HotCommitteeCredential

transDelegatee ∷ Delegatee c → Delegatee

transDRep ∷ DRep c → DRep

transScriptPurpose ∷ (ConwayEraPlutusTxInfo l era, PlutusTxCert l ~ TxCert) ⇒ proxy l → ConwayPlutusPurpose AsIxItem era → Either (ContextError era) ScriptPurpose

In Conway we have Anchors in some certificates and all proposals. However, because we do not translate anchors to plutus context, it is not always possible to deduce which item the script purpose is responsible for, without also including the index for that item. For this reason starting with PlutusV3, besides the item, Certifying and Proposing also have an index. Moreover, other script purposes rely on Ledger Ord instances for types that dictate the order, so it might not be a good idea to pass that information to Plutus for those purposes.

transMap ∷ (t1 → k) → (t2 → v) → Map t1 t2 → Map k v

transTxInInfoV1 ∷ ∀ era. (Inject (BabbageContextError era) (ContextError era), Value era ~ MaryValue (EraCrypto era), BabbageEraTxOut era) ⇒ UTxO era → TxIn (EraCrypto era) → Either (ContextError era) TxInInfo

Given a TxIn, look it up in the UTxO. If it exists, translate it to the V1 context

transTxOutV1 ∷ ∀ era. (Inject (BabbageContextError era) (ContextError era), Value era ~ MaryValue (EraCrypto era), BabbageEraTxOut era) ⇒ TxOutSource (EraCrypto era) → TxOut era → Either (ContextError era) TxOut

Given a TxOut, translate it for V2 and return (Right transalation). If the transaction contains any Byron addresses or Babbage features, return Left.

Orphan instances

Crypto c ⇒ EraPlutusTxInfo 'PlutusV1 (ConwayEra c)

Instance details

Methods

toPlutusTxCert ∷ proxy 'PlutusV1 → TxCert (ConwayEra c) → Either (ContextError (ConwayEra c)) (PlutusTxCert 'PlutusV1)

toPlutusScriptPurpose ∷ proxy 'PlutusV1 → PlutusPurpose AsIxItem (ConwayEra c) → Either (ContextError (ConwayEra c)) (PlutusScriptPurpose 'PlutusV1)

toPlutusTxInfo ∷ proxy 'PlutusV1 → PParams (ConwayEra c) → EpochInfo (Either Text) → SystemStart → UTxO (ConwayEra c) → Tx (ConwayEra c) → Either (ContextError (ConwayEra c)) (PlutusTxInfo 'PlutusV1)

toPlutusScriptContext ∷ proxy 'PlutusV1 → PlutusTxInfo 'PlutusV1 → PlutusPurpose AsIxItem (ConwayEra c) → Either (ContextError (ConwayEra c)) (PlutusScriptContext 'PlutusV1)

Crypto c ⇒ EraPlutusTxInfo 'PlutusV2 (ConwayEra c)

Instance details

Methods

toPlutusTxCert ∷ proxy 'PlutusV2 → TxCert (ConwayEra c) → Either (ContextError (ConwayEra c)) (PlutusTxCert 'PlutusV2)

toPlutusScriptPurpose ∷ proxy 'PlutusV2 → PlutusPurpose AsIxItem (ConwayEra c) → Either (ContextError (ConwayEra c)) (PlutusScriptPurpose 'PlutusV2)

toPlutusTxInfo ∷ proxy 'PlutusV2 → PParams (ConwayEra c) → EpochInfo (Either Text) → SystemStart → UTxO (ConwayEra c) → Tx (ConwayEra c) → Either (ContextError (ConwayEra c)) (PlutusTxInfo 'PlutusV2)

toPlutusScriptContext ∷ proxy 'PlutusV2 → PlutusTxInfo 'PlutusV2 → PlutusPurpose AsIxItem (ConwayEra c) → Either (ContextError (ConwayEra c)) (PlutusScriptContext 'PlutusV2)

Crypto c ⇒ EraPlutusTxInfo 'PlutusV3 (ConwayEra c)

Instance details

Methods

toPlutusTxCert ∷ proxy 'PlutusV3 → TxCert (ConwayEra c) → Either (ContextError (ConwayEra c)) (PlutusTxCert 'PlutusV3)

toPlutusScriptPurpose ∷ proxy 'PlutusV3 → PlutusPurpose AsIxItem (ConwayEra c) → Either (ContextError (ConwayEra c)) (PlutusScriptPurpose 'PlutusV3)

toPlutusTxInfo ∷ proxy 'PlutusV3 → PParams (ConwayEra c) → EpochInfo (Either Text) → SystemStart → UTxO (ConwayEra c) → Tx (ConwayEra c) → Either (ContextError (ConwayEra c)) (PlutusTxInfo 'PlutusV3)

toPlutusScriptContext ∷ proxy 'PlutusV3 → PlutusTxInfo 'PlutusV3 → PlutusPurpose AsIxItem (ConwayEra c) → Either (ContextError (ConwayEra c)) (PlutusScriptContext 'PlutusV3)

Crypto c ⇒ ConwayEraPlutusTxInfo 'PlutusV3 (ConwayEra c)

Instance details

Methods

toPlutusChangedParameters ∷ proxy 'PlutusV3 → PParamsUpdate (ConwayEra c) → ChangedParameters

Crypto c ⇒ EraPlutusContext (ConwayEra c)

Instance details

Associated Types

type ContextError (ConwayEra c) = (r ∷ Type)

Methods

mkPlutusScriptContext ∷ PlutusScript (ConwayEra c) → PlutusPurpose AsIxItem (ConwayEra c) → PParams (ConwayEra c) → EpochInfo (Either Text) → SystemStart → UTxO (ConwayEra c) → Tx (ConwayEra c) → Either (ContextError (ConwayEra c)) (Data (ConwayEra c))

Crypto c ⇒ ToPlutusData (PParamsUpdate (ConwayEra c))

Instance details

Methods

toPlutusData ∷ PParamsUpdate (ConwayEra c) → Data

fromPlutusData ∷ Data → Maybe (PParamsUpdate (ConwayEra c))