Cardano.Ledger.Conway.UTxO

Documentation

conwayConsumed ∷ (EraUTxO era, ConwayEraCertState era) ⇒ PParams era → CertState era → UTxO era → TxBody era → Value era

conwayProducedValue ∷ (ConwayEraTxBody era, Value era ~ MaryValue) ⇒ PParams era → (KeyHash 'StakePool → Bool) → TxBody era → Value era

getConwayWitsVKeyNeeded ∷ (EraTx era, ConwayEraTxBody era) ⇒ UTxO era → TxBody era → Set (KeyHash 'Witness)

getConwayScriptsNeeded ∷ ConwayEraTxBody era ⇒ UTxO era → TxBody era → AlonzoScriptsNeeded era

txNonDistinctRefScriptsSize ∷ (EraTx era, BabbageEraTxBody era) ⇒ UTxO era → Tx era → Int

Calculate the total size of reference scripts used by the transactions. Duplicate scripts will be counted as many times as they occur, since there is never a reason to include an input with the same reference script.

Any input that appears in both regular inputs and reference inputs of a transaction is only used once in this computation.

getConwayMinFeeTxUtxo ∷ (EraTx era, BabbageEraTxBody era) ⇒ PParams era → Tx era → UTxO era → Coin

Orphan instances

AlonzoEraUTxO ConwayEra

Instance details

Methods

getSupplementalDataHashes ∷ UTxO ConwayEra → TxBody ConwayEra → Set DataHash

getSpendingDatum ∷ UTxO ConwayEra → Tx ConwayEra → PlutusPurpose AsItem ConwayEra → Maybe (Data ConwayEra)

EraUTxO ConwayEra

Instance details

Associated Types

type ScriptsNeeded ConwayEra = (r ∷ Type)

Methods

consumed ∷ PParams ConwayEra → CertState ConwayEra → UTxO ConwayEra → TxBody ConwayEra → Value ConwayEra

getConsumedValue ∷ PParams ConwayEra → (Credential 'Staking → Maybe Coin) → (Credential 'DRepRole → Maybe Coin) → UTxO ConwayEra → TxBody ConwayEra → Value ConwayEra

getProducedValue ∷ PParams ConwayEra → (KeyHash 'StakePool → Bool) → TxBody ConwayEra → Value ConwayEra

getScriptsProvided ∷ UTxO ConwayEra → Tx ConwayEra → ScriptsProvided ConwayEra

getScriptsNeeded ∷ UTxO ConwayEra → TxBody ConwayEra → ScriptsNeeded ConwayEra

getScriptsHashesNeeded ∷ ScriptsNeeded ConwayEra → Set ScriptHash

getWitsVKeyNeeded ∷ CertState ConwayEra → UTxO ConwayEra → TxBody ConwayEra → Set (KeyHash 'Witness)

getMinFeeTxUtxo ∷ PParams ConwayEra → Tx ConwayEra → UTxO ConwayEra → Coin