Test.Cardano.Ledger.Conway.ImpTest

Documentation

module Test.Cardano.Ledger.Alonzo.ImpTest

class (AlonzoEraImp era, ConwayEraGov era, ConwayEraTxBody era, STS (EraRule "ENACT" era), BaseM (EraRule "ENACT" era) ~ ShelleyBase, State (EraRule "ENACT" era) ~ EnactState era, Signal (EraRule "ENACT" era) ~ EnactSignal era, Environment (EraRule "ENACT" era) ~ (), NativeScript era ~ Timelock era, Script era ~ AlonzoScript era) ⇒ ConwayEraImp era

enactConstitution ∷ ∀ era. (ConwayEraImp era, HasCallStack) ⇒ StrictMaybe (GovPurposeId 'ConstitutionPurpose era) → Constitution era → Credential 'DRepRole (EraCrypto era) → Credential 'HotCommitteeRole (EraCrypto era) → ImpTestM era (GovActionId (EraCrypto era))

submitGovAction ∷ ∀ era. (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ GovAction era → ImpTestM era (GovActionId (EraCrypto era))

submitGovAction_ ∷ ∀ era. (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ GovAction era → ImpTestM era ()

submitGovActions ∷ ∀ era. (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ NonEmpty (GovAction era) → ImpTestM era (NonEmpty (GovActionId (EraCrypto era)))

submitProposal ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ ProposalProcedure era → ImpTestM era (GovActionId (EraCrypto era))

submitProposal_ ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ ProposalProcedure era → ImpTestM era ()

submitProposals ∷ (ShelleyEraImp era, ConwayEraGov era, ConwayEraTxBody era, HasCallStack) ⇒ NonEmpty (ProposalProcedure era) → ImpTestM era (NonEmpty (GovActionId (EraCrypto era)))

submitFailingProposal ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ ProposalProcedure era → NonEmpty (PredicateFailure (EraRule "LEDGER" era)) → ImpTestM era ()

trySubmitGovAction ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ GovAction era → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) (GovActionId (EraCrypto era)))

Submits a transaction that proposes the given governance action. For proposing multiple actions in the same transaciton use trySubmitGovActions instead.

trySubmitGovActions ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ NonEmpty (GovAction era) → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) (Tx era))

Submits a transaction that proposes the given governance action

trySubmitProposal ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ ProposalProcedure era → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) (GovActionId (EraCrypto era)))

Submits a transaction that proposes the given proposal

trySubmitProposals ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ NonEmpty (ProposalProcedure era) → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) (Tx era))

submitTreasuryWithdrawals ∷ (ShelleyEraImp era, ConwayEraTxBody era, ConwayEraGov era) ⇒ [(RewardAccount (EraCrypto era), Coin)] → ImpTestM era (GovActionId (EraCrypto era))

submitVote ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ Vote → Voter (EraCrypto era) → GovActionId (EraCrypto era) → ImpTestM era (TxId (EraCrypto era))

Submits a transaction with a Vote for the given governance action as some voter

submitVote_ ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ Vote → Voter (EraCrypto era) → GovActionId (EraCrypto era) → ImpTestM era ()

submitYesVote_ ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ Voter (EraCrypto era) → GovActionId (EraCrypto era) → ImpTestM era ()

Submits a transaction that votes Yes for the given governance action as some voter

submitFailingVote ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ Voter (EraCrypto era) → GovActionId (EraCrypto era) → NonEmpty (PredicateFailure (EraRule "LEDGER" era)) → ImpTestM era ()

trySubmitVote ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ Vote → Voter (EraCrypto era) → GovActionId (EraCrypto era) → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) (TxId (EraCrypto era)))

Submits a transaction that votes Yes for the given governance action as some voter, and expects an Either result.

registerDRep ∷ ∀ era. (ShelleyEraImp era, ConwayEraTxCert era) ⇒ ImpTestM era (KeyHash 'DRepRole (EraCrypto era))

Submit a transaction that registers a new DRep and return the keyhash belonging to that DRep

setupSingleDRep ∷ ∀ era. (ConwayEraTxCert era, ShelleyEraImp era) ⇒ Integer → ImpTestM era (KeyHash 'DRepRole (EraCrypto era), KeyHash 'Staking (EraCrypto era), KeyPair 'Payment (EraCrypto era))

Registers a new DRep and delegates the specified amount of ADA to it.

setupPoolWithStake ∷ (ShelleyEraImp era, ConwayEraTxCert era) ⇒ Coin → ImpTestM era (KeyHash 'StakePool (EraCrypto era), Credential 'Payment (EraCrypto era), Credential 'Staking (EraCrypto era))

Sets up a stake pool with coin delegated to it.

NOTE: This uses the RegDepositDelegTxCert for delegating, so it has to be in Conway. The Shelley version of this function would have to separately register the staking credential and then delegate it.

conwayModifyPParams ∷ ConwayEraGov era ⇒ (PParams era → PParams era) → ImpTestM era ()

Modify the PParams in the current state with the given function

getProposals ∷ ConwayEraGov era ⇒ ImpTestM era (Proposals era)

getEnactState ∷ ConwayEraGov era ⇒ ImpTestM era (EnactState era)

getGovActionState ∷ (HasCallStack, ConwayEraGov era) ⇒ GovActionId (EraCrypto era) → ImpTestM era (GovActionState era)

Looks up the governance action state corresponding to the governance action id

lookupGovActionState ∷ ConwayEraGov era ⇒ GovActionId (EraCrypto era) → ImpTestM era (Maybe (GovActionState era))

Looks up the governance action state corresponding to the governance action id

expectPresentGovActionId ∷ (HasCallStack, ConwayEraGov era) ⇒ GovActionId (EraCrypto era) → ImpTestM era ()

expectMissingGovActionId ∷ (HasCallStack, ConwayEraGov era) ⇒ GovActionId (EraCrypto era) → ImpTestM era ()

getRatifyEnv ∷ ConwayEraGov era ⇒ ImpTestM era (RatifyEnv era)

Builds a RatifyEnv from the current state

calculateDRepAcceptedRatio ∷ ∀ era. (HasCallStack, ConwayEraGov era) ⇒ GovActionId (EraCrypto era) → ImpTestM era Rational

Calculates the ratio of DReps that have voted for the governance action

calculatePoolAcceptedRatio ∷ ConwayEraGov era ⇒ GovActionId (EraCrypto era) → ImpTestM era Rational

calculateCommitteeAcceptedRatio ∷ ∀ era. (HasCallStack, ConwayEraGov era) ⇒ GovActionId (EraCrypto era) → ImpTestM era Rational

Calculates the ratio of Committee members that have voted for the governance action

logAcceptedRatio ∷ (HasCallStack, ConwayEraGov era) ⇒ GovActionId (EraCrypto era) → ImpTestM era ()

Logs the ratios of accepted votes per category

isDRepAccepted ∷ (HasCallStack, ConwayEraGov era, ConwayEraPParams era) ⇒ GovActionId (EraCrypto era) → ImpTestM era Bool

Checks whether the governance action has enough DRep votes to be accepted in the next epoch. (Note that no other checks execept DRep votes is used)

isSpoAccepted ∷ (HasCallStack, ConwayEraGov era, ConwayEraPParams era) ⇒ GovActionId (EraCrypto era) → ImpTestM era Bool

logRatificationChecks ∷ (ConwayEraGov era, ConwayEraPParams era) ⇒ GovActionId (EraCrypto era) → ImpTestM era ()

Logs the results of each check required to make the governance action pass

resignCommitteeColdKey ∷ (ShelleyEraImp era, ConwayEraTxCert era) ⇒ Credential 'ColdCommitteeRole (EraCrypto era) → StrictMaybe (Anchor (EraCrypto era)) → ImpTestM era ()

Submits a transaction that resigns the cold key

registerCommitteeHotKey ∷ (ShelleyEraImp era, ConwayEraTxCert era) ⇒ Credential 'ColdCommitteeRole (EraCrypto era) → ImpTestM era (Credential 'HotCommitteeRole (EraCrypto era))

Submits a transaction that registers a hot key for the given cold key. Returns the hot key hash.

logCurPParams ∷ (EraGov era, ToExpr (PParamsHKD Identity era)) ⇒ ImpTestM era ()

electCommittee ∷ ∀ era. (HasCallStack, ConwayEraImp era) ⇒ StrictMaybe (GovPurposeId 'CommitteePurpose era) → KeyHash 'DRepRole (EraCrypto era) → Set (Credential 'ColdCommitteeRole (EraCrypto era)) → Map (Credential 'ColdCommitteeRole (EraCrypto era)) EpochNo → ImpTestM era (GovPurposeId 'CommitteePurpose era)

electBasicCommittee ∷ ∀ era. (HasCallStack, ConwayEraImp era) ⇒ ImpTestM era (Credential 'DRepRole (EraCrypto era), Credential 'HotCommitteeRole (EraCrypto era), GovPurposeId 'CommitteePurpose era)

proposalsShowDebug ∷ Era era ⇒ Proposals era → Bool → String

getGovPolicy ∷ ConwayEraGov era ⇒ ImpTestM era (StrictMaybe (ScriptHash (EraCrypto era)))

submitFailingGovAction ∷ ∀ era. (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ GovAction era → NonEmpty (PredicateFailure (EraRule "LEDGER" era)) → ImpTestM era ()

submitInitConstitutionGovAction ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ ImpTestM era (GovActionId (EraCrypto era))

submitChildConstitutionGovAction ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ GovActionId (EraCrypto era) → ImpTestM era (GovActionId (EraCrypto era))

submitConstitutionGovAction ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ StrictMaybe (GovPurposeId 'ConstitutionPurpose era) → ImpTestM era (GovActionId (EraCrypto era))

submitConstitutionGovActionTree ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ StrictMaybe (GovActionId (EraCrypto era)) → Tree () → ImpTestM era (Tree (GovActionId (EraCrypto era)))

submitConstitutionGovActionForest ∷ ConwayEraImp era ⇒ StrictMaybe (GovActionId (EraCrypto era)) → Forest () → ImpTestM era (Forest (GovActionId (EraCrypto era)))

getProposalsForest ∷ ConwayEraGov era ⇒ ImpTestM era (Forest (StrictMaybe (GovActionId (EraCrypto era))))

logProposalsForest ∷ ConwayEraGov era ⇒ ImpTestM era ()

logProposalsForestDiff ∷ (Era era, ToExpr (PParamsHKD StrictMaybe era)) ⇒ Proposals era → Proposals era → ImpTestM era ()

constitutionShouldBe ∷ (HasCallStack, ConwayEraGov era) ⇒ String → ImpTestM era ()

Asserts that the URL of the current constitution is equal to the given string

ccShouldBeResigned ∷ HasCallStack ⇒ Credential 'ColdCommitteeRole (EraCrypto era) → ImpTestM era ()

Test the resignation status for a CC cold key to be resigned

ccShouldNotBeResigned ∷ HasCallStack ⇒ Credential 'ColdCommitteeRole (EraCrypto era) → ImpTestM era ()

Test the resignation status for a CC cold key to not be resigned

getLastEnactedCommittee ∷ ConwayEraGov era ⇒ ImpTestM era (StrictMaybe (GovPurposeId 'CommitteePurpose era))

getLastEnactedConstitution ∷ ConwayEraGov era ⇒ ImpTestM era (StrictMaybe (GovPurposeId 'ConstitutionPurpose era))

submitParameterChange ∷ ConwayEraImp era ⇒ StrictMaybe (GovPurposeId 'PParamUpdatePurpose era) → PParamsUpdate era → ImpTestM era (GovActionId (EraCrypto era))

getLastEnactedParameterChange ∷ ConwayEraGov era ⇒ ImpTestM era (StrictMaybe (GovPurposeId 'PParamUpdatePurpose era))

getLastEnactedHardForkInitiation ∷ ConwayEraGov era ⇒ ImpTestM era (StrictMaybe (GovPurposeId 'HardForkPurpose era))

getConstitutionProposals ∷ ConwayEraGov era ⇒ ImpTestM era (Map (GovPurposeId 'ConstitutionPurpose era) (PEdges (GovPurposeId 'ConstitutionPurpose era)))

getParameterChangeProposals ∷ ConwayEraGov era ⇒ ImpTestM era (Map (GovPurposeId 'PParamUpdatePurpose era) (PEdges (GovPurposeId 'PParamUpdatePurpose era)))

expectNumDormantEpochs ∷ HasCallStack ⇒ EpochNo → ImpTestM era ()

submitConstitution ∷ ∀ era. ConwayEraImp era ⇒ StrictMaybe (GovPurposeId 'ConstitutionPurpose era) → ImpTestM era (GovActionId (EraCrypto era), Constitution era)

expectExtraDRepExpiry ∷ (HasCallStack, EraGov era, ConwayEraPParams era) ⇒ KeyHash 'DRepRole (EraCrypto era) → EpochNo → ImpTestM era ()

expectCurrentProposals ∷ (HasCallStack, ConwayEraGov era) ⇒ ImpTestM era ()

expectNoCurrentProposals ∷ (HasCallStack, ConwayEraImp era) ⇒ ImpTestM era ()

expectPulserProposals ∷ (HasCallStack, ConwayEraGov era) ⇒ ImpTestM era ()

expectNoPulserProposals ∷ (HasCallStack, ConwayEraGov era) ⇒ ImpTestM era ()

minorFollow ∷ ProtVer → ProtVer

A legal ProtVer that differs in the minor Version

majorFollow ∷ ProtVer → ProtVer

A legal ProtVer that moves to the next major Version

cantFollow ∷ ProtVer → ProtVer

An illegal ProtVer that skips 3 minor versions

Orphan instances

ShelleyEraImp (ConwayEra c) ⇒ AlonzoEraImp (ConwayEra c)

Instance details

Methods

scriptTestContexts ∷ Map (ScriptHash (EraCrypto (ConwayEra c))) ScriptTestContext

(Crypto c, NFData (SigDSIGN (DSIGN c)), NFData (VerKeyDSIGN (DSIGN c)), DSIGN c ~ Ed25519DSIGN, Signable (DSIGN c) (Hash (HASH c) EraIndependentTxBody), Eq (ConwayGovEvent (ConwayEra c))) ⇒ ShelleyEraImp (ConwayEra c)

Instance details

Methods

initImpNES ∷ NewEpochState (ConwayEra c)

impSatisfyNativeScript ∷ Set (KeyHash 'Witness (EraCrypto (ConwayEra c))) → NativeScript (ConwayEra c) → ImpTestM (ConwayEra c) (Maybe (Map (KeyHash 'Witness (EraCrypto (ConwayEra c))) (KeyPair 'Witness (EraCrypto (ConwayEra c)))))

modifyPParams ∷ (PParams (ConwayEra c) → PParams (ConwayEra c)) → ImpTestM (ConwayEra c) ()

fixupTx ∷ Tx (ConwayEra c) → ImpTestM (ConwayEra c) (Tx (ConwayEra c))