Cardano.Ledger.Conway.Governance

Documentation

class (EraPParams era, Eq (GovernanceState era), Show (GovernanceState era), NoThunks (GovernanceState era), NFData (GovernanceState era), EncCBOR (GovernanceState era), DecCBOR (GovernanceState era), ToCBOR (GovernanceState era), FromCBOR (GovernanceState era), Default (GovernanceState era), ToJSON (GovernanceState era)) ⇒ EraGovernance era where

Minimal complete definition

Nothing

Associated Types

type GovernanceState era = (r ∷ Type) | r → era

Methods

emptyGovernanceState ∷ GovernanceState era

Construct empty governance state

getProposedPPUpdates ∷ GovernanceState era → Maybe (ProposedPPUpdates era)

Returns Nothing for all eras starting with Conway, otherwise returns proposed pparams updates

getConstitutionHash ∷ GovernanceState era → Maybe (SafeHash (EraCrypto era) ByteString)

Returns Nothing for all era preceding Conway, otherwise returns the hash of the constitution

newtype ConwayGovState era

Constructors

ConwayGovState

Fields

unConwayGovState ∷ Map (GovernanceActionId (EraCrypto era)) (GovernanceActionState era)

data EnactState era

Constructors

EnactState

Fields

ensCommittee ∷ !(StrictMaybe (Set (KeyHash 'Voting (EraCrypto era)), Rational))

Constitutional Committee

ensConstitution ∷ !(SafeHash (EraCrypto era) ByteString)

Hash of the Constitution

ensProtVer ∷ !ProtVer

ensPParams ∷ !(PParams era)

ensTreasury ∷ !Coin

ensWithdrawals ∷ !(Map (Credential 'Staking (EraCrypto era)) Coin)

data RatifyState era

Constructors

RatifyState

Fields

rsEnactState ∷ !(EnactState era)

rsFuture ∷ !(StrictSeq (GovernanceActionId (EraCrypto era), GovernanceActionState era))

data ConwayGovernance era

Constructors

ConwayGovernance

Fields

cgGov ∷ !(ConwayGovState era)

cgRatify ∷ !(RatifyState era)

cgGovL ∷ Lens' (ConwayGovernance era) (ConwayGovState era)

cgRatifyL ∷ Lens' (ConwayGovernance era) (RatifyState era)

data GovernanceAction era

Constructors

ParameterChange !(PParamsUpdate era)

HardForkInitiation !ProtVer

TreasuryWithdrawals !(Map (Credential 'Staking (EraCrypto era)) Coin)

NoConfidence

NewCommittee !(Set (KeyHash 'Voting (EraCrypto era))) !Rational

NewConstitution !(SafeHash (EraCrypto era) ByteString)

InfoAction

data GovernanceActionState era

Constructors

GovernanceActionState

Fields

gasCommitteeVotes ∷ !(Map (Credential 'CommitteeHotKey (EraCrypto era)) Vote)

gasDRepVotes ∷ !(Map (Credential 'Voting (EraCrypto era)) Vote)

gasStakePoolVotes ∷ !(Map (KeyHash 'StakePool (EraCrypto era)) Vote)

gasDeposit ∷ !Coin

gasReturnAddr ∷ !(KeyHash 'Staking (EraCrypto era))

gasAction ∷ !(GovernanceAction era)

gasProposedIn ∷ !EpochNo

newtype GovernanceActionIx

Constructors

GovernanceActionIx Word64

data GovernanceActionId c

Constructors

GovernanceActionId

Fields

gaidTxId ∷ !(TxId c)

gaidGovActionIx ∷ !GovernanceActionIx

govActionIdToText ∷ GovernanceActionId c → Text

data Voter c

Constructors

CommitteeVoter !(Credential 'CommitteeHotKey c)

DRepVoter !(Credential 'Voting c)

StakePoolVoter !(KeyHash 'StakePool c)

data Vote

Constructors

VoteNo

VoteYes

Abstain

data VotingProcedure era

Constructors

VotingProcedure

Fields

vProcGovActionId ∷ !(GovernanceActionId (EraCrypto era))

vProcVoter ∷ !(Voter (EraCrypto era))

vProcVote ∷ !Vote

vProcAnchor ∷ !(StrictMaybe (Anchor (EraCrypto era)))

data ProposalProcedure era

Constructors

ProposalProcedure

Fields

pProcDeposit ∷ !Coin

pProcReturnAddr ∷ !(KeyHash 'Staking (EraCrypto era))

pProcGovernanceAction ∷ !(GovernanceAction era)

pProcAnchor ∷ !(Anchor (EraCrypto era))

data GovernanceProcedures era

Constructors

GovernanceProcedures

Fields

gpVotingProcedures ∷ !(Seq (VotingProcedure era))

gpProposalProcedures ∷ !(Seq (ProposalProcedure era))

data Anchor c

Constructors

Anchor

Fields

anchorUrl ∷ !Url

anchorDataHash ∷ !(SafeHash c AnchorDataHash)

data AnchorDataHash

ensConstitutionL ∷ Lens' (EnactState era) (SafeHash (EraCrypto era) ByteString)

rsEnactStateL ∷ Lens' (RatifyState era) (EnactState era)

Orphan instances

EraPParams (ConwayEra c) ⇒ EraGovernance (ConwayEra c)

Instance details

Associated Types

type GovernanceState (ConwayEra c) = (r ∷ Type)

Methods

emptyGovernanceState ∷ GovernanceState (ConwayEra c)

getProposedPPUpdates ∷ GovernanceState (ConwayEra c) → Maybe (ProposedPPUpdates (ConwayEra c))

getConstitutionHash ∷ GovernanceState (ConwayEra c) → Maybe (SafeHash (EraCrypto (ConwayEra c)) ByteString)