Cardano.Ledger.Conway.Governance.DRepPulser

Documentation

data DRepPulsingState era

Constructors

DRPulsing !(DRepPulser era Identity (RatifyState era))

DRComplete !(PulsingSnapshot era) !(RatifyState era)

data DRepPulser era (m ∷ Type → Type) ans where

The type of a Pulser which uses computeDRepDistr as its underlying function. Note that we use two type equality (~) constraints to fix both the monad m and the ans type, to the context where we will use the type as a Pulser. The type DRepPulser must have m and ans as its last two parameters so we can make a Pulsable instance. We will always use this instantiation (DRepPulser era Identity (RatifyState era))

Constructors

DRepPulser

Fields

∷ ∀ era ans m. (ans ~ RatifyState era, m ~ Identity, RunConwayRatify era)

⇒ { dpPulseSize ∷ !Int

How many elements of dpUMap to consume each pulse.

, dpUMap ∷ !UMap

Snapshot containing the mapping of stake credentials to DReps, Pools and Rewards.

, dpIndex ∷ !Int

The index of the iterator over dpUMap. Grows with each pulse.

, dpStakeDistr ∷ !(Map (Credential 'Staking) (CompactForm Coin))

Snapshot of the stake distr (comes from the IncrementalStake)

, dpStakePoolDistr ∷ PoolDistr

Snapshot of the pool distr. Lazy on purpose: See ssStakeMarkPoolDistr and ADR-7 for explanation.

, dpDRepDistr ∷ !(Map DRep (CompactForm Coin))

The partial result that grows with each pulse. The purpose of the pulsing.

, dpDRepState ∷ !(Map (Credential 'DRepRole) DRepState)

Snapshot of registered DRep credentials

, dpCurrentEpoch ∷ !EpochNo

Snapshot of the EpochNo this pulser will complete in.

, dpCommitteeState ∷ !(CommitteeState era)

Snapshot of the CommitteeState

, dpEnactState ∷ !(EnactState era)

Snapshot of the EnactState, Used to build the Env of the RATIFY rule

, dpProposals ∷ !(StrictSeq (GovActionState era))

Snapshot of the proposals. This is the Signal for the RATIFY rule

, dpProposalDeposits ∷ !(Map (Credential 'Staking) (CompactForm Coin))

Snapshot of the proposal-deposits per reward-account-staking-credential

, dpGlobals ∷ !Globals

, dpPoolParams ∷ !(Map (KeyHash 'StakePool) PoolParams)

Snapshot of the parameters of stake pools - this is needed to get the reward account for SPO vote calculation

} → DRepPulser era m ans

pulseDRepPulsingState ∷ DRepPulsingState era → DRepPulsingState era

completeDRepPulsingState ∷ DRepPulsingState era → DRepPulsingState era

extractDRepPulsingState ∷ DRepPulsingState era → RatifyState era

finishDRepPulser ∷ ∀ era. DRepPulsingState era → (PulsingSnapshot era, RatifyState era)

computeDRepDistr ∷ Map (Credential 'Staking) (CompactForm Coin) → Map (Credential 'DRepRole) DRepState → Map (Credential 'Staking) (CompactForm Coin) → PoolDistr → Map DRep (CompactForm Coin) → Map (Credential 'Staking) UMElem → (Map DRep (CompactForm Coin), PoolDistr)

We iterate over a pulse-sized chunk of the UMap.

For each staking credential in the chunk that has delegated to a DRep, add the stake distribution, rewards, and proposal deposits for that credential to the DRep distribution, if the DRep is a DRepCredential (also, AlwaysAbstain or AlwaysNoConfidence) and a member of the registered DReps. If the DRepCredential is not a member of the registered DReps, ignore and skip that DRep.

For each staking credential in the chunk that has delegated to an SPO, add only the proposal deposits for that credential to the stake pool distribution, since the rewards and stake are already added to it by the SNAP rule.

Give or take, this operation has roughly O (a * (log(b) + log(c) + log(d) + log(e) + log(f))) complexity, where, (a) is the size of the chunk of the UMap, which is the pulse-size, iterate over (b) is the size of the StakeDistr, lookup (c) is the size of the DRepDistr, insertWith (d) is the size of the dpProposalDeposits, lookup (e) is the size of the registered DReps, lookup (f) is the size of the PoolDistr, insert

psDRepDistrG ∷ SimpleGetter (DRepPulsingState era) (Map DRep (CompactForm Coin))

This is potentially an expensive getter. Make sure not to use it in the first 80% of the epoch.

data PulsingSnapshot era

A snapshot of information from the previous epoch stored inside the Pulser. After the pulser completes, but before the epoch turns, this information is store in the DRComplete constructor of the DRepPulsingState These are the values at the start of the current epoch. This allows the API To access these "previous" values, both during and after pulsing.

Constructors

PulsingSnapshot

Fields

psProposals ∷ !(StrictSeq (GovActionState era))

psDRepDistr ∷ !(Map DRep (CompactForm Coin))

psDRepState ∷ !(Map (Credential 'DRepRole) DRepState)

psPoolDistr ∷ Map (KeyHash 'StakePool) (CompactForm Coin)

psProposalsL ∷ Lens' (PulsingSnapshot era) (StrictSeq (GovActionState era))

psDRepDistrL ∷ Lens' (PulsingSnapshot era) (Map DRep (CompactForm Coin))

psDRepStateL ∷ Lens' (PulsingSnapshot era) (Map (Credential 'DRepRole) DRepState)

psPoolDistrL ∷ Lens' (PulsingSnapshot era) (Map (KeyHash 'StakePool) (CompactForm Coin))

class (STS (ConwayRATIFY era), Signal (ConwayRATIFY era) ~ RatifySignal era, BaseM (ConwayRATIFY era) ~ Reader Globals, Environment (ConwayRATIFY era) ~ RatifyEnv era, State (ConwayRATIFY era) ~ RatifyState era, PredicateFailure (ConwayRATIFY era) ~ Void) ⇒ RunConwayRatify era where

Minimal complete definition

Nothing

Methods

runConwayRatify ∷ Globals → RatifyEnv era → RatifyState era → RatifySignal era → RatifyState era