Cardano.Ledger.Conway.PParams

Description

This module contains the type of protocol parameters and EraPParams instance

Documentation

class BabbageEraPParams era ⇒ ConwayEraPParams era where

Methods

modifiedPPGroups ∷ PParamsUpdate era → Set PPGroups

ppuWellFormed ∷ ProtVer → PParamsUpdate era → Bool

hkdPoolVotingThresholdsL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f PoolVotingThresholds)

hkdDRepVotingThresholdsL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f DRepVotingThresholds)

hkdCommitteeMinSizeL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Natural)

hkdCommitteeMaxTermLengthL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f EpochInterval)

hkdGovActionLifetimeL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f EpochInterval)

hkdGovActionDepositL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

hkdDRepDepositL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f Coin)

hkdDRepActivityL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f EpochInterval)

hkdMinFeeRefScriptCostPerByteL ∷ HKDFunctor f ⇒ Lens' (PParamsHKD f era) (HKD f NonNegativeInterval)

ppPoolVotingThresholdsL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) PoolVotingThresholds

ppDRepVotingThresholdsL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) DRepVotingThresholds

ppCommitteeMinSizeL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) Natural

ppCommitteeMaxTermLengthL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) EpochInterval

ppGovActionLifetimeL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) EpochInterval

ppGovActionDepositL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) Coin

ppDRepDepositL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) Coin

ppDRepActivityL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) EpochInterval

ppMinFeeRefScriptCostPerByteL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParams era) NonNegativeInterval

ppuPoolVotingThresholdsL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe PoolVotingThresholds)

ppuDRepVotingThresholdsL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe DRepVotingThresholds)

ppuCommitteeMinSizeL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Natural)

ppuCommitteeMaxTermLengthL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

ppuGovActionLifetimeL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

ppuGovActionDepositL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

ppuDRepDepositL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe Coin)

ppuDRepActivityL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe EpochInterval)

ppuMinFeeRefScriptCostPerByteL ∷ ∀ era. ConwayEraPParams era ⇒ Lens' (PParamsUpdate era) (StrictMaybe NonNegativeInterval)

data PoolVotingThresholds

Constructors

PoolVotingThresholds

Fields

pvtMotionNoConfidence ∷ !UnitInterval

pvtCommitteeNormal ∷ !UnitInterval

pvtCommitteeNoConfidence ∷ !UnitInterval

pvtHardForkInitiation ∷ !UnitInterval

pvtPPSecurityGroup ∷ !UnitInterval

pvtCommitteeNoConfidenceL ∷ Lens' PoolVotingThresholds UnitInterval

pvtCommitteeNormalL ∷ Lens' PoolVotingThresholds UnitInterval

pvtPPSecurityGroupL ∷ Lens' PoolVotingThresholds UnitInterval

data DRepVotingThresholds

Constructors

DRepVotingThresholds

Fields

dvtMotionNoConfidence ∷ !UnitInterval

dvtCommitteeNormal ∷ !UnitInterval

dvtCommitteeNoConfidence ∷ !UnitInterval

dvtUpdateToConstitution ∷ !UnitInterval

dvtHardForkInitiation ∷ !UnitInterval

dvtPPNetworkGroup ∷ !UnitInterval

dvtPPEconomicGroup ∷ !UnitInterval

dvtPPTechnicalGroup ∷ !UnitInterval

dvtPPGovGroup ∷ !UnitInterval

dvtTreasuryWithdrawal ∷ !UnitInterval

dvtCommitteeNoConfidenceL ∷ Lens' DRepVotingThresholds UnitInterval

dvtCommitteeNormalL ∷ Lens' DRepVotingThresholds UnitInterval

dvtHardForkInitiationL ∷ Lens' DRepVotingThresholds UnitInterval

dvtMotionNoConfidenceL ∷ Lens' DRepVotingThresholds UnitInterval

dvtPPNetworkGroupL ∷ Lens' DRepVotingThresholds UnitInterval

dvtPPGovGroupL ∷ Lens' DRepVotingThresholds UnitInterval

dvtPPTechnicalGroupL ∷ Lens' DRepVotingThresholds UnitInterval

dvtPPEconomicGroupL ∷ Lens' DRepVotingThresholds UnitInterval

dvtTreasuryWithdrawalL ∷ Lens' DRepVotingThresholds UnitInterval

dvtUpdateToConstitutionL ∷ Lens' DRepVotingThresholds UnitInterval

data ConwayPParams f era

Conway Protocol parameters. The following parameters have been added since Babbage: * poolVotingThresholds * dRepVotingThresholds * committeeMinSize * committeeMaxTermLength * govActionLifetime * govActionDeposit * dRepDeposit * dRepActivity

Constructors

ConwayPParams

Fields

cppMinFeeA ∷ !(THKD ('PPGroups 'EconomicGroup 'SecurityGroup) f Coin)

The linear factor for the minimum fee calculation

cppMinFeeB ∷ !(THKD ('PPGroups 'EconomicGroup 'SecurityGroup) f Coin)

The constant factor for the minimum fee calculation

cppMaxBBSize ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f Word32)

Maximal block body size

cppMaxTxSize ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f Word32)

Maximal transaction size

cppMaxBHSize ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f Word16)

Maximal block header size

cppKeyDeposit ∷ !(THKD ('PPGroups 'EconomicGroup 'NoStakePoolGroup) f Coin)

The amount of a key registration deposit

cppPoolDeposit ∷ !(THKD ('PPGroups 'EconomicGroup 'NoStakePoolGroup) f Coin)

The amount of a pool registration deposit

cppEMax ∷ !(THKD ('PPGroups 'TechnicalGroup 'NoStakePoolGroup) f EpochInterval)

Maximum number of epochs in the future a pool retirement is allowed to be scheduled for.

cppNOpt ∷ !(THKD ('PPGroups 'TechnicalGroup 'NoStakePoolGroup) f Word16)

Desired number of pools

cppA0 ∷ !(THKD ('PPGroups 'TechnicalGroup 'NoStakePoolGroup) f NonNegativeInterval)

Pool influence

cppRho ∷ !(THKD ('PPGroups 'EconomicGroup 'NoStakePoolGroup) f UnitInterval)

Monetary expansion

cppTau ∷ !(THKD ('PPGroups 'EconomicGroup 'NoStakePoolGroup) f UnitInterval)

Treasury expansion

cppProtocolVersion ∷ !(HKDNoUpdate f ProtVer)

Protocol version

cppMinPoolCost ∷ !(THKD ('PPGroups 'EconomicGroup 'NoStakePoolGroup) f Coin)

Minimum Stake Pool Cost

cppCoinsPerUTxOByte ∷ !(THKD ('PPGroups 'EconomicGroup 'SecurityGroup) f CoinPerByte)

Cost in lovelace per byte of UTxO storage

cppCostModels ∷ !(THKD ('PPGroups 'TechnicalGroup 'NoStakePoolGroup) f CostModels)

Cost models for non-native script languages

cppPrices ∷ !(THKD ('PPGroups 'EconomicGroup 'NoStakePoolGroup) f Prices)

Prices of execution units (for non-native script languages)

cppMaxTxExUnits ∷ !(THKD ('PPGroups 'NetworkGroup 'NoStakePoolGroup) f OrdExUnits)

Max total script execution resources units allowed per tx

cppMaxBlockExUnits ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f OrdExUnits)

Max total script execution resources units allowed per block

cppMaxValSize ∷ !(THKD ('PPGroups 'NetworkGroup 'SecurityGroup) f Word32)

Max size of a Value in an output

cppCollateralPercentage ∷ !(THKD ('PPGroups 'TechnicalGroup 'NoStakePoolGroup) f Word16)

Percentage of the txfee which must be provided as collateral when including non-native scripts.

cppMaxCollateralInputs ∷ !(THKD ('PPGroups 'NetworkGroup 'NoStakePoolGroup) f Word16)

Maximum number of collateral inputs allowed in a transaction

cppPoolVotingThresholds ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f PoolVotingThresholds)

Thresholds for SPO votes

cppDRepVotingThresholds ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f DRepVotingThresholds)

Thresholds for DRep votes

cppCommitteeMinSize ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f Word16)

Minimum size of the Constitutional Committee

cppCommitteeMaxTermLength ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f EpochInterval)

The Constitutional Committee Term limit in number of Slots

cppGovActionLifetime ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f EpochInterval)

Gov action lifetime in number of Epochs

cppGovActionDeposit ∷ !(THKD ('PPGroups 'GovGroup 'SecurityGroup) f Coin)

The amount of the Gov Action deposit

cppDRepDeposit ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f Coin)

The amount of a DRep registration deposit

cppDRepActivity ∷ !(THKD ('PPGroups 'GovGroup 'NoStakePoolGroup) f EpochInterval)

The number of Epochs that a DRep can perform no activity without losing their Active status.

cppMinFeeRefScriptCostPerByte ∷ !(THKD ('PPGroups 'EconomicGroup 'SecurityGroup) f NonNegativeInterval)

Reference scripts fee for the minimum fee calculation

getLanguageView ∷ AlonzoEraPParams era ⇒ PParams era → Language → LangDepView

data LangDepView

Constructors

LangDepView

Fields

tag ∷ ByteString

params ∷ ByteString

encodeLangViews ∷ Set LangDepView → Encoding

upgradeConwayPParams ∷ ∀ f c. HKDApplicative f ⇒ UpgradeConwayPParams f → PParamsHKD f (BabbageEra c) → ConwayPParams f (ConwayEra c)

data UpgradeConwayPParams f

Constructors

UpgradeConwayPParams

Fields

ucppPoolVotingThresholds ∷ !(HKD f PoolVotingThresholds)

ucppDRepVotingThresholds ∷ !(HKD f DRepVotingThresholds)

ucppCommitteeMinSize ∷ !(HKD f Word16)

ucppCommitteeMaxTermLength ∷ !(HKD f EpochInterval)

ucppGovActionLifetime ∷ !(HKD f EpochInterval)

ucppGovActionDeposit ∷ !(HKD f Coin)

ucppDRepDeposit ∷ !(HKD f Coin)

ucppDRepActivity ∷ !(HKD f EpochInterval)

ucppMinFeeRefScriptCostPerByte ∷ !(HKD f NonNegativeInterval)

ucppPlutusV3CostModel ∷ !(HKD f CostModel)

toUpgradeConwayPParamsUpdatePairs ∷ KeyValue e a ⇒ UpgradeConwayPParams Identity → [a]

newtype THKD (t ∷ PPGroups) f a

HKD that is tagged with a group

Constructors

THKD

Fields

unTHKD ∷ HKD f a

data DRepGroup

Protocol parameter groups that dictate different thresholds for DReps.

Constructors

NetworkGroup

EconomicGroup

TechnicalGroup

GovGroup

data PPGroups

Constructors

PPGroups DRepGroup StakePoolGroup

data StakePoolGroup

Constructors

SecurityGroup

NoStakePoolGroup

conwayModifiedPPGroups ∷ ConwayPParams StrictMaybe era → Set PPGroups

pvtHardForkInitiationL ∷ Lens' PoolVotingThresholds UnitInterval

pvtMotionNoConfidenceL ∷ Lens' PoolVotingThresholds UnitInterval

Orphan instances

Crypto c ⇒ AlonzoEraPParams (ConwayEra c)

Instance details

Methods

hkdCoinsPerUTxOWordL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ExactEra AlonzoEra (ConwayEra c)) ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f CoinPerWord)

hkdCostModelsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f CostModels)

hkdPricesL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Prices)

hkdMaxTxExUnitsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f ExUnits)

hkdMaxBlockExUnitsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f ExUnits)

hkdMaxValSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

hkdCollateralPercentageL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

hkdMaxCollateralInputsL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

Crypto c ⇒ BabbageEraPParams (ConwayEra c)

Instance details

Methods

hkdCoinsPerUTxOByteL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f CoinPerByte)

Crypto c ⇒ EraPParams (ConwayEra c)

Instance details

Associated Types

type PParamsHKD f (ConwayEra c) = (r ∷ Type)

type UpgradePParams f (ConwayEra c)

type DowngradePParams f (ConwayEra c)

Methods

applyPPUpdates ∷ PParams (ConwayEra c) → PParamsUpdate (ConwayEra c) → PParams (ConwayEra c)

emptyPParamsIdentity ∷ PParamsHKD Identity (ConwayEra c)

emptyPParamsStrictMaybe ∷ PParamsHKD StrictMaybe (ConwayEra c)

upgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDApplicative f, EraPParams (PreviousEra (ConwayEra c))) ⇒ UpgradePParams f (ConwayEra c) → PParamsHKD f (PreviousEra (ConwayEra c)) → PParamsHKD f (ConwayEra c)

downgradePParamsHKD ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, EraPParams (PreviousEra (ConwayEra c))) ⇒ DowngradePParams f (ConwayEra c) → PParamsHKD f (ConwayEra c) → PParamsHKD f (PreviousEra (ConwayEra c))

hkdMinFeeAL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)

hkdMinFeeBL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)

hkdMaxBBSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Word32)

hkdMaxTxSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Word32)

hkdMaxBHSizeL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Word16)

hkdKeyDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)

hkdPoolDepositL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)

hkdEMaxL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f EpochInterval)

hkdNOptL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Natural)

hkdA0L ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f NonNegativeInterval)

hkdRhoL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f UnitInterval)

hkdTauL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f UnitInterval)

hkdDL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost (ConwayEra c) 6) ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f UnitInterval)

ppDG ∷ SimpleGetter (PParams (ConwayEra c)) UnitInterval

hkdExtraEntropyL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost (ConwayEra c) 6) ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Nonce)

hkdProtocolVersionL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost (ConwayEra c) 8) ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f ProtVer)

ppProtocolVersionL ∷ Lens' (PParams (ConwayEra c)) ProtVer

ppuProtocolVersionL ∷ Lens' (PParamsUpdate (ConwayEra c)) (StrictMaybe ProtVer)

hkdMinUTxOValueL ∷ ∀ (f ∷ Type → Type). (HKDFunctor f, ProtVerAtMost (ConwayEra c) 4) ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)

hkdMinPoolCostL ∷ ∀ (f ∷ Type → Type). HKDFunctor f ⇒ Lens' (PParamsHKD f (ConwayEra c)) (HKD f Coin)