Cardano.Ledger.Conway.State

Documentation

module Cardano.Ledger.Babbage.State

data ConwayAccountState era

Constructors

ConwayAccountState

Fields

casBalance ∷ !(CompactForm Coin)

Current balance of the account

casDeposit ∷ !(CompactForm Coin)

Deposit amount that was left when staking credential was registered

casStakePoolDelegation ∷ !(StrictMaybe (KeyHash 'StakePool))

Potential delegation to a stake pool

casDRepDelegation ∷ !(StrictMaybe DRep)

Potential delegation to a DRep

newtype ConwayAccounts era

Constructors

ConwayAccounts

Fields

caStates ∷ Map (Credential 'Staking) (ConwayAccountState era)

Map from a staking credential to the account state.

class EraAccounts era ⇒ ConwayEraAccounts era where

Minimal complete definition

dRepDelegationAccountStateL

Methods

mkConwayAccountState ∷ CompactForm Coin → AccountState era

default mkConwayAccountState ∷ AccountState era ~ ConwayAccountState era ⇒ CompactForm Coin → AccountState era

dRepDelegationAccountStateL ∷ Lens' (AccountState era) (Maybe DRep)

accountStateDelegatee ∷ ConwayEraAccounts era ⇒ AccountState era → Maybe Delegatee

registerConwayAccount

Arguments

∷ ConwayEraAccounts era

⇒ Credential 'Staking

→ CompactForm Coin
Deposit

→ Maybe Delegatee

→ Accounts era

→ Accounts era

unregisterConwayAccount

Arguments

∷ EraAccounts era

⇒ Credential 'Staking
Credential to unregister

→ Accounts era
Accounts to remove the account state from

→ (Maybe (AccountState era), Accounts era)
Returns Just whenever account was registered and Nothing otherwise. Produced Accounts will have the account state removed, if it was present there to begin with.

lookupDRepDelegation ∷ ConwayEraAccounts era ⇒ Credential 'Staking → Accounts era → Maybe DRep

data ConwayCertState era

Constructors

ConwayCertState

Fields

conwayCertVState ∷ !(VState era)

conwayCertPState ∷ !(PState era)

conwayCertDState ∷ !(DState era)

class (EraCertState era, ConwayEraAccounts era) ⇒ ConwayEraCertState era where

Methods

certVStateL ∷ Lens' (CertState era) (VState era)

csCommitteeCredsL ∷ Lens' (CommitteeState era) (Map (Credential 'ColdCommitteeRole) CommitteeAuthorization)

epochStateRegDrepL ∷ ConwayEraCertState era ⇒ Lens' (EpochState era) (Map (Credential 'DRepRole) DRepState)

mkConwayCertState ∷ ConwayEraCertState era ⇒ VState era → PState era → DState era → CertState era

conwayCertDStateL ∷ Lens' (ConwayCertState era) (DState era)

conwayCertPStateL ∷ Lens' (ConwayCertState era) (PState era)

conwayCertVStateL ∷ Lens' (ConwayCertState era) (VState era)

conwayObligationCertState ∷ ConwayEraCertState era ⇒ CertState era → Obligations

conwayCertsTotalDepositsTxBody ∷ EraTxBody era ⇒ PParams era → ConwayCertState era → TxBody era → Coin

conwayCertsTotalRefundsTxBody ∷ (EraTxBody era, EraAccounts era) ⇒ PParams era → ConwayCertState era → TxBody era → Coin

newtype ConwayInstantStake era

Constructors

ConwayInstantStake

Fields

cisCredentialStake ∷ Map (Credential 'Staking) (CompactForm Coin)

conwayInstantStakeCredentialsL ∷ Lens' (ConwayInstantStake era) (Map (Credential 'Staking) (CompactForm Coin))

addConwayInstantStake ∷ EraTxOut era ⇒ UTxO era → ConwayInstantStake era → ConwayInstantStake era

deleteConwayInstantStake ∷ EraTxOut era ⇒ UTxO era → ConwayInstantStake era → ConwayInstantStake era

resolveConwayInstantStake ∷ (EraStake era, InstantStake era ~ ConwayInstantStake era) ⇒ ConwayInstantStake era → Accounts era → Stake

data VState era

The state that tracks the voting entities (DReps and Constitutional Committee members). In the formal ledger specification this type is called GState

Constructors

VState

Fields

vsDReps ∷ !(Map (Credential 'DRepRole) DRepState)

vsCommitteeState ∷ !(CommitteeState era)

vsNumDormantEpochs ∷ !EpochNo

Number of contiguous epochs in which there are exactly zero active governance proposals to vote on. It is incremented in every EPOCH rule if the number of active governance proposals to vote on continues to be zero. It is reset to zero when a new governance action is successfully proposed. We need this counter in order to bump DRep expiries through dormant periods when DReps do not have an opportunity to vote on anything.

vsDRepsL ∷ Lens' (VState era) (Map (Credential 'DRepRole) DRepState)

vsCommitteeStateL ∷ Lens' (VState era) (CommitteeState era)

vsNumDormantEpochsL ∷ Lens' (VState era) EpochNo

vsActualDRepExpiry ∷ Credential 'DRepRole → VState era → Maybe EpochNo

lookupDepositVState ∷ VState era → Credential 'DRepRole → Maybe Coin

Function that looks up the deposit for currently registered DRep