Cardano.Ledger.Conway.Tx

Documentation

data AlonzoTx era

Constructors

AlonzoTx

Fields

body ∷ !(TxBody era)

wits ∷ !(TxWits era)

isValid ∷ !IsValid

auxiliaryData ∷ !(StrictMaybe (TxAuxData era))

class (EraTx era, AlonzoEraTxBody era, AlonzoEraTxWits era, AlonzoEraScript era) ⇒ AlonzoEraTx era where

Methods

isValidTxL ∷ Lens' (Tx era) IsValid

tierRefScriptFee

Arguments

∷ HasCallStack

⇒ Rational
Growth factor or step multiplier

→ Int
Increment size in which price grows linearly according to the price

→ Rational
Base fee. Currently this is customizable by ppMinFeeRefScriptCostPerByteL

→ Int
Total RefScript size in bytes

→ Coin

Calculate the fee for reference scripts using an expoential growth of the price per byte with linear increments

refScriptCostStride ∷ Int

25 KiB

refScriptCostMultiplier ∷ Rational

Orphan instances

Crypto c ⇒ AlonzoEraTx (ConwayEra c)

Instance details

Methods

isValidTxL ∷ Lens' (Tx (ConwayEra c)) IsValid

Crypto c ⇒ EraSegWits (ConwayEra c)

Instance details

Associated Types

type TxSeq (ConwayEra c) = (r ∷ Type)

Methods

fromTxSeq ∷ TxSeq (ConwayEra c) → StrictSeq (Tx (ConwayEra c))

toTxSeq ∷ StrictSeq (Tx (ConwayEra c)) → TxSeq (ConwayEra c)

hashTxSeq ∷ TxSeq (ConwayEra c) → Hash (HASH (EraCrypto (ConwayEra c))) EraIndependentBlockBody

numSegComponents ∷ Word64

Crypto c ⇒ EraTx (ConwayEra c)

Instance details

Associated Types

type Tx (ConwayEra c) = (r ∷ Type)

type TxUpgradeError (ConwayEra c)

Methods

mkBasicTx ∷ TxBody (ConwayEra c) → Tx (ConwayEra c)

bodyTxL ∷ Lens' (Tx (ConwayEra c)) (TxBody (ConwayEra c))

witsTxL ∷ Lens' (Tx (ConwayEra c)) (TxWits (ConwayEra c))

auxDataTxL ∷ Lens' (Tx (ConwayEra c)) (StrictMaybe (AuxiliaryData (ConwayEra c)))

sizeTxF ∷ SimpleGetter (Tx (ConwayEra c)) Integer

wireSizeTxF ∷ SimpleGetter (Tx (ConwayEra c)) Word32

validateNativeScript ∷ Tx (ConwayEra c) → NativeScript (ConwayEra c) → Bool

getMinFeeTx ∷ PParams (ConwayEra c) → Tx (ConwayEra c) → Int → Coin

upgradeTx ∷ Tx (PreviousEra (ConwayEra c)) → Either (TxUpgradeError (ConwayEra c)) (Tx (ConwayEra c))