Cardano.Ledger.Conway.TxBody

Documentation

class (BabbageEraTxBody era, ConwayEraTxCert era, ConwayEraPParams era, ConwayEraScript era) ⇒ ConwayEraTxBody era where

Methods

currentTreasuryValueTxBodyL ∷ Lens' (TxBody era) (StrictMaybe Coin)

Lens for getting and setting number of Coin that is expected to be in the Treasury at the current Epoch

votingProceduresTxBodyL ∷ Lens' (TxBody era) (VotingProcedures era)

Lens for getting and setting VotingProcedures.

proposalProceduresTxBodyL ∷ Lens' (TxBody era) (OSet (ProposalProcedure era))

Lens for getting and setting ProposalProcedures.

treasuryDonationTxBodyL ∷ Lens' (TxBody era) Coin

data family TxBody era

The body of a transaction.

data ConwayTxBodyRaw

Constructors

ConwayTxBodyRaw

Fields

ctbrSpendInputs ∷ !(Set TxIn)

ctbrCollateralInputs ∷ !(Set TxIn)

ctbrReferenceInputs ∷ !(Set TxIn)

ctbrOutputs ∷ !(StrictSeq (Sized (TxOut ConwayEra)))

ctbrCollateralReturn ∷ !(StrictMaybe (Sized (TxOut ConwayEra)))

ctbrTotalCollateral ∷ !(StrictMaybe Coin)

ctbrCerts ∷ !(OSet (TxCert ConwayEra))

ctbrWithdrawals ∷ !Withdrawals

ctbrFee ∷ !Coin

ctbrVldt ∷ !ValidityInterval

ctbrReqSignerHashes ∷ !(Set (KeyHash 'Witness))

ctbrMint ∷ !MultiAsset

ctbrScriptIntegrityHash ∷ !(StrictMaybe ScriptIntegrityHash)

ctbrAuxDataHash ∷ !(StrictMaybe TxAuxDataHash)

ctbrNetworkId ∷ !(StrictMaybe Network)

ctbrVotingProcedures ∷ !(VotingProcedures ConwayEra)

ctbrProposalProcedures ∷ !(OSet (ProposalProcedure ConwayEra))

ctbrCurrentTreasuryValue ∷ !(StrictMaybe Coin)

ctbrTreasuryDonation ∷ !Coin

conwayTotalDepositsTxBody ∷ PParams ConwayEra → (KeyHash 'StakePool → Bool) → TxBody ConwayEra → Coin

Compute all the deposits in a TxBody. This includes deposits for:

registering Stake

registering a StakePool

registering a DRep

submitting a Proposal

This is the contribution of a TxBody towards the total Obligations

conwayProposalsDeposits ∷ ConwayEraTxBody era ⇒ PParams era → TxBody era → Coin

Total number of deposits in the proposals in TxBody

conwayRedeemerPointer ∷ ∀ era. ConwayEraTxBody era ⇒ TxBody era → ConwayPlutusPurpose AsItem era → StrictMaybe (ConwayPlutusPurpose AsIx era)

conwayRedeemerPointerInverse ∷ ConwayEraTxBody era ⇒ TxBody era → ConwayPlutusPurpose AsIx era → StrictMaybe (ConwayPlutusPurpose AsIxItem era)

upgradeBabbageTxOut ∷ (Value era ~ Value (PreviousEra era), EraScript (PreviousEra era), EraScript era) ⇒ BabbageTxOut (PreviousEra era) → BabbageTxOut era

Orphan instances

AllegraEraTxBody ConwayEra

Instance details

Methods

vldtTxBodyL ∷ Lens' (TxBody ConwayEra) ValidityInterval

AlonzoEraTxBody ConwayEra

Instance details

Methods

collateralInputsTxBodyL ∷ Lens' (TxBody ConwayEra) (Set TxIn)

reqSignerHashesTxBodyL ∷ Lens' (TxBody ConwayEra) (Set (KeyHash 'Witness))

reqSignerHashesTxBodyG ∷ SimpleGetter (TxBody ConwayEra) (Set (KeyHash 'Witness))

scriptIntegrityHashTxBodyL ∷ Lens' (TxBody ConwayEra) (StrictMaybe ScriptIntegrityHash)

networkIdTxBodyL ∷ Lens' (TxBody ConwayEra) (StrictMaybe Network)

redeemerPointer ∷ TxBody ConwayEra → PlutusPurpose AsItem ConwayEra → StrictMaybe (PlutusPurpose AsIx ConwayEra)

redeemerPointerInverse ∷ TxBody ConwayEra → PlutusPurpose AsIx ConwayEra → StrictMaybe (PlutusPurpose AsIxItem ConwayEra)

BabbageEraTxBody ConwayEra

Instance details

Methods

sizedOutputsTxBodyL ∷ Lens' (TxBody ConwayEra) (StrictSeq (Sized (TxOut ConwayEra)))

referenceInputsTxBodyL ∷ Lens' (TxBody ConwayEra) (Set TxIn)

totalCollateralTxBodyL ∷ Lens' (TxBody ConwayEra) (StrictMaybe Coin)

collateralReturnTxBodyL ∷ Lens' (TxBody ConwayEra) (StrictMaybe (TxOut ConwayEra))

sizedCollateralReturnTxBodyL ∷ Lens' (TxBody ConwayEra) (StrictMaybe (Sized (TxOut ConwayEra)))

allSizedOutputsTxBodyF ∷ SimpleGetter (TxBody ConwayEra) (StrictSeq (Sized (TxOut ConwayEra)))

EraTxBody ConwayEra

Instance details

Associated Types

data TxBody ConwayEra

Methods

mkBasicTxBody ∷ TxBody ConwayEra

inputsTxBodyL ∷ Lens' (TxBody ConwayEra) (Set TxIn)

outputsTxBodyL ∷ Lens' (TxBody ConwayEra) (StrictSeq (TxOut ConwayEra))

feeTxBodyL ∷ Lens' (TxBody ConwayEra) Coin

withdrawalsTxBodyL ∷ Lens' (TxBody ConwayEra) Withdrawals

auxDataHashTxBodyL ∷ Lens' (TxBody ConwayEra) (StrictMaybe TxAuxDataHash)

spendableInputsTxBodyF ∷ SimpleGetter (TxBody ConwayEra) (Set TxIn)

allInputsTxBodyF ∷ SimpleGetter (TxBody ConwayEra) (Set TxIn)

certsTxBodyL ∷ Lens' (TxBody ConwayEra) (StrictSeq (TxCert ConwayEra))

getTotalDepositsTxBody ∷ PParams ConwayEra → (KeyHash 'StakePool → Bool) → TxBody ConwayEra → Coin

getTotalRefundsTxBody ∷ PParams ConwayEra → (Credential 'Staking → Maybe Coin) → (Credential 'DRepRole → Maybe Coin) → TxBody ConwayEra → Coin

getGenesisKeyHashCountTxBody ∷ TxBody ConwayEra → Int

MaryEraTxBody ConwayEra

Instance details

Methods

mintTxBodyL ∷ Lens' (TxBody ConwayEra) MultiAsset

mintedTxBodyF ∷ SimpleGetter (TxBody ConwayEra) (Set PolicyID)

mintValueTxBodyF ∷ SimpleGetter (TxBody ConwayEra) (Value ConwayEra)

Generic (TxBody ConwayEra)

Instance details

Associated Types

type Rep (TxBody ConwayEra) ∷ Type → Type

Methods

from ∷ TxBody ConwayEra → Rep (TxBody ConwayEra) x

to ∷ Rep (TxBody ConwayEra) x → TxBody ConwayEra

Show (TxBody ConwayEra)

Instance details

Methods

showsPrec ∷ Int → TxBody ConwayEra → ShowS

show ∷ TxBody ConwayEra → String

showList ∷ [TxBody ConwayEra] → ShowS

ToCBOR (TxBody ConwayEra)

Instance details

Methods

toCBOR ∷ TxBody ConwayEra → Encoding

encodedSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy (TxBody ConwayEra) → Size

encodedListSizeExpr ∷ (∀ t. ToCBOR t ⇒ Proxy t → Size) → Proxy [TxBody ConwayEra] → Size

DecCBOR (Annotator (TxBody ConwayEra))

Instance details

Methods

decCBOR ∷ Decoder s (Annotator (TxBody ConwayEra))

dropCBOR ∷ Proxy (Annotator (TxBody ConwayEra)) → Decoder s ()

label ∷ Proxy (Annotator (TxBody ConwayEra)) → Text

EncCBOR (TxBody ConwayEra)
Encodes memoized bytes created upon construction.

Instance details

Methods

encCBOR ∷ TxBody ConwayEra → Encoding

encodedSizeExpr ∷ (∀ t. EncCBOR t ⇒ Proxy t → Size) → Proxy (TxBody ConwayEra) → Size

encodedListSizeExpr ∷ (∀ t. EncCBOR t ⇒ Proxy t → Size) → Proxy [TxBody ConwayEra] → Size

SafeToHash (TxBody ConwayEra)

Instance details

Methods

originalBytes ∷ TxBody ConwayEra → ByteString

originalBytesSize ∷ TxBody ConwayEra → Int

makeHashWithExplicitProxys ∷ Proxy i → TxBody ConwayEra → SafeHash i

EqRaw (TxBody ConwayEra)

Instance details

Methods

eqRaw ∷ TxBody ConwayEra → TxBody ConwayEra → Bool

Memoized (TxBody ConwayEra)

Instance details

Associated Types

type RawType (TxBody ConwayEra) = (r ∷ Type)

Methods

getMemoBytes ∷ TxBody ConwayEra → MemoBytes (RawType (TxBody ConwayEra))

wrapMemoBytes ∷ MemoBytes (RawType (TxBody ConwayEra)) → TxBody ConwayEra

NFData (TxBody ConwayEra)

Instance details

Methods

rnf ∷ TxBody ConwayEra → ()

Eq (TxBody ConwayEra)

Instance details

Methods

(==) ∷ TxBody ConwayEra → TxBody ConwayEra → Bool

(/=) ∷ TxBody ConwayEra → TxBody ConwayEra → Bool

NoThunks (TxBody ConwayEra)

Instance details

Methods

noThunks ∷ Context → TxBody ConwayEra → IO (Maybe ThunkInfo)

wNoThunks ∷ Context → TxBody ConwayEra → IO (Maybe ThunkInfo)

showTypeOf ∷ Proxy (TxBody ConwayEra) → String

HashAnnotated (TxBody ConwayEra) EraIndependentTxBody

Instance details

Methods

hashAnnotated ∷ TxBody ConwayEra → SafeHash EraIndependentTxBody