Cardano.Ledger.Conway.TxBody

Documentation

class (BabbageEraTxBody era, ConwayEraTxCert era, ConwayEraPParams era, ConwayEraScript era) ⇒ ConwayEraTxBody era where

Methods

currentTreasuryValueTxBodyL ∷ Lens' (TxBody era) (StrictMaybe Coin)

Lens for getting and setting number of Coin that is expected to be in the Treasury at the current Epoch

votingProceduresTxBodyL ∷ Lens' (TxBody era) (VotingProcedures era)

Lens for getting and setting VotingProcedures.

proposalProceduresTxBodyL ∷ Lens' (TxBody era) (OSet (ProposalProcedure era))

Lens for getting and setting ProposalProcedures.

treasuryDonationTxBodyL ∷ Lens' (TxBody era) Coin

data ConwayTxBody era where

Bundled Patterns

pattern ConwayTxBody ∷ ConwayEraTxBody era ⇒ Set (TxIn (EraCrypto era)) → Set (TxIn (EraCrypto era)) → Set (TxIn (EraCrypto era)) → StrictSeq (Sized (TxOut era)) → StrictMaybe (Sized (TxOut era)) → StrictMaybe Coin → OSet (ConwayTxCert era) → Withdrawals (EraCrypto era) → Coin → ValidityInterval → Set (KeyHash 'Witness (EraCrypto era)) → MultiAsset (EraCrypto era) → StrictMaybe (ScriptIntegrityHash (EraCrypto era)) → StrictMaybe (AuxiliaryDataHash (EraCrypto era)) → StrictMaybe Network → VotingProcedures era → OSet (ProposalProcedure era) → StrictMaybe Coin → Coin → ConwayTxBody era

data ConwayTxBodyRaw era

conwayTotalDepositsTxBody ∷ ConwayEraTxBody era ⇒ PParams era → (KeyHash 'StakePool (EraCrypto era) → Bool) → TxBody era → Coin

Compute all the deposits in a TxBody. This includes deposits for:

registering Stake

registering a StakePool

registering a DRep

submitting a Proposal

This is the contribution of a TxBody towards the total Obligations

conwayProposalsDeposits ∷ ConwayEraTxBody era ⇒ PParams era → TxBody era → Coin

Total number of deposits in the proposals in TxBody