Cardano.Ledger.Conway.TxCert

Documentation

data ConwayTxCert era

Constructors

ConwayTxCertDeleg !(ConwayDelegCert (EraCrypto era))

ConwayTxCertPool !(PoolCert (EraCrypto era))

ConwayTxCertConstitutional !(ConstitutionalDelegCert (EraCrypto era))

data ConwayDelegCert c

Certificates for registration and delegation of stake to Pools and DReps. Comparing to previous eras, there is now ability to:

Register and delegate with a single certificate: ConwayRegDelegCert

Ability to delegate to DReps with DelegVote and DelegStakeVote

Ability to specify the deposit amount. Deposits during registration and unregistration in Conway are optional, which will change in the future era. They are optional only for the smooth transition from Babbage to Conway. Validity of deposits is checked by the CERT rule.

Constructors

ConwayRegCert !(StakeCredential c) !(StrictMaybe Coin)
Register staking credential. Deposit, when present, must match the expected deposit amount specified by ppKeyDepositL in the protocol parameters.

ConwayUnRegCert !(StakeCredential c) !(StrictMaybe Coin)
De-Register the staking credential. Deposit, if present, must match the amount that was left as a deposit upon stake credential registration.

ConwayDelegCert !(StakeCredential c) !(Delegatee c)
Redelegate to another delegatee. Staking credential must already be registered.

ConwayRegDelegCert !(StakeCredential c) !(Delegatee c) !Coin
This is a new type of certificate, which allows to register staking credential and delegate within a single certificate. Deposit is required and must match the expected deposit amount specified by ppKeyDepositL in the protocol parameters.

data Delegatee c

First type argument is the deposit

Constructors

DelegStake !(KeyHash 'StakePool c)

DelegVote !(Credential 'Voting c)

DelegStakeVote !(KeyHash 'StakePool c) !(Credential 'Voting c)

class ShelleyEraTxCert era ⇒ ConwayEraTxCert era where

Methods

mkConwayTxCertDeleg ∷ ConwayDelegCert (EraCrypto era) → TxCert era

getConwayTxCertDeleg ∷ TxCert era → Maybe (ConwayDelegCert (EraCrypto era))

fromShelleyDelegCert ∷ ShelleyDelegCert c → ConwayDelegCert c

toShelleyDelegCert ∷ ConwayDelegCert c → Maybe (ShelleyDelegCert c)

Orphan instances

Crypto c ⇒ EraTxCert (ConwayEra c)

Instance details

Associated Types

type TxCert (ConwayEra c) = (r ∷ Type)

Methods

mkTxCertPool ∷ PoolCert (EraCrypto (ConwayEra c)) → TxCert (ConwayEra c)

getTxCertPool ∷ TxCert (ConwayEra c) → Maybe (PoolCert (EraCrypto (ConwayEra c)))

mkTxCertGenesis ∷ ConstitutionalDelegCert (EraCrypto (ConwayEra c)) → TxCert (ConwayEra c)

getTxCertGenesis ∷ TxCert (ConwayEra c) → Maybe (ConstitutionalDelegCert (EraCrypto (ConwayEra c)))

Crypto c ⇒ ShelleyEraTxCert (ConwayEra c)

Instance details

Methods

mkTxCertMir ∷ MIRCert (EraCrypto (ConwayEra c)) → TxCert (ConwayEra c)

getTxCertMir ∷ TxCert (ConwayEra c) → Maybe (MIRCert (EraCrypto (ConwayEra c)))

mkShelleyTxCertDeleg ∷ ShelleyDelegCert (EraCrypto (ConwayEra c)) → TxCert (ConwayEra c)

getShelleyTxCertDeleg ∷ TxCert (ConwayEra c) → Maybe (ShelleyDelegCert (EraCrypto (ConwayEra c)))