Cardano.Ledger.Conway.TxCert

Documentation

data ConwayTxCert era

Constructors

ConwayTxCertDeleg !(ConwayDelegCert (EraCrypto era))

ConwayTxCertPool !(PoolCert (EraCrypto era))

ConwayTxCertCommittee !(ConwayCommitteeCert (EraCrypto era))

data ConwayDelegCert c

Certificates for registration and delegation of stake to Pools and DReps. Comparing to previous eras, there is now ability to:

Register and delegate with a single certificate: ConwayRegDelegCert

Ability to delegate to DReps with DelegVote and DelegStakeVote

Ability to specify the deposit amount. Deposits during registration and unregistration in Conway are optional, which will change in the future era. They are optional only for the smooth transition from Babbage to Conway. Validity of deposits is checked by the CERT rule.

Constructors

ConwayRegCert !(StakeCredential c) !(StrictMaybe Coin)
Register staking credential. Deposit, when present, must match the expected deposit amount specified by ppKeyDepositL in the protocol parameters.

ConwayUnRegCert !(StakeCredential c) !(StrictMaybe Coin)
De-Register the staking credential. Deposit, if present, must match the amount that was left as a deposit upon stake credential registration.

ConwayDelegCert !(StakeCredential c) !(Delegatee c)
Redelegate to another delegatee. Staking credential must already be registered.

ConwayRegDelegCert !(StakeCredential c) !(Delegatee c) !Coin
This is a new type of certificate, which allows to register staking credential and delegate within a single certificate. Deposit is required and must match the expected deposit amount specified by ppKeyDepositL in the protocol parameters.

data ConwayCommitteeCert c

Constructors

ConwayRegDRep !(Credential 'Voting c) !Coin

ConwayUnRegDRep !(Credential 'Voting c) !Coin

ConwayAuthCommitteeHotKey !(KeyHash 'CommitteeColdKey c) !(KeyHash 'CommitteeHotKey c)

ConwayResignCommitteeColdKey !(KeyHash 'CommitteeColdKey c)

data Delegatee c

First type argument is the deposit

Constructors

DelegStake !(KeyHash 'StakePool c)

DelegVote !(DRep c)

DelegStakeVote !(KeyHash 'StakePool c) !(DRep c)

class ShelleyEraTxCert era ⇒ ConwayEraTxCert era where

Methods

mkRegDepositTxCert ∷ StakeCredential (EraCrypto era) → Coin → TxCert era

getRegDepositTxCert ∷ TxCert era → Maybe (StakeCredential (EraCrypto era), Coin)

mkUnRegDepositTxCert ∷ StakeCredential (EraCrypto era) → Coin → TxCert era

getUnRegDepositTxCert ∷ TxCert era → Maybe (StakeCredential (EraCrypto era), Coin)

mkDelegTxCert ∷ StakeCredential (EraCrypto era) → Delegatee (EraCrypto era) → TxCert era

getDelegTxCert ∷ TxCert era → Maybe (StakeCredential (EraCrypto era), Delegatee (EraCrypto era))

mkRegDepositDelegTxCert ∷ StakeCredential (EraCrypto era) → Delegatee (EraCrypto era) → Coin → TxCert era

getRegDepositDelegTxCert ∷ TxCert era → Maybe (StakeCredential (EraCrypto era), Delegatee (EraCrypto era), Coin)

mkAuthCommitteeHotKeyTxCert ∷ KeyHash 'CommitteeColdKey (EraCrypto era) → KeyHash 'CommitteeHotKey (EraCrypto era) → TxCert era

getAuthCommitteeHotKeyTxCert ∷ TxCert era → Maybe (KeyHash 'CommitteeColdKey (EraCrypto era), KeyHash 'CommitteeHotKey (EraCrypto era))

mkResignCommitteeColdTxCert ∷ KeyHash 'CommitteeColdKey (EraCrypto era) → TxCert era

getResignCommitteeColdTxCert ∷ TxCert era → Maybe (KeyHash 'CommitteeColdKey (EraCrypto era))

mkRegDRepTxCert ∷ Credential 'Voting (EraCrypto era) → Coin → TxCert era

getRegDRepTxCert ∷ TxCert era → Maybe (Credential 'Voting (EraCrypto era), Coin)

mkUnRegDRepTxCert ∷ Credential 'Voting (EraCrypto era) → Coin → TxCert era

getUnRegDRepTxCert ∷ TxCert era → Maybe (Credential 'Voting (EraCrypto era), Coin)

fromShelleyDelegCert ∷ ShelleyDelegCert c → ConwayDelegCert c

toShelleyDelegCert ∷ ConwayDelegCert c → Maybe (ShelleyDelegCert c)

getScriptWitnessConwayTxCert ∷ ConwayTxCert era → Maybe (ScriptHash (EraCrypto era))

getVKeyWitnessConwayTxCert ∷ ConwayTxCert era → Maybe (KeyHash 'Witness (EraCrypto era))

pattern RegDepositTxCert ∷ ConwayEraTxCert era ⇒ StakeCredential (EraCrypto era) → Coin → TxCert era

pattern UnRegDepositTxCert ∷ ConwayEraTxCert era ⇒ StakeCredential (EraCrypto era) → Coin → TxCert era

pattern DelegTxCert ∷ ConwayEraTxCert era ⇒ StakeCredential (EraCrypto era) → Delegatee (EraCrypto era) → TxCert era

pattern RegDepositDelegTxCert ∷ ConwayEraTxCert era ⇒ StakeCredential (EraCrypto era) → Delegatee (EraCrypto era) → Coin → TxCert era

pattern AuthCommitteeHotKeyTxCert ∷ ConwayEraTxCert era ⇒ KeyHash 'CommitteeColdKey (EraCrypto era) → KeyHash 'CommitteeHotKey (EraCrypto era) → TxCert era

pattern ResignCommitteeColdTxCert ∷ ConwayEraTxCert era ⇒ KeyHash 'CommitteeColdKey (EraCrypto era) → TxCert era

pattern RegDRepTxCert ∷ ConwayEraTxCert era ⇒ Credential 'Voting (EraCrypto era) → Coin → TxCert era

pattern UnRegDRepTxCert ∷ ConwayEraTxCert era ⇒ Credential 'Voting (EraCrypto era) → Coin → TxCert era

Orphan instances

Crypto c ⇒ EraTxCert (ConwayEra c)

Instance details

Associated Types

type TxCert (ConwayEra c) = (r ∷ Type)

Methods

getVKeyWitnessTxCert ∷ TxCert (ConwayEra c) → Maybe (KeyHash 'Witness (EraCrypto (ConwayEra c)))

getScriptWitnessTxCert ∷ TxCert (ConwayEra c) → Maybe (ScriptHash (EraCrypto (ConwayEra c)))

mkRegPoolTxCert ∷ PoolParams (EraCrypto (ConwayEra c)) → TxCert (ConwayEra c)

getRegPoolTxCert ∷ TxCert (ConwayEra c) → Maybe (PoolParams (EraCrypto (ConwayEra c)))

mkRetirePoolTxCert ∷ KeyHash 'StakePool (EraCrypto (ConwayEra c)) → EpochNo → TxCert (ConwayEra c)

getRetirePoolTxCert ∷ TxCert (ConwayEra c) → Maybe (KeyHash 'StakePool (EraCrypto (ConwayEra c)), EpochNo)

Crypto c ⇒ ShelleyEraTxCert (ConwayEra c)

Instance details

Methods

mkRegTxCert ∷ StakeCredential (EraCrypto (ConwayEra c)) → TxCert (ConwayEra c)

getRegTxCert ∷ TxCert (ConwayEra c) → Maybe (StakeCredential (EraCrypto (ConwayEra c)))

mkUnRegTxCert ∷ StakeCredential (EraCrypto (ConwayEra c)) → TxCert (ConwayEra c)

getUnRegTxCert ∷ TxCert (ConwayEra c) → Maybe (StakeCredential (EraCrypto (ConwayEra c)))

mkDelegStakeTxCert ∷ StakeCredential (EraCrypto (ConwayEra c)) → KeyHash 'StakePool (EraCrypto (ConwayEra c)) → TxCert (ConwayEra c)

getDelegStakeTxCert ∷ TxCert (ConwayEra c) → Maybe (StakeCredential (EraCrypto (ConwayEra c)), KeyHash 'StakePool (EraCrypto (ConwayEra c)))

mkGenesisDelegTxCert ∷ GenesisDelegCert (EraCrypto (ConwayEra c)) → TxCert (ConwayEra c)

getGenesisDelegTxCert ∷ TxCert (ConwayEra c) → Maybe (GenesisDelegCert (EraCrypto (ConwayEra c)))

mkMirTxCert ∷ MIRCert (EraCrypto (ConwayEra c)) → TxCert (ConwayEra c)

getMirTxCert ∷ TxCert (ConwayEra c) → Maybe (MIRCert (EraCrypto (ConwayEra c)))