Cardano.Ledger.Conway.TxInfo

Documentation

data ConwayContextError era

Constructors

BabbageContextError !(BabbageContextError era)

CertificateNotSupported !(TxCert era)

PlutusPurposeNotSupported !(PlutusPurpose AsItem era)

CurrentTreasuryFieldNotSupported !Coin

VotingProceduresFieldNotSupported !(VotingProcedures era)

ProposalProceduresFieldNotSupported !(OSet (ProposalProcedure era))

TreasuryDonationFieldNotSupported !Coin

ReferenceInputsNotDisjointFromInputs !(NonEmpty TxIn)

class EraPlutusTxInfo l era ⇒ ConwayEraPlutusTxInfo (l ∷ Language) era where

A class to compute the changed parameters in the TxInfo given a ToPlutusData instance for PParamsUpdate

Methods

toPlutusChangedParameters ∷ proxy l → PParamsUpdate era → ChangedParameters

transTxBodyWithdrawals ∷ EraTxBody era ⇒ TxBody era → Map Credential Lovelace

Translate all Withdrawals from within a TxBody

transTxCert ∷ (ShelleyEraTxCert era, ConwayEraTxCert era, TxCert era ~ ConwayTxCert era) ⇒ ProtVer → TxCert era → TxCert

In protocol version 9, a bug in RegTxCert and UnRegTxCert pattern definitions was causing the deposit in RegDepositTxCert and UnRegDepositTxCert to be omitted. We need to keep this behavior for version 9, so, now that the bug in the patterns has been fixed, we are explicitly omitting the deposit in these cases. It has been confirmed that this buggy behavior for protocol version 9 has been exercised on Mainnet, therefore this conditional translation can never be removed for Conway era (#4863)

transDRepCred ∷ Credential 'DRepRole → DRepCredential

transColdCommitteeCred ∷ Credential 'ColdCommitteeRole → ColdCommitteeCredential

transHotCommitteeCred ∷ Credential 'HotCommitteeRole → HotCommitteeCredential

transDelegatee ∷ Delegatee → Delegatee

transDRep ∷ DRep → DRep

transMap ∷ (t1 → k) → (t2 → v) → Map t1 t2 → Map k v

transTxInInfoV1 ∷ ∀ era. (Inject (BabbageContextError era) (ContextError era), Value era ~ MaryValue, BabbageEraTxOut era) ⇒ UTxO era → TxIn → Either (ContextError era) TxInInfo

Given a TxIn, look it up in the UTxO. If it exists, translate it to the V1 context

transTxOutV1 ∷ ∀ era. (Inject (BabbageContextError era) (ContextError era), Value era ~ MaryValue, BabbageEraTxOut era) ⇒ TxOutSource → TxOut era → Either (ContextError era) TxOut

Given a TxOut, translate it for V2 and return (Right transalation). If the transaction contains any Byron addresses or Babbage features, return Left.

transMintValue ∷ MultiAsset → MintValue

transTxBodyId ∷ EraTxBody era ⇒ TxBody era → TxId

transValidityInterval ∷ ∀ proxy era a. Inject (AlonzoContextError era) a ⇒ proxy era → EpochInfo (Either Text) → SystemStart → ValidityInterval → Either a POSIXTimeRange

Translate a validity interval to POSIX time

transVotingProcedures ∷ VotingProcedures era → Map Voter (Map GovernanceActionId Vote)

transProposal ∷ ConwayEraPlutusTxInfo l era ⇒ proxy l → ProposalProcedure era → ProposalProcedure

toPlutusV3Args ∷ EraPlutusTxInfo 'PlutusV3 era ⇒ proxy 'PlutusV3 → ProtVer → TxInfo → PlutusPurpose AsIxItem era → Maybe (Data era) → Data era → Either (ContextError era) (PlutusArgs 'PlutusV3)

transTxCertV1V2 ∷ (ShelleyEraTxCert era, ConwayEraTxCert era, Inject (ConwayContextError era) (ContextError era)) ⇒ TxCert era → Either (ContextError era) DCert

transPlutusPurposeV1V2 ∷ ∀ l era proxy. (PlutusTxCert l ~ DCert, EraPlutusTxInfo l era, Inject (ConwayContextError era) (ContextError era), Inject (ConwayPlutusPurpose AsItem era) (PlutusPurpose AsItem era)) ⇒ proxy l → ProtVer → ConwayPlutusPurpose AsItem era → Either (ContextError era) ScriptPurpose

transPlutusPurposeV3 ∷ (ConwayEraPlutusTxInfo l era, PlutusTxCert l ~ TxCert) ⇒ proxy l → ProtVer → ConwayPlutusPurpose AsIxItem era → Either (ContextError era) ScriptPurpose

In Conway we have Anchors in some certificates and all proposals. However, because we do not translate anchors to plutus context, it is not always possible to deduce which item the script purpose is responsible for, without also including the index for that item. For this reason starting with PlutusV3, besides the item, Certifying and Proposing also have an index. Moreover, other script purposes rely on Ledger Ord instances for types that dictate the order, so it might not be a good idea to pass that information to Plutus for those purposes.

guardConwayFeaturesForPlutusV1V2 ∷ ∀ era. (EraTx era, ConwayEraTxBody era, Inject (ConwayContextError era) (ContextError era)) ⇒ Tx era → Either (ContextError era) ()

transTxInInfoV3 ∷ ∀ era. (Inject (BabbageContextError era) (ContextError era), Value era ~ MaryValue, BabbageEraTxOut era) ⇒ UTxO era → TxIn → Either (ContextError era) TxInInfo

Given a TxIn, look it up in the UTxO. If it exists, translate it to the V3 context

scriptPurposeToScriptInfo ∷ ScriptPurpose → Maybe Datum → ScriptInfo

Orphan instances

EraPlutusContext ConwayEra

Instance details

Associated Types

type ContextError ConwayEra = (r ∷ Type)

data TxInfoResult ConwayEra

Methods

mkSupportedLanguage ∷ Language → Maybe (SupportedLanguage ConwayEra)

mkTxInfoResult ∷ LedgerTxInfo ConwayEra → TxInfoResult ConwayEra

lookupTxInfoResult ∷ ∀ (l ∷ Language). EraPlutusTxInfo l ConwayEra ⇒ SLanguage l → TxInfoResult ConwayEra → Either (ContextError ConwayEra) (PlutusTxInfo l)

mkPlutusWithContext ∷ PlutusScript ConwayEra → ScriptHash → PlutusPurpose AsIxItem ConwayEra → LedgerTxInfo ConwayEra → TxInfoResult ConwayEra → (Data ConwayEra, ExUnits) → CostModel → Either (ContextError ConwayEra) PlutusWithContext

EraPlutusTxInfo 'PlutusV1 ConwayEra

Instance details

Methods

toPlutusTxCert ∷ proxy 'PlutusV1 → ProtVer → TxCert ConwayEra → Either (ContextError ConwayEra) (PlutusTxCert 'PlutusV1)

toPlutusScriptPurpose ∷ proxy 'PlutusV1 → ProtVer → PlutusPurpose AsIxItem ConwayEra → Either (ContextError ConwayEra) (PlutusScriptPurpose 'PlutusV1)

toPlutusTxInfo ∷ proxy 'PlutusV1 → LedgerTxInfo ConwayEra → Either (ContextError ConwayEra) (PlutusTxInfo 'PlutusV1)

toPlutusArgs ∷ proxy 'PlutusV1 → ProtVer → PlutusTxInfo 'PlutusV1 → PlutusPurpose AsIxItem ConwayEra → Maybe (Data ConwayEra) → Data ConwayEra → Either (ContextError ConwayEra) (PlutusArgs 'PlutusV1)

toPlutusTxInInfo ∷ proxy 'PlutusV1 → UTxO ConwayEra → TxIn → Either (ContextError ConwayEra) (PlutusTxInInfo ConwayEra 'PlutusV1)

EraPlutusTxInfo 'PlutusV2 ConwayEra

Instance details

Methods

toPlutusTxCert ∷ proxy 'PlutusV2 → ProtVer → TxCert ConwayEra → Either (ContextError ConwayEra) (PlutusTxCert 'PlutusV2)

toPlutusScriptPurpose ∷ proxy 'PlutusV2 → ProtVer → PlutusPurpose AsIxItem ConwayEra → Either (ContextError ConwayEra) (PlutusScriptPurpose 'PlutusV2)

toPlutusTxInfo ∷ proxy 'PlutusV2 → LedgerTxInfo ConwayEra → Either (ContextError ConwayEra) (PlutusTxInfo 'PlutusV2)

toPlutusArgs ∷ proxy 'PlutusV2 → ProtVer → PlutusTxInfo 'PlutusV2 → PlutusPurpose AsIxItem ConwayEra → Maybe (Data ConwayEra) → Data ConwayEra → Either (ContextError ConwayEra) (PlutusArgs 'PlutusV2)

toPlutusTxInInfo ∷ proxy 'PlutusV2 → UTxO ConwayEra → TxIn → Either (ContextError ConwayEra) (PlutusTxInInfo ConwayEra 'PlutusV2)

EraPlutusTxInfo 'PlutusV3 ConwayEra

Instance details

Methods

toPlutusTxCert ∷ proxy 'PlutusV3 → ProtVer → TxCert ConwayEra → Either (ContextError ConwayEra) (PlutusTxCert 'PlutusV3)

toPlutusScriptPurpose ∷ proxy 'PlutusV3 → ProtVer → PlutusPurpose AsIxItem ConwayEra → Either (ContextError ConwayEra) (PlutusScriptPurpose 'PlutusV3)

toPlutusTxInfo ∷ proxy 'PlutusV3 → LedgerTxInfo ConwayEra → Either (ContextError ConwayEra) (PlutusTxInfo 'PlutusV3)

toPlutusArgs ∷ proxy 'PlutusV3 → ProtVer → PlutusTxInfo 'PlutusV3 → PlutusPurpose AsIxItem ConwayEra → Maybe (Data ConwayEra) → Data ConwayEra → Either (ContextError ConwayEra) (PlutusArgs 'PlutusV3)

toPlutusTxInInfo ∷ proxy 'PlutusV3 → UTxO ConwayEra → TxIn → Either (ContextError ConwayEra) (PlutusTxInInfo ConwayEra 'PlutusV3)