Cardano.Ledger.Conway.UTxO

Documentation

conwayProducedValue ∷ (ConwayEraTxBody era, Value era ~ MaryValue (EraCrypto era)) ⇒ PParams era → (KeyHash 'StakePool (EraCrypto era) → Bool) → TxBody era → Value era

getConwayWitsVKeyNeeded ∷ (EraTx era, ConwayEraTxBody era) ⇒ UTxO era → TxBody era → Set (KeyHash 'Witness (EraCrypto era))

txNonDistinctRefScriptsSize ∷ (EraTx era, BabbageEraTxBody era) ⇒ UTxO era → Tx era → Int

Calculate the total size of reference scripts used by the transactions. Duplicate scripts will be counted as many times as they occur, since there is never a reason to include an input with the same reference script.

Any input that appears in both regular inputs and reference inputs of a transaction is only used once in this computation.

Orphan instances

Crypto c ⇒ AlonzoEraUTxO (ConwayEra c)

Instance details

Methods

getSupplementalDataHashes ∷ UTxO (ConwayEra c) → TxBody (ConwayEra c) → Set (DataHash (EraCrypto (ConwayEra c)))

getSpendingDatum ∷ UTxO (ConwayEra c) → Tx (ConwayEra c) → PlutusPurpose AsItem (ConwayEra c) → Maybe (Data (ConwayEra c))

Crypto c ⇒ EraUTxO (ConwayEra c)

Instance details

Associated Types

type ScriptsNeeded (ConwayEra c) = (r ∷ Type)

Methods

getConsumedValue ∷ PParams (ConwayEra c) → (Credential 'Staking (EraCrypto (ConwayEra c)) → Maybe Coin) → (Credential 'DRepRole (EraCrypto (ConwayEra c)) → Maybe Coin) → UTxO (ConwayEra c) → TxBody (ConwayEra c) → Value (ConwayEra c)

getProducedValue ∷ PParams (ConwayEra c) → (KeyHash 'StakePool (EraCrypto (ConwayEra c)) → Bool) → TxBody (ConwayEra c) → Value (ConwayEra c)

getScriptsProvided ∷ UTxO (ConwayEra c) → Tx (ConwayEra c) → ScriptsProvided (ConwayEra c)

getScriptsNeeded ∷ UTxO (ConwayEra c) → TxBody (ConwayEra c) → ScriptsNeeded (ConwayEra c)

getScriptsHashesNeeded ∷ ScriptsNeeded (ConwayEra c) → Set (ScriptHash (EraCrypto (ConwayEra c)))

getWitsVKeyNeeded ∷ CertState (ConwayEra c) → UTxO (ConwayEra c) → TxBody (ConwayEra c) → Set (KeyHash 'Witness (EraCrypto (ConwayEra c)))

getMinFeeTxUtxo ∷ PParams (ConwayEra c) → Tx (ConwayEra c) → UTxO (ConwayEra c) → Coin