Test.Cardano.Ledger.Conway.ImpTest

Documentation

module Test.Cardano.Ledger.Babbage.ImpTest

class (BabbageEraImp era, ConwayEraTest era, STS (EraRule "ENACT" era), BaseM (EraRule "ENACT" era) ~ ShelleyBase, State (EraRule "ENACT" era) ~ EnactState era, Signal (EraRule "ENACT" era) ~ EnactSignal era, Environment (EraRule "ENACT" era) ~ (), NFData (Event (EraRule "ENACT" era)), ToExpr (Event (EraRule "ENACT" era)), Typeable (Event (EraRule "ENACT" era)), Eq (Event (EraRule "ENACT" era)), GovState era ~ ConwayGovState era, InjectRuleFailure "LEDGER" ConwayUtxowPredFailure era, InjectRuleFailure "LEDGER" ConwayUtxosPredFailure era, InjectRuleFailure "LEDGER" ConwayLedgerPredFailure era, InjectRuleFailure "LEDGER" ConwayCertsPredFailure era, InjectRuleFailure "LEDGER" ConwayDelegPredFailure era, InjectRuleFailure "LEDGER" ConwayGovPredFailure era, InjectRuleFailure "LEDGER" ConwayUtxoPredFailure era, InjectRuleFailure "LEDGER" ConwayGovCertPredFailure era, InjectRuleFailure "BBODY" ConwayBbodyPredFailure era, InjectRuleEvent "TICK" ConwayHardForkEvent era, InjectRuleEvent "TICK" ConwayEpochEvent era) ⇒ ConwayEraImp era

conwayModifyImpInitProtVer ∷ ∀ era. ConwayEraImp era ⇒ Version → SpecWith (ImpInit (LedgerSpec era)) → SpecWith (ImpInit (LedgerSpec era))

enactConstitution ∷ ∀ era. (ConwayEraImp era, HasCallStack) ⇒ StrictMaybe (GovPurposeId 'ConstitutionPurpose) → Constitution era → Credential 'DRepRole → NonEmpty (Credential 'HotCommitteeRole) → ImpTestM era GovActionId

enactTreasuryWithdrawals ∷ ConwayEraImp era ⇒ [(RewardAccount, Coin)] → Credential 'DRepRole → NonEmpty (Credential 'HotCommitteeRole) → ImpTestM era GovActionId

submitGovAction ∷ ∀ era. (ConwayEraImp era, HasCallStack) ⇒ GovAction era → ImpTestM era GovActionId

submitGovAction_ ∷ ∀ era. (ConwayEraImp era, HasCallStack) ⇒ GovAction era → ImpTestM era ()

submitGovActions ∷ ∀ era. (ConwayEraImp era, HasCallStack) ⇒ NonEmpty (GovAction era) → ImpTestM era (NonEmpty GovActionId)

submitProposal ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ ProposalProcedure era → ImpTestM era GovActionId

submitAndExpireProposalToMakeReward ∷ ConwayEraImp era ⇒ Credential 'Staking → ImpTestM era ()

submitProposal_ ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ ProposalProcedure era → ImpTestM era ()

submitProposals ∷ (ShelleyEraImp era, ConwayEraGov era, ConwayEraTxBody era, HasCallStack) ⇒ NonEmpty (ProposalProcedure era) → ImpTestM era (NonEmpty GovActionId)

submitFailingProposal ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ ProposalProcedure era → NonEmpty (PredicateFailure (EraRule "LEDGER" era)) → ImpTestM era ()

trySubmitGovAction ∷ ConwayEraImp era ⇒ GovAction era → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) GovActionId)

Submits a transaction that proposes the given governance action. For proposing multiple actions in the same transaciton use trySubmitGovActions instead.

trySubmitGovActions ∷ ConwayEraImp era ⇒ NonEmpty (GovAction era) → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era)), Tx era) (Tx era))

Submits a transaction that proposes the given governance action

trySubmitProposal ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ ProposalProcedure era → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) GovActionId)

Submits a transaction that proposes the given proposal

trySubmitProposals ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ NonEmpty (ProposalProcedure era) → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era)), Tx era) (Tx era))

mkConstitutionProposal ∷ ConwayEraImp era ⇒ StrictMaybe (GovPurposeId 'ConstitutionPurpose) → ImpTestM era (ProposalProcedure era, Constitution era)

mkProposal ∷ ConwayEraImp era ⇒ GovAction era → ImpTestM era (ProposalProcedure era)

mkProposalWithRewardAccount ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ GovAction era → RewardAccount → ImpTestM era (ProposalProcedure era)

mkTreasuryWithdrawalsGovAction ∷ ConwayEraGov era ⇒ [(RewardAccount, Coin)] → ImpTestM era (GovAction era)

submitTreasuryWithdrawals ∷ ConwayEraImp era ⇒ [(RewardAccount, Coin)] → ImpTestM era GovActionId

submitVote ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ Vote → Voter → GovActionId → ImpTestM era TxId

Submits a transaction with a Vote for the given governance action as some voter

submitVote_ ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ Vote → Voter → GovActionId → ImpTestM era ()

submitYesVote_ ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ Voter → GovActionId → ImpTestM era ()

Submits a transaction that votes Yes for the given governance action as some voter

submitFailingVote ∷ (ShelleyEraImp era, ConwayEraTxBody era, HasCallStack) ⇒ Voter → GovActionId → NonEmpty (PredicateFailure (EraRule "LEDGER" era)) → ImpTestM era ()

trySubmitVote ∷ (ShelleyEraImp era, ConwayEraTxBody era) ⇒ Vote → Voter → GovActionId → ImpTestM era (Either (NonEmpty (PredicateFailure (EraRule "LEDGER" era))) TxId)

Submits a transaction that votes Yes for the given governance action as some voter, and expects an Either result.

registerDRep ∷ ConwayEraImp era ⇒ ImpTestM era (KeyHash 'DRepRole)

Submit a transaction that registers a new DRep and return the keyhash belonging to that DRep

unRegisterDRep ∷ ∀ era. (ShelleyEraImp era, ConwayEraTxCert era, ConwayEraCertState era) ⇒ Credential 'DRepRole → ImpTestM era ()

Submit a transaction that unregisters a given DRep

updateDRep ∷ ∀ era. (ShelleyEraImp era, ConwayEraTxCert era) ⇒ Credential 'DRepRole → ImpTestM era ()

Submit a transaction that updates a given DRep

delegateToDRep ∷ ConwayEraImp era ⇒ Credential 'Staking → Coin → DRep → ImpTestM era (KeyPair 'Payment)

setupSingleDRep ∷ ConwayEraImp era ⇒ Integer → ImpTestM era (Credential 'DRepRole, Credential 'Staking, KeyPair 'Payment)

Registers a new DRep, registers its stake credentials and delegates the specified amount of ADA to it.

setupDRepWithoutStake ∷ ConwayEraImp era ⇒ ImpTestM era (KeyHash 'DRepRole, KeyHash 'Staking)

In contrast to setupSingleDRep, this function does not make a UTxO entry that could count as delegated stake to the DRep

setupPoolWithStake ∷ ConwayEraImp era ⇒ Coin → ImpTestM era (KeyHash 'StakePool, Credential 'Payment, Credential 'Staking)

Sets up a stake pool with coin delegated to it.

NOTE: This uses the RegDepositDelegTxCert for delegating, so it has to be in Conway. The Shelley version of this function would have to separately register the staking credential and then delegate it.

setupPoolWithoutStake ∷ ConwayEraImp era ⇒ ImpTestM era (KeyHash 'StakePool, Credential 'Staking)

conwayModifyPParams ∷ ConwayEraGov era ⇒ (PParams era → PParams era) → ImpTestM era ()

Modify the PParams in the current state with the given function

getProposals ∷ ConwayEraGov era ⇒ ImpTestM era (Proposals era)

getEnactState ∷ ConwayEraGov era ⇒ ImpTestM era (EnactState era)

getGovActionState ∷ (HasCallStack, ConwayEraGov era) ⇒ GovActionId → ImpTestM era (GovActionState era)

Looks up the governance action state corresponding to the governance action id

lookupGovActionState ∷ ConwayEraGov era ⇒ GovActionId → ImpTestM era (Maybe (GovActionState era))

Looks up the governance action state corresponding to the governance action id

expectPresentGovActionId ∷ (HasCallStack, ConwayEraGov era) ⇒ GovActionId → ImpTestM era ()

expectMissingGovActionId ∷ (HasCallStack, ConwayEraGov era) ⇒ GovActionId → ImpTestM era ()

getRatifyEnv ∷ (ConwayEraGov era, ConwayEraCertState era) ⇒ ImpTestM era (RatifyEnv era)

Builds a RatifyEnv from the current state

calculateDRepAcceptedRatio ∷ ∀ era. (HasCallStack, ConwayEraGov era, ConwayEraCertState era) ⇒ GovActionId → ImpTestM era Rational

Calculates the ratio of DReps that have voted for the governance action

calculatePoolAcceptedRatio ∷ (ConwayEraGov era, ConwayEraCertState era) ⇒ GovActionId → ImpTestM era Rational

calculateCommitteeAcceptedRatio ∷ ∀ era. (HasCallStack, ConwayEraGov era, ConwayEraCertState era) ⇒ GovActionId → ImpTestM era Rational

Calculates the ratio of Committee members that have voted for the governance action

logAcceptedRatio ∷ (HasCallStack, ConwayEraGov era, ConwayEraCertState era) ⇒ GovActionId → ImpTestM era ()

Logs the ratios of accepted votes per category

isDRepAccepted ∷ (HasCallStack, ConwayEraGov era, ConwayEraPParams era, ConwayEraCertState era) ⇒ GovActionId → ImpTestM era Bool

Checks whether the governance action has enough DRep votes to be accepted in the next epoch. (Note that no other checks except DRep votes are used)

isSpoAccepted ∷ (HasCallStack, ConwayEraGov era, ConwayEraPParams era, ConwayEraCertState era) ⇒ GovActionId → ImpTestM era Bool

isCommitteeAccepted ∷ (HasCallStack, ConwayEraGov era, ConwayEraPParams era, ConwayEraCertState era) ⇒ GovActionId → ImpTestM era Bool

getCommitteeMembers ∷ ConwayEraImp era ⇒ ImpTestM era (Set (Credential 'ColdCommitteeRole))

getConstitution ∷ ConwayEraImp era ⇒ ImpTestM era (Constitution era)

registerInitialCommittee ∷ (HasCallStack, ConwayEraImp era) ⇒ ImpTestM era (NonEmpty (Credential 'HotCommitteeRole))

logRatificationChecks ∷ (ConwayEraGov era, ConwayEraPParams era, HasCallStack, ConwayEraCertState era) ⇒ GovActionId → ImpTestM era ()

Logs the results of each check required to make the governance action pass

resignCommitteeColdKey ∷ (ShelleyEraImp era, ConwayEraTxCert era, ConwayEraCertState era) ⇒ Credential 'ColdCommitteeRole → StrictMaybe Anchor → ImpTestM era (Maybe (Credential 'HotCommitteeRole))

Submits a transaction that resigns the cold key. Prior to resignation if there was hot credential authorization for this committee member it will be returned.

registerCommitteeHotKey ∷ (ShelleyEraImp era, ConwayEraTxCert era) ⇒ Credential 'ColdCommitteeRole → ImpTestM era (Credential 'HotCommitteeRole)

Submits a transaction that registers a hot key for the given cold key. Returns the hot key hash.

registerCommitteeHotKeys

Arguments

∷ (ShelleyEraImp era, ConwayEraTxCert era)

⇒ ImpTestM era (Credential 'HotCommitteeRole)
Hot Credential generator

→ NonEmpty (Credential 'ColdCommitteeRole)

→ ImpTestM era (NonEmpty (Credential 'HotCommitteeRole))

logCurPParams ∷ (EraGov era, ToExpr (PParamsHKD Identity era), HasCallStack) ⇒ ImpTestM era ()

submitCommitteeElection ∷ ∀ era. (HasCallStack, ConwayEraImp era) ⇒ StrictMaybe (GovPurposeId 'CommitteePurpose) → Credential 'DRepRole → Set (Credential 'ColdCommitteeRole) → Map (Credential 'ColdCommitteeRole) EpochNo → ImpTestM era (GovPurposeId 'CommitteePurpose)

electBasicCommittee ∷ ∀ era. (HasCallStack, ConwayEraImp era) ⇒ ImpTestM era (Credential 'DRepRole, Credential 'HotCommitteeRole, GovPurposeId 'CommitteePurpose)

proposalsShowDebug ∷ Proposals era → Bool → Doc AnsiStyle

getGovPolicy ∷ ConwayEraGov era ⇒ ImpTestM era (StrictMaybe ScriptHash)

submitFailingGovAction ∷ ∀ era. (ConwayEraImp era, HasCallStack) ⇒ GovAction era → NonEmpty (PredicateFailure (EraRule "LEDGER" era)) → ImpTestM era ()

submitGovActionForest ∷ (StrictMaybe GovActionId → ImpTestM era GovActionId) → StrictMaybe GovActionId → Forest () → ImpTestM era (Forest GovActionId)

submitGovActionTree ∷ (StrictMaybe GovActionId → ImpTestM era GovActionId) → StrictMaybe GovActionId → Tree () → ImpTestM era (Tree GovActionId)

getProposalsForest ∷ ConwayEraGov era ⇒ ImpTestM era (Forest (StrictMaybe GovActionId))

logProposalsForest ∷ (ConwayEraGov era, HasCallStack) ⇒ ImpTestM era ()

logProposalsForestDiff ∷ (Era era, ToExpr (PParamsHKD StrictMaybe era), HasCallStack) ⇒ Proposals era → Proposals era → ImpTestM era ()

getCCExpiry ∷ (HasCallStack, ConwayEraGov era) ⇒ Credential 'ColdCommitteeRole → ImpTestM era EpochNo

ccShouldBeExpired ∷ (HasCallStack, ConwayEraGov era) ⇒ Credential 'ColdCommitteeRole → ImpTestM era ()

ccShouldNotBeExpired ∷ (HasCallStack, ConwayEraGov era) ⇒ Credential 'ColdCommitteeRole → ImpTestM era ()

ccShouldBeResigned ∷ (HasCallStack, ConwayEraCertState era) ⇒ Credential 'ColdCommitteeRole → ImpTestM era ()

Test the resignation status for a CC cold key to be resigned

ccShouldNotBeResigned ∷ (HasCallStack, ConwayEraCertState era) ⇒ Credential 'ColdCommitteeRole → ImpTestM era ()

Test the resignation status for a CC cold key to not be resigned

getLastEnactedCommittee ∷ ConwayEraGov era ⇒ ImpTestM era (StrictMaybe (GovPurposeId 'CommitteePurpose))

getLastEnactedConstitution ∷ ConwayEraGov era ⇒ ImpTestM era (StrictMaybe (GovPurposeId 'ConstitutionPurpose))

submitParameterChange ∷ ConwayEraImp era ⇒ StrictMaybe GovActionId → PParamsUpdate era → ImpTestM era GovActionId

mkMinFeeUpdateGovAction ∷ ConwayEraImp era ⇒ StrictMaybe GovActionId → ImpTestM era (GovAction era)

mkParameterChangeGovAction ∷ ConwayEraImp era ⇒ StrictMaybe GovActionId → PParamsUpdate era → ImpTestM era (GovAction era)

mkUpdateCommitteeProposal

Arguments

∷ ConwayEraImp era

⇒ Maybe (StrictMaybe (GovPurposeId 'CommitteePurpose))
Set the parent. When Nothing is supplied latest parent will be used.

→ Set (Credential 'ColdCommitteeRole)
CC members to remove

→ [(Credential 'ColdCommitteeRole, EpochInterval)]
CC members to add

→ UnitInterval

→ ImpTestM era (ProposalProcedure era)

submitUpdateCommittee

Arguments

∷ ConwayEraImp era

⇒ Maybe (StrictMaybe (GovPurposeId 'CommitteePurpose))
Set the parent. When Nothing is supplied latest parent will be used.

→ Set (Credential 'ColdCommitteeRole)
CC members to remove

→ [(Credential 'ColdCommitteeRole, EpochInterval)]
CC members to add

→ UnitInterval

→ ImpTestM era GovActionId

expectCommitteeMemberPresence ∷ (HasCallStack, ConwayEraGov era) ⇒ Credential 'ColdCommitteeRole → ImpTestM era ()

expectCommitteeMemberAbsence ∷ (HasCallStack, ConwayEraGov era) ⇒ Credential 'ColdCommitteeRole → ImpTestM era ()

getLastEnactedParameterChange ∷ ConwayEraGov era ⇒ ImpTestM era (StrictMaybe (GovPurposeId 'PParamUpdatePurpose))

getLastEnactedHardForkInitiation ∷ ConwayEraGov era ⇒ ImpTestM era (StrictMaybe (GovPurposeId 'HardForkPurpose))

getConstitutionProposals ∷ ConwayEraGov era ⇒ ImpTestM era (Map (GovPurposeId 'ConstitutionPurpose) (PEdges (GovPurposeId 'ConstitutionPurpose)))

getParameterChangeProposals ∷ ConwayEraGov era ⇒ ImpTestM era (Map (GovPurposeId 'PParamUpdatePurpose) (PEdges (GovPurposeId 'PParamUpdatePurpose)))

expectNumDormantEpochs ∷ (HasCallStack, ConwayEraCertState era) ⇒ EpochNo → ImpTestM era ()

submitConstitution ∷ ∀ era. ConwayEraImp era ⇒ StrictMaybe (GovPurposeId 'ConstitutionPurpose) → ImpTestM era GovActionId

isDRepExpired ∷ (HasCallStack, ConwayEraCertState era) ⇒ Credential 'DRepRole → ImpTestM era Bool

expectDRepExpiry ∷ (HasCallStack, ConwayEraCertState era) ⇒ Credential 'DRepRole → EpochNo → ImpTestM era ()

expectActualDRepExpiry ∷ (HasCallStack, ConwayEraCertState era) ⇒ Credential 'DRepRole → EpochNo → ImpTestM era ()

expectDRepNotRegistered ∷ (HasCallStack, ConwayEraCertState era) ⇒ Credential 'DRepRole → ImpTestM era ()

expectCurrentProposals ∷ (HasCallStack, ConwayEraGov era) ⇒ ImpTestM era ()

expectNoCurrentProposals ∷ (HasCallStack, ConwayEraImp era) ⇒ ImpTestM era ()

expectPulserProposals ∷ (HasCallStack, ConwayEraGov era) ⇒ ImpTestM era ()

expectNoPulserProposals ∷ (HasCallStack, ConwayEraGov era) ⇒ ImpTestM era ()

currentProposalsShouldContain ∷ (HasCallStack, ConwayEraGov era) ⇒ GovActionId → ImpTestM era ()

ifBootstrap ∷ EraGov era ⇒ ImpTestM era a → ImpTestM era a → ImpTestM era a

whenBootstrap ∷ EraGov era ⇒ ImpTestM era () → ImpTestM era ()

whenPostBootstrap ∷ EraGov era ⇒ ImpTestM era () → ImpTestM era ()

submitYesVoteCCs_ ∷ ∀ era f. (ConwayEraImp era, Foldable f) ⇒ f (Credential 'HotCommitteeRole) → GovActionId → ImpTestM era ()

donateToTreasury ∷ ConwayEraImp era ⇒ Coin → ImpTestM era ()

expectMembers ∷ (HasCallStack, ConwayEraGov era) ⇒ Set (Credential 'ColdCommitteeRole) → ImpTestM era ()

showConwayTxBalance ∷ (EraUTxO era, ConwayEraTxBody era, ConwayEraCertState era) ⇒ PParams era → CertState era → UTxO era → Tx era → String

logConwayTxBalance ∷ (EraUTxO era, EraGov era, ConwayEraTxBody era, ConwayEraCertState era) ⇒ Tx era → ImpTestM era ()

submitBootstrapAware ∷ EraGov era ⇒ ImpTestM era a → (NonEmpty (PredicateFailure (EraRule "LEDGER" era)) → ImpTestM era a) → SubmitFailureExpectation era → ImpTestM era a

submitBootstrapAwareFailingVote ∷ ConwayEraImp era ⇒ Vote → Voter → GovActionId → SubmitFailureExpectation era → ImpTestM era ()

submitBootstrapAwareFailingProposal ∷ ConwayEraImp era ⇒ ProposalProcedure era → SubmitFailureExpectation era → ImpTestM era (Maybe GovActionId)

submitBootstrapAwareFailingProposal_ ∷ ConwayEraImp era ⇒ ProposalProcedure era → SubmitFailureExpectation era → ImpTestM era ()

data SubmitFailureExpectation era

Constructors

FailBootstrap (NonEmpty (PredicateFailure (EraRule "LEDGER" era)))

FailPostBootstrap (NonEmpty (PredicateFailure (EraRule "LEDGER" era)))

FailBootstrapAndPostBootstrap (FailBoth era)

data FailBoth era

Constructors

FailBoth

Fields

bootstrapFailures ∷ NonEmpty (PredicateFailure (EraRule "LEDGER" era))

postBootstrapFailures ∷ NonEmpty (PredicateFailure (EraRule "LEDGER" era))

delegateSPORewardAddressToDRep_ ∷ ConwayEraImp era ⇒ KeyHash 'StakePool → Coin → DRep → ImpTestM era ()

getCommittee ∷ ConwayEraGov era ⇒ ImpTestM era (StrictMaybe (Committee era))

Orphan instances

AlonzoEraImp ConwayEra

Instance details

Methods

scriptTestContexts ∷ Map ScriptHash ScriptTestContext

BabbageEraImp ConwayEra

Instance details

MaryEraImp ConwayEra

Instance details

ShelleyEraImp ConwayEra

Instance details

Methods

initGenesis ∷ (HasKeyPairs s, MonadState s m, HasStatefulGen g m, MonadFail m) ⇒ m (Genesis ConwayEra)

initNewEpochState ∷ (HasKeyPairs s, MonadState s m, HasStatefulGen g m, MonadFail m) ⇒ m (NewEpochState ConwayEra)

initImpTestState ∷ (HasKeyPairs s, MonadState s m, HasStatefulGen g m, MonadFail m) ⇒ m (ImpTestState ConwayEra)

impSatisfyNativeScript ∷ Set (KeyHash 'Witness) → TxBody ConwayEra → NativeScript ConwayEra → ImpTestM ConwayEra (Maybe (Map (KeyHash 'Witness) (KeyPair 'Witness)))

modifyPParams ∷ (PParams ConwayEra → PParams ConwayEra) → ImpTestM ConwayEra ()

modifyImpInitProtVer ∷ Version → SpecWith (ImpInit (LedgerSpec ConwayEra)) → SpecWith (ImpInit (LedgerSpec ConwayEra))

fixupTx ∷ Tx ConwayEra → ImpTestM ConwayEra (Tx ConwayEra)

expectTxSuccess ∷ Tx ConwayEra → ImpTestM ConwayEra ()

genRegTxCert ∷ Credential 'Staking → ImpTestM ConwayEra (TxCert ConwayEra)

genUnRegTxCert ∷ Credential 'Staking → ImpTestM ConwayEra (TxCert ConwayEra)

InjectRuleFailure "CERT" ShelleyDelegPredFailure ConwayEra

Instance details

Methods

injectFailure ∷ ShelleyDelegPredFailure ConwayEra → EraRuleFailure "CERT" ConwayEra

InjectRuleFailure "CERTS" ShelleyDelegPredFailure ConwayEra

Instance details

Methods

injectFailure ∷ ShelleyDelegPredFailure ConwayEra → EraRuleFailure "CERTS" ConwayEra

InjectRuleFailure "DELEG" ShelleyDelegPredFailure ConwayEra

Instance details

Methods

injectFailure ∷ ShelleyDelegPredFailure ConwayEra → EraRuleFailure "DELEG" ConwayEra

InjectRuleFailure "LEDGER" ShelleyDelegPredFailure ConwayEra

Instance details

Methods

injectFailure ∷ ShelleyDelegPredFailure ConwayEra → EraRuleFailure "LEDGER" ConwayEra